% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael01_nollt.R
\docType{data}
\name{ael01_nollt_1_main}
\alias{ael01_nollt_1_main}
\alias{ael01_nollt_1_pre}
\alias{ael01_nollt_1_check}
\alias{ael01_nollt_1_post}
\alias{ael01_nollt_1}
\title{\code{AEL01_NOLLT} Listing 1 (Default) Glossary of Preferred Terms and Investigator-Specified Terms.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
ael01_nollt_1_main(adam_db, dataset = "adae", deco = std_deco("AEL01_NOLLT"))

ael01_nollt_1_pre(adam_db, dataset = "adae")

ael01_nollt_1_check(adam_db, dataset = "adae", vars)

ael01_nollt_1_post(tlg)

ael01_nollt_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{AEL01_NOLLT} Listing 1 (Default) Glossary of Preferred Terms and Investigator-Specified Terms.
}
\details{
\itemize{
\item Removes duplicate rows.
\item Using dataset \code{adae}, sorts by key columns \code{AEBODSYS} and \code{AEDECOD}.
\item Using dataset \code{admh}, sorts by key columns \code{MHBODSYS} and \code{MHDECOD}.
}
}
\section{Functions}{
\itemize{
\item \code{ael01_nollt_1_main()}: Main TLG function

\item \code{ael01_nollt_1_pre()}: Preprocessing

\item \code{ael01_nollt_1_check()}: Checks

\item \code{ael01_nollt_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adae} table with columns \code{AEBODSYS}, \code{AEDECOD}, and \code{AETERM}.
\item \code{adam_db} object must contain an \code{admh} table with columns \code{MHBODSYS}, \code{MHDECOD}, and \code{MHTERM}.
}
}
\examples{
run(ael01_nollt_1, syn_data, dataset = "adae")
}
\keyword{datasets}
