% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael04.R
\docType{data}
\name{ael04_main}
\alias{ael04_main}
\alias{ael04_pre}
\alias{ael04}
\title{\code{AEL04} Listing 1 (Default) Listing of Patient Deaths.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
ael04_main(
  adam_db,
  dataset = "adsl",
  key_cols = "ID",
  disp_cols = c("ASR", "TRTSDTM", "EOSDY", "DTHADY", "DTHCAUS", "ADTHAUT"),
  split_into_pages_by_var = "ACTARM",
  ...
)

ael04_pre(adam_db, dataset = "adsl", split_into_pages_by_var = "ACTARM", ...)

ael04
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{key_cols}{(\code{character}) names of columns that should be treated as key columns when rendering the listing.
Key columns allow you to group repeat occurrences.}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{split_into_pages_by_var}{(\code{character} or \code{NULL}) the name of the variable to split the listing by.}

\item{...}{not used.}
}
\value{
the main function returns an \code{rlistings} or a \code{list} object.
}
\description{
\code{AEL04} Listing 1 (Default) Listing of Patient Deaths.
}
\section{Functions}{
\itemize{
\item \code{ael04_main()}: Main TLG function

\item \code{ael04_pre()}: Preprocessing

}}
\examples{
res <- run(ael04, syn_data)
}
\keyword{datasets}
