% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\name{aet01_1}
\alias{aet01_1}
\title{\code{AET01} Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.}
\usage{
aet01_1(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET01"),
  .study = list(armvar = "ARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{

}
\description{
Overview of death and summary of adverse events.
}
\examples{
library(dm)

db <- syn_test_data() \%>\%
  preprocess_data("aet01_1")

aet01_1(db, armvar = "ARM")

}
