% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01_aesi.R
\docType{data}
\name{aet01_aesi_1_main}
\alias{aet01_aesi_1_main}
\alias{aet01_aesi_1_lyt}
\alias{aet01_aesi_1_pre}
\alias{aet01_aesi_1_check}
\alias{aet01_aesi_post}
\alias{aet01_aesi_1}
\title{\code{AET01_AESI} Table 1 (Default) Adverse Event of Special Interest Summary Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet01_aesi_1_main(
  adam_db,
  armvar = "ACTARM",
  aesi_vars = NA,
  deco = std_deco("AET01_AESI"),
  ...
)

aet01_aesi_1_lyt(armvar, aesi_vars, deco, lbl_aesi_vars, ...)

aet01_aesi_1_pre(adam_db, ...)

aet01_aesi_1_check(
  adam_db,
  req_tables = c("adsl", "adae"),
  armvar = "ACTARM",
  aesi_vars = c("WD", "DSM", "CONTRT", "ALL_RESOLVED", "ALLRESWD", "ALLRESDSM",
    "ALLRESCONTRT", "NOT_RESOLVED", "NOTRESWD", "NOTRESDSM", "NOTRESCONTRT", "SER",
    "SERWD", "SERDSM", "SERCONTRT", "REL", "RELWD", "RELDSM", "RELCONTRT", "RELSER"),
  ...
)

aet01_aesi_post(tlg, prune_0 = FALSE, deco = std_deco("AET01_AESI"), ...)

aet01_aesi_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{aesi_vars}{(\code{character}) the AESI variables to be included in the summary. Defaults to \code{NA}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used}

\item{lbl_aesi_vars}{(\code{character}) the labels of the AESI variables to be summarized.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
\code{AET01_AESI} Table 1 (Default) Adverse Event of Special Interest Summary Table.
}
\details{
\itemize{
\item Does not remove rows with zero counts by default.
}
}
\section{Functions}{
\itemize{
\item \code{aet01_aesi_1_main()}: Main TLG function

\item \code{aet01_aesi_1_lyt()}: Layout

\item \code{aet01_aesi_1_pre()}: Preprocessing

\item \code{aet01_aesi_1_check()}: Checks

\item \code{aet01_aesi_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adae} table with columns \code{"WD"}, \code{"DSM"}, \code{"CONTRT"}, \code{"ALL_RESOLVED"},
\code{"NOT_RESOLVED"}, \code{"SER"}, \code{"REL"}, the column specified by \code{armvar}, and the columns passed to \code{aesi_vars}.
\item \code{aesi_vars} may contain any/all of the following variables to display: \code{"ALLRESWD"}, \code{"ALLRESDSM"},
\code{"ALLRESCONTRT"}, \code{"NOTRESWD"}, \code{"NOTRESDSM"}, \code{"NOTRESCONTRT"}, \code{"SERWD"}, \code{"SERDSM"}, \code{"SERCONTRT"},
\code{"RELWD"}, \code{"RELDSM"}, \code{"RELCONTRT"}, \code{"RELSER"}.
\item \code{aesi_vars} variable prefixes are defined as follows:
\itemize{
\item \code{"ALLRES"} = "all non-fatal adverse events resolved"
\item \code{"NOTRES"} = "at least one unresolved or ongoing non-fatal adverse event"
\item \code{"SER"} = "serious adverse event"
\item \code{"REL"} = "related adverse event"
}
\item \code{aesi_vars} variable suffixes are defined as follows:
\itemize{
\item \code{"WD"} = "patients with study drug withdrawn"
\item \code{"DSM"} = "patients with dose modified/interrupted"
\item \code{"CONTRT"} = "patients with treatment received"
}
}
}
\examples{
run(aet01_aesi_1, syn_data)
}
\keyword{datasets}
