% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01_aesi.R
\name{aet01_aesi_lyt}
\alias{aet01_aesi_lyt}
\title{\code{aet01_aesi} Layout}
\usage{
aet01_aesi_lyt(arm_var, aesi_vars, lbl_overall, lbl_aesi_vars, grade_groups)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aesi_vars}{(\code{character}) the labels of the AESI variables to be summarized.}
}
\description{
\code{aet01_aesi} Layout
}
\keyword{internal}
