% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\name{aet01_lyt}
\alias{aet01_lyt}
\title{\code{aet01} Layout}
\usage{
aet01_lyt(arm_var, lbl_overall, anl_vars, anl_lbls, lbl_vars)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{anl_vars}{Named (\code{list}) of analysis variables.}

\item{anl_lbls}{(\code{character}) of labels.}

\item{lbl_vars}{Named (\code{list}) of analysis labels.}
}
\description{
\code{aet01} Layout
}
\keyword{internal}
