% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\docType{data}
\name{aet02_2}
\alias{aet02_2}
\title{\code{AET02_2}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet02_2
}
\description{
\code{AET02_2}
}
\seealso{
\code{\link[=aet02_2_main]{aet02_2_main()}}
}
\keyword{datasets}
