% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\docType{data}
\name{aet02_3}
\alias{aet02_3}
\title{\code{AET02_3}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet02_3
}
\description{
\code{AET02_3}
}
\seealso{
\code{\link[=aet02_3_main]{aet02_3_main()}}
}
\keyword{datasets}
