% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_3_main}
\alias{aet02_3_main}
\alias{aet02_3_lyt}
\alias{aet02_3_pre}
\title{\code{AET02} Table 3 (Supplementary) Adverse Events by Dictionary-Derived Term Table 3.}
\usage{
aet02_3_main(
  adam_db,
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

aet02_3_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  lbl_aedecod = "MedDRA Preferred Term",
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

aet02_3_pre(adam_db, ...)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{...}{not used.}
}
\description{
The \code{AET02_3} table provides an overview of the number of patients experiencing adverse events and the number of
adverse events categorized by Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Dictionary-Derived Code by highest overall frequencies.
\item Missing values of \code{AEDECOD} in \code{aead} are labeled by \verb{No Coding available}.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_3_lyt()}: \code{aet02_3} Layout

\item \code{aet02_3_pre()}: \code{aet02_3} Preprocessing

}}
\examples{
library(dm)

db <- syn_test_data() \%>\%
  aet02_3_pre()

aet02_3_main(adam_db = db) \%>\% head()

aet02_3_main(db, lbl_overall = "All Patients") \%>\% head()
aet02_3_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  lbl_aedecod = "Dictionary-Derived Term",
  deco = std_deco("AET02")
)
syn_test_data() \%>\%
  aet02_3_pre()
}
