% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_lyt}
\alias{aet02_lyt}
\title{aet02 Layout}
\usage{
aet02_lyt(arm_var, lbl_overall, row_split_var, lbl_row_split, lbl_aedecod)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aedecod}{(\code{string}) text label for \code{AEDECOD}.}

\item{lbl_aebodsys}{(\code{string}) text label for \code{AEBODSYS}.}
}
\description{
aet02 Layout
}
\keyword{internal}
