% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\name{aet03_1}
\alias{aet03_1}
\alias{aet03_1_lyt}
\title{AET03 Table 1 (Default) Advert Events by Greatest Intensity}
\usage{
aet03_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)

aet03_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "",
  lbl_AEDECOD = "",
  lbl_AESEV = "",
  gradation = .study$gradation,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = "", gradation = c("MILD", "MODERATE",
    "SEVERE", "LIFE THREATENING"))
)
}
\arguments{
\item{adsl}{(\code{data.frame}) Subject Level Analysis Dataset}

\item{adae}{(\code{data.frame}) Adverse Events Analysis Dataset}

\item{armvar}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{list}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AEBODSYS}{(\code{string}) text label for AEBODSYS.}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}

\item{lbl_AESEV}{(\code{string}) text label for AESEV.}

\item{gradation}{(\verb{vector of strings}) describing the severity levels present in the dataset.}
}
\value{

}
\description{
An adverse events table categorized by System Organ Class, Dictionary-Derived Term  and Greatest intensity
}
\details{
\itemize{
\item Default Adverse Events by Greatest Intensity table
\item Numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm
\item Does not include a total column by default
\item Sort by Body System or Organ Class (SOC) and Dictionary-Derived Term (PT)
}
}
\section{Functions}{
\itemize{
\item \code{aet03_1_lyt}: 
}}

\examples{

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

aet03_1(adsl, adae)
aet02_1(adsl, adae, lbl_overall = "All Patients")

aet03_1_lyt(armvar = "ACTARM",
 lbl_overall = "",
 lbl_AEBODSYS = "Body System or Organ Class",
 lbl_AESEV = "Severity/Intensity",
 lbl_AEDECOD = "Dictionary-Derived Term",
 deco = std_deco("AET03"))
}
