% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\name{aet03_1_lyt}
\alias{aet03_1_lyt}
\title{AET03 Layout 1 (Default)}
\usage{
aet03_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "",
  lbl_AEDECOD = "",
  lbl_AESEV = "",
  gradation = .study$gradation,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = "", gradation = c("MILD", "MODERATE",
    "SEVERE", "LIFE THREATENING"))
)
}
\arguments{
\item{lbl_AEBODSYS}{(\code{string}) text label for AEBODSYS.}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}

\item{lbl_AESEV}{(\code{string}) text label for AESEV.}

\item{gradation}{(\verb{vector of strings}) describing the severity levels present in the dataset.}
}
\value{

}
\description{
AET03 Layout 1 (Default)
}
\examples{
aet03_1_lyt(armvar = "ACTARM",
 lbl_overall = "",
 lbl_AEBODSYS = "Body System or Organ Class",
 lbl_AESEV = "Severity/Intensity",
 lbl_AEDECOD = "Dictionary-Derived Term",
 deco = std_deco("AET03"))
}
