% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\name{aet03_1_main}
\alias{aet03_1_main}
\alias{aet03_1_lyt}
\alias{aet03_1_pre}
\title{\code{AET03} Table 1 (Default) Advert Events by Greatest Intensity Table 1.}
\usage{
aet03_1_main(
  adam_db,
  armvar = .study$actualarm,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

aet03_1_lyt(
  armvar = .study$actualarm,
  lbl_aebodsys = "",
  lbl_aedecod = "",
  severity_grade = .study$severity_grade,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, severity_grade = c("MILD",
    "MODERATE", "SEVERE", "LIFE THREATENING"))
)

aet03_1_pre(adam_db, ...)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{severity_grade}{(\verb{vector of character}) describing the severity levels present in the data set.}

\item{...}{not used.}
}
\description{
An adverse events table categorized by System
Organ Class, Dictionary-Derived Term  and Greatest intensity.
}
\details{
\itemize{
\item Default Adverse Events by Greatest Intensity table.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by Body System or Organ Class (\code{SOC}) and Dictionary-Derived Term (\code{PT}).
}
}
\section{Functions}{
\itemize{
\item \code{aet03_1_lyt()}: \code{aet03_1} Layout

\item \code{aet03_1_pre()}: \code{aet03_1} Preprocessing

}}
\examples{
library(dm)

db <- syn_test_data() \%>\%
  aet03_1_pre()

aet03_1_main(db)
aet03_1_main(db, lbl_overall = "All Patients")
aet03_1_lyt(
  armvar = "ACTARM",
  lbl_aebodsys = "Body System or Organ Class",
  lbl_aedecod = "Dictionary-Derived Term",
  lbl_overall = NULL,
  deco = std_deco("AET03")
)
syn_test_data() \%>\%
  aet03_1_pre()
}
