% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\name{aet03_lyt}
\alias{aet03_lyt}
\title{\code{aet03} Layout}
\usage{
aet03_lyt(arm_var, lbl_overall, lbl_aebodsys, lbl_aedecod, intensity_grade)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aebodsys}{(\code{string}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{string}) text label for \code{AEDECOD}.}

\item{intensity_grade}{(\code{character}) describing the intensity levels present in the dataset.}
}
\description{
\code{aet03} Layout
}
\keyword{internal}
