% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\name{aet04_1_lyt}
\alias{aet04_1_lyt}
\title{AET04 Layout 1 (Default)}
\usage{
aet04_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "AEBODSYS",
  lbl_AEDECOD = "AEDECOD",
  gr_grp = .study$gr_grp,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = "", gr_grp = list(`- Any Grade -` =
    c("1", "2", "3", "4", "5"), `Grade 1-2` = c("1", "2"), `Grade 3-4` = c("3", "4"),
    `Grade 5` = c("5")))
)
}
\arguments{
\item{lbl_AEBODSYS}{(\code{string}) text label for AEBODSYS.}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}

\item{gr_grp}{(\code{list}) putting in correspondence severity levels and labels.}
}
\value{

}
\description{
AET04 Layout 1 (Default)
}
\examples{
aet04_1_lyt(armvar = "ACTARM")
}
