% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\name{aet04_lyt}
\alias{aet04_lyt}
\title{\code{aet04} Layout}
\usage{
aet04_lyt(
  arm_var,
  total_var,
  lbl_overall,
  lbl_aebodsys,
  lbl_aedecod,
  grade_groups
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{total_var}{(\code{string}) variable to create summary of all variables.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aebodsys}{(\code{string}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{string}) text label for \code{AEDECOD}.}

\item{grade_groups}{(\code{list}) putting in correspondence toxicity grades and labels.}
}
\description{
\code{aet04} Layout
}
\keyword{internal}
