% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet05_all.R
\name{aet05_all_lyt}
\alias{aet05_all_lyt}
\title{\code{aet05_all} Layout}
\usage{
aet05_all_lyt(arm_var)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}
}
\description{
\code{aet05_all} Layout
}
\keyword{internal}
