% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet05.R
\name{aet05_lyt}
\alias{aet05_lyt}
\title{\code{aet05} Layout}
\usage{
aet05_lyt(arm_var, vars, n_events, control)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}
}
\description{
\code{aet05} Layout
}
\keyword{internal}
