% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet10.R
\docType{data}
\name{aet10_1_lyt}
\alias{aet10_1_lyt}
\alias{aet10_1_pre}
\alias{aet10_1_check}
\alias{aet10_1_post}
\alias{aet10_1}
\title{\code{AET10} Table 1 (Default) Most Common (xx\%) Adverse Events Preferred Terms Table 1}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet10_1_lyt(arm_var, lbl_overall, lbl_aedecod = "MedDRA Preferred Term", deco)

aet10_1_pre(adam_db)

aet10_1_check(adam_db, req_tables = c("adsl", "adae"), arm_var = "ACTARM")

aet10_1_post(tlg, prune_0 = TRUE, atleast = 0.05)

aet10_1
}
\arguments{
\item{arm_var}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{atleast}{given cut-off default is \code{0.05}}
}
\description{
The \code{AET10} table Include Adverse Events occurring with user-specified threshold X\% in at least
one of the treatment groups. Standard table summarized by preferred term (PT).
Order the data by total column frequency from most to least frequently reported PT (regardless of SOC).
}
\section{Functions}{
\itemize{
\item \code{aet10_1_lyt()}: Layout

\item \code{aet10_1_pre()}: Preprocessing

\item \code{aet10_1_check()}: Checks

\item \code{aet10_1_post()}: Postprocessing

}}
\examples{
run(aet10_1, syn_data)
}
\keyword{datasets}
