% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet10.R
\name{aet10_lyt}
\alias{aet10_lyt}
\title{\code{aet10} Layout}
\usage{
aet10_lyt(arm_var, row_split_var, lbl_row_split, lbl_overall, lbl_aedecod)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{row_split_var}{(\code{character}) row split variables.}

\item{lbl_row_split}{(\code{character}) label for \code{row_split_var}.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aedecod}{(\code{string}) text label for \code{AEDECOD}.}
}
\description{
\code{aet10} Layout
}
\keyword{internal}
