% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{afun_p}
\alias{afun_p}
\title{Analyze with defined precision}
\usage{
afun_p(
  x,
  .N_col,
  .spl_context,
  precision,
  .N_row,
  .var = NULL,
  .df_row = NULL,
  .stats = NULL,
  .labels = NULL,
  .indent_mods = NULL,
  ...
)
}
\arguments{
\item{x}{value to analyze}

\item{.N_col}{(\code{int}) See \code{tern::analyze_variables}.}

\item{.spl_context}{split context.}

\item{precision}{(named \code{list} of \code{integer}) where names of columns found in \code{.df_row} and the values indicate the
number of digits in statistics for numeric value. If \code{default} is set, and parameter precision not specified, the
value for \code{default} will be used. If neither are provided, auto determination is used. See \code{\link[tern:format_auto]{tern::format_auto}}.}

\item{.N_row}{(\code{int}) See \code{tern::analyze_variables}.}

\item{.var}{variable name.}

\item{.stats}{(named \code{list} of character)  where names of columns found in \code{.df_row} and the values indicate the
statistical analysis to perform. If \code{default} is set, and parameter precision not specified, the
value for \code{default} will be used.}

\item{.labels}{(\code{character}) See \code{tern::analyze_variables}.}

\item{.indent_mods}{(\code{integer}) See \code{tern::analyze_variables}.}

\item{...}{additional arguments for \code{tern::a_summary}.}
}
\description{
Analyze with defined precision
}
\keyword{internal}
