% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{afun_skip_baseline}
\alias{afun_skip_baseline}
\title{Analyze skip baseline}
\usage{
afun_skip_baseline(
  x,
  .var,
  .spl_context,
  paramcdvar,
  visitvar,
  skip,
  precision,
  default_precision,
  ...
)
}
\arguments{
\item{x}{value to analyze}

\item{.var}{variable name.}

\item{.spl_context}{split context.}

\item{paramcdvar}{(\code{string}) name of parameter code.}

\item{visitvar}{(\code{string}) name of the visit variable.}

\item{skip}{Named (\code{character}) indicating the pairs to skip in analyze.}

\item{precision}{(named \code{list} of \code{integer}) where names are values found in the \code{PARAMCD} column and the the values
indicate the number of digits that should be represented for \code{min}, \code{max} and \code{median}. \code{Mean} and \code{sd} are
represented with one more decimal of precision.}

\item{default_precision}{(\code{integer}) the default number of digits.}

\item{...}{not used.}
}
\description{
Analyze skip baseline
}
\keyword{internal}
