% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{append_to_pmap}
\alias{append_to_pmap}
\title{Append A Preprocessing Map Entry}
\usage{
append_to_pmap(x, y)
}
\arguments{
\item{x}{(\code{data.frame}) in the structure returned from \code{std_pmap} or \code{pmap_entry}}

\item{y}{(\code{data.frame}) in the structure returned from \code{std_pmap} or \code{pmap_entry}}
}
\description{
Append A Preprocessing Map Entry
}
