% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{args_ls}
\alias{args_ls}
\alias{args_ls,chevron_tlg-method}
\title{Get Arguments List}
\usage{
args_ls(x, simplify = FALSE, omit = NULL)

\S4method{args_ls}{chevron_tlg}(x, simplify = FALSE, omit = NULL)
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{simplify}{(\code{flag}) whether to simplify the output, coalescing the values of the parameters. The order of
priority for the value of the parameters is: \code{main}, \code{preprocess} and \code{postprocess}.}

\item{omit}{(\code{character}) the names of the argument to omit from the output.}
}
\description{
Get Arguments List
}
