% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_args}
\alias{assert_args}
\title{Assert Validity of Arguments}
\usage{
assert_args(object, args_names)
}
\arguments{
\item{object}{(\code{chevron_tlg}) defining the valid arguments.}

\item{args_names}{(\code{character}) to assert.}
}
\value{
invisible \code{NULL} if all \code{args_names} are valid argument for the \code{chevron_tlg} object or an error otherwise.
}
\description{
Assert Validity of Arguments
}
\examples{
\dontrun{
assert_args(
  aet01_1,
  list(lbl_ = "xxx", "lbl_overall" = "All Patient", armvar = "ACTARM", "_var" = "ACT", "xxxx" = 12)
)
assert_args(aet01_1, list("lbl_overall" = "All Patient"))
}
}
\keyword{internal}
