% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_colnames}
\alias{assert_colnames}
\title{Check if strings are column names of a data frame}
\usage{
assert_colnames(df, x)
}
\arguments{
\item{df}{\verb{(data frame)}}

\item{x}{\verb{(vector of strings)}}
}
\value{

}
\description{
Check if strings are column names of a data frame
}
\examples{

assert_colnames(mtcars, c("mpg","cyl"))

}
