% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\name{assert_fun_or_list}
\alias{assert_fun_or_list}
\title{Assertion that an object is a function or a list of functions}
\usage{
assert_fun_or_list(x, add = NULL)
}
\arguments{
\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}

\item{db}{(\code{x}) input to check for type.}
}
\description{
Assertion that an object is a function or a list of functions
}
\keyword{internal}
