% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_one_tablenames}
\alias{assert_one_tablenames}
\title{Assert that at least one name is among table names of a \code{dm} object}
\usage{
assert_one_tablenames(db, tab, null_ok = TRUE, qualifier = NULL)
}
\arguments{
\item{db}{(\code{dm}) input to check for the presence or tables.}

\item{tab}{(\code{character}) the names of the tables to be checked.}

\item{null_ok}{(\code{flag}) can \code{x} be NULL.}

\item{qualifier}{(\code{string}) to be returned if the check fails.}
}
\description{
Assert that at least one name is among table names of a \code{dm} object
}
\examples{
\dontrun{
assert_one_tablenames(dm::dm_nycflights13(), c("mpg", "new", "hoho"), qualifier = "first test:")
}
}
\keyword{internal}
