% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_single_paramcd}
\alias{assert_single_paramcd}
\title{Check to have only one PARAMCD in the analysis dataset}
\usage{
assert_single_paramcd(param_val)
}
\arguments{
\item{param_val}{value of PARAMCD}
}
\description{
Check to have only one PARAMCD in the analysis dataset
}
