% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_single_value}
\alias{assert_single_value}
\title{Check variable only has one unique value.}
\usage{
assert_single_value(x, label = deparse(substitute(x)))
}
\arguments{
\item{x}{value vector.}

\item{label}{(\code{string}) label of input.}
}
\description{
Check variable only has one unique value.
}
