% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_study_object}
\alias{assert_study_object}
\title{Assertion on the Study Object}
\usage{
assert_study_object(x)
}
\arguments{
\item{x}{(\code{list}) to be assessed.}
}
\value{
invisible \code{NULL} if the object fits the criteria of a \code{.study} object.
}
\description{
Assertion on the Study Object
}
\examples{
assert_study_object(std_study_object)

}
