% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_subset_suggest}
\alias{assert_subset_suggest}
\title{Assert Subset and Provide Suggestions}
\usage{
assert_subset_suggest(x, choices)
}
\arguments{
\item{x}{(\code{character}) to be asserted}

\item{choices}{(\code{character}) set of possible values.}
}
\value{
invisible \code{NULL} if all \code{x} are subset of \code{choices} or an informative warning otherwise.
}
\description{
Assert Subset and Provide Suggestions
}
\examples{
\dontrun{
assert_subset_suggest(c("lbl_overall"), "lbl_overall")
assert_subset_suggest(
  c("lbl_", "xxx", "arm_var"),
  c("lbl_overall", "arm_var")
)
}
}
\keyword{internal}
