% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_col_var}
\alias{assert_valid_col_var}
\title{Check col_split var is valid}
\usage{
assert_valid_col_var(x, label = "variable")
}
\arguments{
\item{x}{value of col_split variable}

\item{label}{(\code{string}) hints.}
}
\description{
Check col_split var is valid
}
