% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_var_pair}
\alias{assert_valid_var_pair}
\title{Check col_split vars are of same levels}
\usage{
assert_valid_var_pair(x, y, lab1 = "var1", lab2 = "var2")
}
\arguments{
\item{x}{value of col_split variable 1.}

\item{y}{value of col_split variable 2.}

\item{lab1}{(\code{string}) label hint for variable 1.}

\item{lab2}{(\code{string}) label hint for variable 2.}
}
\description{
Check col_split vars are of same levels
}
