% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_type}
\alias{assert_valid_type}
\title{Check variable is of correct type}
\usage{
assert_valid_type(x, types, label = deparse(substitute(x)))
}
\description{
Check variable is of correct type
}
