% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_var_pair}
\alias{assert_valid_var_pair}
\title{Check variables are of same levels}
\usage{
assert_valid_var_pair(
  df1,
  df2,
  var,
  lab1 = deparse(substitute(df1)),
  lab2 = deparse(substitute(df2))
)
}
\arguments{
\item{df1}{(\code{data.frame}) input.}

\item{df2}{(\code{data.frame}) input.}

\item{var}{(\code{string}) variable to check.}

\item{lab1}{(\code{string}) label hint for df1.}

\item{lab2}{(\code{string}) label hint for df2.}
}
\description{
Check variables are of same levels
}
