% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_all_colnames}
\alias{check_all_colnames}
\title{Check that all names are among column names}
\usage{
check_all_colnames(df, x, null_ok = TRUE, qualifier = NULL)
}
\arguments{
\item{df}{(\code{data.frame})}

\item{x}{(\code{character}) the names of the columns to be checked.}

\item{null_ok}{(\code{flag}) can \code{x} be NULL.}

\item{qualifier}{(\code{string}) to be returned if the check fails.}
}
\description{
Check that all names are among column names
}
\examples{
\dontrun{
check_all_colnames(mtcars, c("x", "y"))
}
}
\keyword{internal}
