% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_col_contains}
\alias{check_col_contains}
\title{Check that value exists in the column}
\usage{
check_col_contains(
  df,
  value,
  x,
  null_ok = TRUE,
  df_name = deparse(substitute(df))
)
}
\arguments{
\item{df}{(\code{data.frame})}

\item{value}{(\code{string}) the value to be checked.}

\item{x}{(\code{string}) the names of the column.}

\item{df_name}{(\code{string}) name of the data frame}
}
\description{
Check that value exists in the column
}
\examples{
\dontrun{
check_col_contains(iris, "setosa", "Species")
}
}
\keyword{internal}
