% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_one_colnames}
\alias{check_one_colnames}
\title{Check that at least one name is among column names}
\usage{
check_one_colnames(df, x, null_ok = TRUE, qualifier = NULL)
}
\arguments{
\item{df}{(\code{data.frame})}

\item{x}{(\code{character}) the names of the columns to be checked.}

\item{null_ok}{(\code{flag}) can \code{x} be NULL.}

\item{qualifier}{(\code{string}) to be returned if the check fails.}
}
\description{
Check that at least one name is among column names
}
\examples{
check_one_colnames(mtcars, c("new", "hoho"))
}
\keyword{internal}
