% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{chevron-package}
\alias{chevron}
\alias{chevron-package}
\title{\code{chevron} package}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Provide standard tables, listings, and graphs (TLGs) libraries used in clinical trials. This package implements a structure to reformat the data with 'dunlin', create reporting tables using 'rtables' and 'tern' with standardized input arguments to enable quick generation of standard outputs. In addition, it also provides comprehensive data checks and script generation functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/chevron/}
  \item \url{https://github.com/insightsengineering/chevron/}
  \item Report bugs at \url{https://github.com/insightsengineering/chevron/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Liming Li \email{liming.li@roche.com} (\href{https://orcid.org/0009-0008-6870-0878}{ORCID}) (Original creator of the package)
  \item Benoit Falquet \email{benoit.falquet@roche.com} (\href{https://orcid.org/0000-0002-4434-3799}{ORCID}) (Original creator of the package)
  \item Xiaoli Duan \email{xiaoli.duan@roche.com}
}

Other contributors:
\itemize{
  \item Adrian Waddell \email{adrian.waddell@gmail.com} [contributor]
  \item Chenkai Lv \email{chenkai.lv@roche.com} [contributor]
  \item Pawel Rucki \email{pawel.rucki@roche.com} [contributor]
  \item Tim Barnett \email{timothy.barnett@roche.com} [contributor]
  \item Tian Fang \email{tian.fang@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
