% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\name{chevron_l}
\alias{chevron_l}
\title{\code{chevron_l} constructor}
\usage{
chevron_l(
  main = function(adam_db, lyt_ls, ...) lyt_ls[[1]](adam_db),
  lyt = list(function(...) list(...)),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) {
     tlg
 },
  adam_datasets = NA_character_
)
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}, with \code{adam_db} as first argument and \code{...} as last argument. Typically
one of the \verb{_main} function of \code{chevron}.}

\item{lyt}{(a single \code{function} returning \code{PreDataTableLayouts} object or \code{PreDataTableLayouts} object or \code{list} of
either \code{functions} or \code{PreDataTableLayouts} type of elements) typically one of the \verb{_lyt} function of \code{chevron}.
Functions passed to \code{lyt}, whether as a single \code{function} or as a \verb{list of functions}, must have the \code{...} formal
argument.}

\item{preprocess}{(\code{function}) returning a pre-processed \code{dm} object, with \code{adam_db} as first argument and \code{...} as
last argument. Typically one of the \verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post-processed \code{tlg}, with \code{tlg} as first argument.}

\item{adam_datasets}{(\code{character}) representing the names of the tables from an \code{ADaM} dataset required for \code{tlg}
creation.}
}
\description{
\code{chevron_l} constructor
}
