% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\name{chevron_tlg}
\alias{chevron_tlg}
\alias{chevron_tlg-internal}
\title{\code{chevron_tlg} constructor}
\usage{
chevron_tlg(
  main,
  preprocess,
  postprocess,
  adam_datasets,
  type = c(NA, "table", "listing", "graph"),
  ...
)
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}, with \code{adam_db} as first argument and \code{...} as last argument. Typically
one of the \verb{_main} function of \code{chevron}.}

\item{preprocess}{(\code{function}) returning a pre-processed \code{dm} object, with \code{adam_db} as first argument and \code{...} as
last argument. Typically one of the \verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post-processed \code{tlg}, with \code{tlg} as first argument.}

\item{adam_datasets}{(\code{character}) representing the names of the tables from an \code{ADaM} data set required for \code{tlg}
creation.}

\item{type}{(\code{string}) indicating the subclass.}

\item{...}{used to pass additional class specific argument. see \link{chevron_t}}
}
\description{
\code{chevron_tlg} constructor
}
\keyword{internal}
