% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\docType{data}
\name{cmt01a_1}
\alias{cmt01a_1}
\title{\code{CMT01A_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
cmt01a_1
}
\description{
\code{CMT01A_1}
}
\seealso{
\code{\link[=cmt01a_1_main]{cmt01a_1_main()}}
}
\keyword{datasets}
