% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\docType{data}
\name{cmt01a_2}
\alias{cmt01a_2}
\title{\code{CMT01A_2}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
cmt01a_2
}
\description{
\code{CMT01A_2}
}
\seealso{
\code{\link[=cmt01a_2_main]{cmt01a_2_main()}}
}
\keyword{datasets}
