% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\docType{data}
\name{cmt01a_3}
\alias{cmt01a_3}
\title{\code{CMT01A_3} \code{chevron_tlg}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
cmt01a_3
}
\description{
\code{CMT01A_3} \code{chevron_tlg}
}
\seealso{
\code{\link[=cmt01a_3_main]{cmt01a_3_main()}}
}
\keyword{datasets}
