% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\name{cmt01a_lyt}
\alias{cmt01a_lyt}
\title{cmt01a Layout}
\usage{
cmt01a_lyt(
  arm_var,
  lbl_overall,
  incl_n_treatment = TRUE,
  row_split_var,
  lbl_row_split,
  medname_var,
  lbl_medname_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{incl_n_treatment}{(\code{flag}) include total number of treatments per medication.}

\item{row_split_var}{(\code{character}) the variable defining the medication category. By default \code{ATC2}.}

\item{medname_var}{(\code{string}) variable name of medical treatment name.}

\item{lbl_medname_var}{(\code{string}) label for the variable defining the medication name.}
}
\description{
cmt01a Layout
}
\keyword{internal}
