% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt02_pt.R
\docType{data}
\name{cmt02_pt_1}
\alias{cmt02_pt_1}
\title{\code{CMT02_PT_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
cmt02_pt_1
}
\description{
\code{CMT02_PT_1}
}
\seealso{
\code{\link[=cmt02_pt_1_main]{cmt02_pt_1_main()}}
}
\keyword{datasets}
