% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt02_pt.R
\name{cmt02_pt_1_main}
\alias{cmt02_pt_1_main}
\alias{cmt02_pt_1_lyt}
\alias{cmt02_pt_1_pre}
\title{\code{CMT02_PT} Table 1 (Default) Concomitant Medications by Preferred Name.}
\usage{
cmt02_pt_1_main(
  adam_db,
  armvar = .study$planarm,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("CMT02_PT"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

cmt02_pt_1_lyt(
  armvar = .study$planarm,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("CMT02_PT"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

cmt02_pt_1_pre(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
A concomitant medication table with the
number of subjects and the total number of treatments by medication name sorted by frequencies.
}
\details{
\itemize{
\item Data should be filtered for concomitant medication. \code{(ATIREL == "CONCOMITANT")}.
\item Order by decreasing total number of patients with the specific medication.
\item Does not include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{cmt02_pt_1_lyt}: \code{cmt02_pt_1} Layout

\item \code{cmt02_pt_1_pre}: \code{cmt02_pt_1} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  dm_zoom_to("adcm") \%>\%
  filter(.data$ATIREL == "CONCOMITANT") \%>\%
  dm_update_zoomed() \%>\%
  cmt02_pt_1_pre()

cmt02_pt_1_main(adam_db = db)
cmt02_pt_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
syn_test_data() \%>\%
  cmt02_pt_1_pre()
}
