% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt02_pt.R
\name{cmt02_pt_lyt}
\alias{cmt02_pt_lyt}
\title{cmt02_pt Layout}
\usage{
cmt02_pt_lyt(arm_var, lbl_overall, medname_var, medname_lbl)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
cmt02_pt Layout
}
\keyword{internal}
