% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_month}
\alias{convert_to_month}
\title{Helper function to convert to months if needed}
\usage{
convert_to_month(x, unit)
}
\arguments{
\item{x}{(\code{numeric}) time.}

\item{unit}{(\code{character}) or (\code{factor}) time unit.}
}
\value{
A \code{numeric} vector with the time in months.
}
\description{
Helper function to convert to months if needed
}
