% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{count_or_summarize}
\alias{count_or_summarize}
\title{Count or summarize by groups}
\usage{
count_or_summarize(lyt, var, level, detail_vars, indent_mod = 0L, ...)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts}) \code{rtable} layout.}

\item{var}{(\code{string}) of analysis variable.}

\item{level}{(\code{string}) level to be displayed.}

\item{detail_vars}{(\code{character}) of variables for detail information.}
}
\description{
Count or summarize by groups
}
\keyword{internal}
