% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{count_patients_recursive}
\alias{count_patients_recursive}
\title{Count patients recursively}
\usage{
count_patients_recursive(lyt, anl_vars, anl_lbls, lbl_vars)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts}) rtable layout.}

\item{anl_vars}{Named (\code{list}) of analysis variables.}

\item{anl_lbls}{(\code{character}) of labels.}

\item{lbl_vars}{Named (\code{list}) of analysis labels.}
}
\description{
Count patients recursively
}
\keyword{internal}
