% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxt02.R
\name{coxt02_lyt}
\alias{coxt02_lyt}
\title{\code{coxt02} Layout}
\usage{
coxt02_lyt(variables, lbl_vars, control)
}
\arguments{
\item{variables}{(\code{list}) list of variables in a Cox proportional hazards regression model.}
}
\description{
\code{coxt02} Layout
}
\keyword{internal}
