% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_id_listings}
\alias{create_id_listings}
\title{Concatenate Site and Subject ID}
\usage{
create_id_listings(site, subject, sep = "/")
}
\arguments{
\item{site}{(\code{string})}

\item{subject}{(\code{string})}

\item{sep}{(\code{string})}
}
\description{
Concatenate Site and Subject ID
}
\note{
the \code{{Patient_label}} whisker placeholder will be used in the label.
}
\examples{
create_id_listings("BRA-1", "xxx-1234")
}
