% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctc_dir}
\alias{ctc_dir}
\title{CTC Grade Direction Data}
\format{
An object of class \code{data.frame} with 69 rows and 3 columns.
}
\usage{
ctc_dir
}
\description{
CTC Grade Direction Data
}
\keyword{datasets}
