% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctcv4_dir}
\alias{ctcv4_dir}
\title{CTC version 4 Grade Direction Data}
\format{
An object of class \code{data.frame} with 35 rows and 3 columns.
}
\usage{
ctcv4_dir
}
\description{
CTC version 4 Grade Direction Data
}
\keyword{datasets}
