% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctcv5_dir}
\alias{ctcv5_dir}
\title{CTC version 5 Grade Direction Data}
\format{
An object of class \code{data.frame} with 35 rows and 3 columns.
}
\usage{
ctcv5_dir
}
\description{
CTC version 5 Grade Direction Data
}
\keyword{datasets}
