% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_by_group}
\alias{cut_by_group}
\title{Cut by group}
\usage{
cut_by_group(df, col_data, col_group, group, new_col, as_factor = FALSE)
}
\arguments{
\item{df}{(\code{dataframe}) with a column of data to be cut and a column specifying the group of each observation.}

\item{col_data}{(\code{string}) the column containing the data to be cut.}

\item{col_group}{(\code{string}) the column containing the names of the groups according to which the data should be
split.}

\item{group}{(\verb{nested list}) providing for each parameter value that should be analyzed in a categorical way: the
name of the parameter (\code{string}), a series of breakpoints (\code{vector}) where the first breakpoints is typically
\code{-Inf} and the last \code{Inf}, and a series of name which will describe each category (\code{vector}).}

\item{new_col}{(\code{string}) the name of the new column in which the cut label should he stored.}

\item{as_factor}{(\code{logical}) if TRUE, the new column is of type \code{factor} else \code{character}.}
}
\description{
Cut by group
}
\examples{
group <- list(
  list(
    "Dose administered during constant dosing interval",
    c(-Inf, 700, 900, 1200, Inf),
    c("<700", "700-900", "900-1200", ">1200")
  ),
  list(
    "Total dose administered",
    c(-Inf, 5000, 7000, 9000, Inf),
    c("<5000", "5000-7000", "7000-9000", ">9000")
  )
)

library(scda)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adex <- sd$adex \%>\%
  dplyr::mutate(ANL01FL = "Y")

adex_gp <- cut_by_group(adex, "AVAL", "PARAM", group, "AVAL_gp")

head(adex_gp[, c("PARAM", "AVAL", "AVAL_gp")])
}
