% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{dataset}
\alias{dataset}
\alias{dataset,chevron_tlg-method}
\alias{dataset<-}
\alias{dataset<-,chevron_tlg-method}
\title{Data Set}
\usage{
dataset(x)

\S4method{dataset}{chevron_tlg}(x)

dataset(x) <- value

\S4method{dataset}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{value}{(\code{character} or \code{NULL}) defining the dataset slot.}
}
\value{
the \code{character} stored in the \code{dataset} slot of the \code{x} argument.
}
\description{
retrieve or set \code{dataset} slot.
}
