% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{datasets}
\alias{datasets}
\alias{datasets,chevron_tlg-method}
\alias{datasets<-}
\alias{datasets<-,chevron_tlg-method}
\title{Retrieve Data Sets}
\usage{
datasets(x)

\S4method{datasets}{chevron_tlg}(x)

datasets(x) <- value

\S4method{datasets}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{value}{(\code{character}) representing the name of the table from an \code{ADaM} dataset required for \code{tlg} creation.}
}
\description{
Retrieve Data Sets

Set Data Sets
}
