% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{deparse_print}
\alias{deparse_print}
\title{Deparse print}
\usage{
deparse_print(x, indent, max_line = getOption("chevron.arg_max_line", 5L))
}
\description{
Deparse print
}
\keyword{internal}
