% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_explicit_na.R
\name{dm_explicit_na}
\alias{dm_explicit_na}
\title{Encode Categorical Missing Values in a \code{DM} Object}
\usage{
dm_explicit_na(
  data,
  omit_tables = NULL,
  omit_columns = NULL,
  char_as_factor = TRUE,
  logical_as_factor = FALSE,
  na_level = "<Missing>"
)
}
\arguments{
\item{data}{(\code{dm}) object to be transformed.}

\item{omit_tables}{(\code{character}) the names of the table to omit from processing.}

\item{omit_columns}{(\code{character}) the names of the columns to omit from processing.}

\item{char_as_factor}{(\code{logical}) should character columns be transformed into factor.}

\item{logical_as_factor}{(\code{logical}) should logical columns be transformed into factor.}

\item{na_level}{(\code{character}) the label to encode missing levels.}
}
\value{
\code{dm} object with explicit missing levels
}
\description{
Encode Categorical Missing Values in a \code{DM} Object
}
\details{
This is a helper function to encode missing entries across groups of categorical variables in potentially
all tables of a \code{dm} object.
}
\examples{
library(dm)

df1 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1")),
  "logi" = c(NA, FALSE, TRUE, NA, FALSE, NA)
)
df2 <- data.frame(
  "char" = c("a", "b", NA, "a", "k", "x"),
  "fact" = factor(c("f1", "f2", NA, NA, "f1", "f1")),
  "num" = 1:6
)

db <- dm(df1, df2)

dm_fact <- dm_explicit_na(db)
dm_fact$df1
dm_fact$df2
}
