% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dm_unite}
\alias{dm_unite}
\title{Unite Columns of a Table in a \code{dm} object.}
\usage{
dm_unite(adam_db, dataset, cols, sep = ".", new = NULL)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{cols}{(\code{character}) the name of the columns to unite.}

\item{sep}{(\code{string}) the separator for the new column name.}

\item{new}{(\code{string}) the name of the new column. If \code{NULL} the concatenation of \code{cols} separated by \code{sep} is used.}
}
\value{
\code{dm} object with a united column.
}
\description{
Unite Columns of a Table in a \code{dm} object.
}
\examples{
x <- dm_unite(dm::dm_nycflights13(), "airlines", c("carrier", "name"), new = "FUSION")
x$airlines
}
\keyword{internal}
