% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml01.R
\docType{data}
\name{dml01_main}
\alias{dml01_main}
\alias{dml01_pre}
\alias{dml01}
\title{\code{DML01} Listing 1 (Default) Demographics and Baseline Characteristics.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
dml01_main(
  adam_db,
  dataset = "adsl",
  arm_var = "ARM",
  disp_cols = c("ID", "ASR", "ETHNIC"),
  ...
)

dml01_pre(adam_db, arm_var = "ARM", dataset = "adsl", ...)

dml01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{...}{not used.}
}
\description{
\code{DML01} Listing 1 (Default) Demographics and Baseline Characteristics.
}
\section{Functions}{
\itemize{
\item \code{dml01_main()}: Main TLG function

\item \code{dml01_pre()}: Preprocessing

}}
\examples{
res <- run(dml01, syn_data)
}
\keyword{datasets}
