% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml02.R
\docType{data}
\name{dml02_main}
\alias{dml02_main}
\alias{dml02_pre}
\alias{dml02}
\title{\code{DML02} Listing 1 (Default) Randomization.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
dml02_main(
  adam_db,
  dataset = "adsl",
  arm_var = "ARM",
  disp_cols = c("ID", arm_var, "RANDDT", "TRTSDT"),
  ...
)

dml02_pre(adam_db, dataset = "adsl", arm_var = "ARM", ...)

dml02
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{...}{not used.}
}
\description{
\code{DML02} Listing 1 (Default) Randomization.
}
\section{Functions}{
\itemize{
\item \code{dml02_main()}: Main TLG function

\item \code{dml02_pre()}: Preprocessing

}}
\examples{
res <- run(dml02, syn_data)
}
\keyword{datasets}
