% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\docType{data}
\name{dmt01_1}
\alias{dmt01_1}
\title{\code{DMT01_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
dmt01_1
}
\description{
\code{DMT01_1}
}
\seealso{
\code{\link[=dmt01_1_main]{dmt01_1_main()}}
}
\keyword{datasets}
