% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\name{dmt01_1_lyt}
\alias{dmt01_1_lyt}
\title{DMT01 Layout 1 (Default)}
\usage{
dmt01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$summaryvars,
  summaryvars_lbls = summaryvars,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DMT01"),
  .study = list(armvar = "ARM", summaryvars = c("AAGE", "AGEGR1", "SEX", "ETHNIC",
    "RACE", "BWGHTSI"), overall_col = "All Patients")
)
}
\arguments{
\item{summaryvars}{(\verb{vector of strings}) variables summarized in demographic table.}

\item{summaryvars_lbls}{(\verb{vector of strings}) labels corresponding to the analyzed variables.}
}
\value{

}
\description{
DMT01 Layout 1 (Default)
}
\examples{
dmt01_1_lyt(armvar = "ACTARM",
summaryvars =  c("AAGE", "SEX", "ETHNIC"),
summaryvars_lbls = c("AGE", "SEX", "ETHNICITY"))
}
