% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\name{dmt01_lyt}
\alias{dmt01_lyt}
\title{\code{dmt01} Layout}
\usage{
dmt01_lyt(arm_var, lbl_overall, summaryvars, summaryvars_lbls)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{summaryvars_lbls}{(\code{character}) labels corresponding to the analyzed variables.}
}
\description{
\code{dmt01} Layout
}
\keyword{internal}
