% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\docType{data}
\name{dose_change_rule}
\alias{dose_change_rule}
\title{Dose Change Rule}
\format{
An object of class \code{rule} (inherits from \code{character}) of length 9.
}
\usage{
dose_change_rule
}
\description{
Dose Change Rule
}
\keyword{datasets}
