% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_1_lyt}
\alias{dst01_1_lyt}
\title{DST01 Layout (Default)}
\usage{
dst01_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS")
)
}
\arguments{
\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{status}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a variable
name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis period.}

\item{reason}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS}, however can
also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\value{

}
\description{
DST01 Layout (Default)
}
\examples{
dst01_1_lyt(armvar = "ACTARM",
status = "EOP01STT",
reason = "DCP01RS")
}
