% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_2}
\alias{dst01_2}
\title{Title}
\usage{
dst01_2(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(armvar = "ACTARM", lbl_overall = "All patients")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
Title
}
\details{
\itemize{
\item Non-standard
}
}
\examples{

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

dst01_2(adsl, adae)
dst01_2(adsl, adae, lbl_overall = "All patients")
}
