% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_2_main}
\alias{dst01_2_main}
\alias{dst01_2_lyt}
\alias{dst01_2_pre}
\title{DST01 Table 2 (Supplementary) Patient Disposition Table 2.}
\usage{
dst01_2_main(
  adam_db,
  armvar = .study$planarm,
  status_var = .study$status_var,
  disc_reason_var = .study$disc_reason_var,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", disc_reason_var =
    "DCSREAS", status_var = "EOSSTT")
)

dst01_2_lyt(
  armvar = .study$planarm,
  status = .study$status,
  disc_reason_var = .study$disc_reason_var,
  completed_lbl = "COMPLETED",
  ongoing_lbl = "ONGOING",
  discontinued_lbl = "DISCONTINUED",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", status = "EOSSTT",
    disc_reason_var = "DCSREAS")
)

dst01_2_pre(
  adam_db,
  reason = .study$disc_reason_var,
  .study = list(disc_reason_var = "DCSREAS"),
  ...
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting.}

\item{status_var}{(\code{character}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a
variable name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{disc_reason_var}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS},
however can also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{status}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a variable
name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis period.}

\item{completed_lbl}{(\code{string}) associated with completed study and found in the columns given by \code{status}. By
Default \code{COMPLETED}.}

\item{ongoing_lbl}{(\code{string}) associated with ongoing treatment and found in the columns given by \code{status_treatment}.
By Default `ONGOING.}

\item{discontinued_lbl}{(\code{string}) associated with discontinued study and found in the columns given by \code{status}. By
Default \code{DISCONTINUED}.}

\item{reason}{(\code{string}) the variable name for variable with the reason for discontinuation.}

\item{...}{not used.}
}
\description{
The DST01_2 Disposition Table provides an overview of
patients study completion. For patients who discontinued the study, the reason provided is categorized as "Safety" or
"Non-Safety" issue.
}
\details{
\itemize{
\item Non-standard disposition table summarizing the reasons for patient withdrawal.
\item Withdrawal reasons are grouped into Safety and Non-Safety issues.
\item Safety issues include Death and Adverse events.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows.
\item Split columns by arm.
\item Include a total column by default.
\item Sort withdrawal reasons by alphabetic order.
}

Since the two parts of the tables are pruned differently, the layout function returns a list of layouts,
which allows the tables to be constructed and pruned separately before binding.
}
\section{Functions}{
\itemize{
\item \code{dst01_2_lyt}: \code{dst01_2} Layout

\item \code{dst01_2_pre}: \code{dst01_2} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  dst01_2_pre()

dst01_2_main(db)
dst01_2_main(db, lbl_overall = NULL)
dst01_2_lyt(
  armvar = "ACTARM",
  status = "EOP01STT",
  disc_reason_var = "DCP01RS"
)
syn_test_data() \%>\%
  dst01_2_pre()
}
