% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_3}
\alias{dst01_3}
\alias{dst01_3_lyt}
\title{DST01 Table 3 (Supplementary) Patient Disposition Table 3}
\usage{
dst01_3(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = "EOSSTT",
  reason = "DCSREAS",
  prune_0 = TRUE,
  status_treatment = "EOTSTT",
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients")
)

dst01_3_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status_treatment = .study$status,
  deco = std_deco("DST01"),
  completed_lbl = "COMPLETED",
  discontinued_lbl = "DISCONTINUED",
  ongoing_lbl = "ONGOING",
  .study = list(armvar = "ARM", lbl_overall = "All patients", status_treatment =
    "EOTSTT")
)
}
\arguments{
\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{status}{(\code{character}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a
variable name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{reason}{(\code{character}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS}, however can
also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{status_treatment}{(\code{string}) variable used to define the treatment status of the patients. Default is \code{EOTSTT},
however can also be a variable with the pattern \code{EOTxxSTT} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{completed_lbl}{(\code{string}) associated with completed treatment and found in the columns given by
\code{status_treatment}. By Default \code{COMPLETED}.}

\item{discontinued_lbl}{(\code{string}) associated with discontinued treatment and found in the columns given by
\code{status_treatment}. By Default \code{DISCONTINUED}.}

\item{ongoing_lbl}{(\code{string}) associated with ongoing treatment and found in the columns given by \code{status_treatment}.
By Default `ONGOING.}
}
\value{

}
\description{
The DST01_3 Disposition Table provides an overview of patients study treatment status.
For patients who discontinued the study, the reason provided is categorized as "Safety" or "Non-Safety" issue.
}
\details{
\itemize{
\item Non-standard disposition table summarizing the reasons for patient withdrawal and treatment status.
\item Withdrawal reasons are grouped into Safety and Non-Safety issues.
\item Safety issues include Death and Adverse Events.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows.
\item Split columns by arm.
\item Include a total column by default.
\item Sort withdrawal reasons by alphabetic order.
}
}
\section{Functions}{
\itemize{
\item \code{dst01_3_lyt}: 
}}

\examples{
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
   dm_select_tbl(adsl)

db <- db \%>\%
  (std_filter("dst01_3"))() \%>\%
  (std_mutate("dst01_3"))()

dst01_3(db)
dst01_3(db, lbl_overall = "")

dst01_3_lyt(armvar = "ACTARM",
status_treatment = "EOTxx01")
}
