% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_3_lyt}
\alias{dst01_3_lyt}
\title{DST03 Layout (Supplementary)}
\usage{
dst01_3_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status_treatment = .study$status,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status_treatment =
    "EOTSTT")
)
}
\arguments{
\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{status_treatment}{(\code{string}) variable used to define the treatment status of the patients. Default is \code{EOTSTT},
however can also be a variable with the pattern \code{EOTxxSTT} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\value{

}
\description{
DST03 Layout (Supplementary)
}
\examples{
dst01_3_lyt(armvar = "ACTARM",
status_treatment = "EOTxx01")
}
