% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_lyt}
\alias{dst01_lyt}
\title{dst01 Layout}
\usage{
dst01_lyt(arm_var, study_status_var, detail_vars, trt_status_var, lbl_overall)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{study_status_var}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a
variable name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{detail_vars}{Named (\code{list}) of grouped display of \code{study_status_var}.}

\item{trt_status_var}{(\code{string}) variable of treatment status.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
dst01 Layout
}
\keyword{internal}
