% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtht01.R
\name{dtht01_2}
\alias{dtht01_2}
\alias{dtht01_2_lyt}
\title{\code{DTHT01} Table 2 (Supplementary) Death Table.}
\usage{
dtht01_2(
  adam_db,
  armvar = .study$armvar,
  add_time_to_event = TRUE,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DTHT01"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

dtht01_2_lyt(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DTHT01"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{add_time_to_event}{(\code{logical}) should the time to event information be displayed.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
A description of the causes of death with post-study reporting of death.
}
\details{
\itemize{
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Does not include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{dtht01_2_lyt}: 
}}

\examples{
library(tern)
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
 dm_zoom_to(adsl) \%>\%
 mutate(DTHCAT = explicit_na(DTHCAT)) \%>\%
 mutate(LDDTHGR1 = explicit_na(LDDTHGR1)) \%>\%
 mutate(DTHCAUS = explicit_na(DTHCAUS)) \%>\%
 dm_update_zoomed() \%>\%
 preprocess_data("dtht01_2")

dtht01_2(adam_db = db)

dtht01_2_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
}
