% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_template.R
\name{dummy_template}
\alias{dummy_template}
\title{Dummy template.}
\usage{
dummy_template
}
\description{
This template creates a dummy output.
}
\examples{
run(dummy_template, syn_data)
}
