% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt01.R
\docType{data}
\name{egt01_1}
\alias{egt01_1}
\title{\code{EGT01_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
egt01_1
}
\description{
\code{EGT01_1}
}
\seealso{
\code{\link[=egt01_1_main]{egt01_1_main()}}
}
\keyword{datasets}
