% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt02.R
\name{egt02_lyt}
\alias{egt02_lyt}
\title{\code{egt02} Layout}
\usage{
egt02_lyt(
  arm_var = "ACTARM",
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall,
  exclude_base_abn
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{string}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{string}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{exclude_base_abn}{(\code{flag}) whether to exclude subjects with baseline abnormality from numerator and
denominator.}
}
\description{
\code{egt02} Layout
}
\keyword{internal}
