% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt03.R
\name{egt03_lyt}
\alias{egt03_lyt}
\title{egt03 Layout}
\usage{
egt03_lyt(arm_var, splitvar, summaryvar, lbl_armvar, lbl_summaryvars)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{splitvar}{(\code{character}) variables to be analyzed, typically \code{"ANRIND"}. Labels of the corresponding columns are
used as subtitles.}

\item{summaryvar}{(\code{character}) variables to be analyzed, typically \code{"BNRIND"}. Labels of the corresponding columns
are used as subtitles.}

\item{lbl_armvar}{(\code{string}) label of the \code{arm_var} variable.}

\item{lbl_summaryvars}{(\code{string}) label of the \code{summaryvar} variable.}
}
\description{
egt03 Layout
}
\keyword{internal}
