% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt05_qtcat.R
\docType{data}
\name{egt05_qtcat_1_main}
\alias{egt05_qtcat_1_main}
\alias{egt05_qtcat_1_lyt}
\alias{egt05_qtcat_1_pre}
\alias{egt05_qtcat_1_post}
\alias{egt05_qtcat_1}
\title{\code{EGT05_QTCAT} Table 1 (Default) ECG Actual Values and Changes from Baseline by Visit Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt05_qtcat_1_main(
  adam_db,
  armvar = "ACTARM",
  summaryvars = c(`Value at Visit` = "AVALCAT1", `Change from Baseline` = "CHGCAT1"),
  lbl_overall = NULL,
  visitvar = "AVISIT",
  deco = std_deco("EGT05_QTCAT"),
  lbl_cat = "Category",
  lbl_headvisit = "Analysis Visit",
  ...
)

egt05_qtcat_1_lyt(
  armvar,
  summaryvars,
  summaryvars_lbls,
  lbl_overall = NULL,
  visitvar,
  lbl_avisit,
  deco,
  lbl_cat,
  lbl_headvisit,
  ...
)

egt05_qtcat_1_pre(adam_db, ...)

egt05_qtcat_1_post(tlg, prune_0 = TRUE, ...)

egt05_qtcat_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. \code{AVALCAT1} and \code{CHGCAT1} by default.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{lbl_cat}{(\code{character}) label of the Category of \code{summaryvars} variable. Default as \code{Category}.}

\item{lbl_headvisit}{(\code{character}) label of Visits in the header. Default as \verb{Analysis Visit}.}

\item{...}{not used.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{lbl_avisit}{(\code{character}) label of the \code{visitvar} variable.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
The \code{EGT05_QTCAT} table 1 summarizes
several electrocardiogram parameters and their evolution throughout the study.
}
\details{
\itemize{
\item The \verb{Value at Visit} column, displays the categories of QT value for patients, "<=450 msec", ">450 to <=480 msec",
">480 to <= 500 msec", ">500 msec", and "\if{html}{\out{<Missing>}}" for each visit.
\item The \verb{Change from Baseline} column, displays the categories of QT value change from baseline for patients,
"<=30 msec", ">30 to <=60 msec", ">60 msec", and "\if{html}{\out{<Missing>}}"
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted based on factor level; by chronological time point given by one of \code{"AVISIT"} (Default) or \code{"ATPTN"}.
Re-level to customize order.
}
}
\section{Functions}{
\itemize{
\item \code{egt05_qtcat_1_main()}: Main TLG function

\item \code{egt05_qtcat_1_lyt()}: Layout

\item \code{egt05_qtcat_1_pre()}: Preprocessing

\item \code{egt05_qtcat_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adeg} table with a \code{"PARAM"} column contains 'QT' as well as column
specified in \code{visitvar}.
For \code{summaryvars}, if \code{AVALCAT1} and \code{CHGCAT1} columns are not existed in input data sets, \code{AVAL} and \code{CHG}
columns must be contained to re-derive \code{AVALCAT1} and \code{CHGCAT1}.
}
}
\examples{
db <- syn_data
run(egt05_qtcat_1, db)
run(egt05_qtcat_1, db, summaryvars = c("AVALCAT1", "Change" = "CHGCAT1"))
}
\keyword{datasets}
