% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\docType{data}
\name{empty_rule}
\alias{empty_rule}
\title{Empty rule}
\format{
An object of class \code{empty_rule} (inherits from \code{rule}, \code{character}) of length 0.
}
\usage{
empty_rule
}
\description{
Empty rule
}
\keyword{datasets}
