% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_with_args}
\alias{execute_with_args}
\title{Execute function with given arguments}
\usage{
execute_with_args(fun, ...)
}
\description{
Execute function with given arguments
}
\details{
If the function has \code{...}, this function will not pass other arguments to \code{...}.
Only named arguments are passed.
}
\keyword{internal}
