% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\docType{data}
\name{ext01_1}
\alias{ext01_1}
\title{\code{EXT01_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
ext01_1
}
\description{
\code{EXT01_1}
}
\seealso{
\code{\link[=ext01_1_main]{ext01_1_main()}}
}
\keyword{datasets}
