% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\docType{data}
\name{ext01_2}
\alias{ext01_2}
\title{\code{EXT01_2}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
ext01_2
}
\description{
\code{EXT01_2}
}
\seealso{
\code{\link[=ext01_2_main]{ext01_2_main()}}
}
\keyword{datasets}
