% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\name{ext01_2_main}
\alias{ext01_2_main}
\alias{ext01_2_lyt}
\alias{ext01_2_pre}
\title{EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options}
\usage{
ext01_2_main(
  adam_db,
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

ext01_2_lyt(
  armvar = .study$actualarm,
  summaryvars = c("AVAL", "AVALCAT1"),
  summaryvars_lbls = c("Summary", "Categories"),
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

ext01_2_pre(
  adam_db,
  show_stats = .study$show_cont_stats,
  show_bins = .study$show_cat_stats,
  .study = list(show_cont_stats = c("ALL"), show_cat_stats = c("ALL")),
  ...
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{summaryvars}{(\code{string}) the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{summaryvars_lbls}{(\code{string}) the label associated with the analyzed variable.}

\item{show_stats}{(\verb{vector of character}) providing the name of the parameters whose statistical summary should be
presented. To analyze all, provide \code{show_stats = "ALL"} (Default), to analyze none, provide \code{show_stats = ""}.}

\item{show_bins}{(\verb{vector of character}) providing the name of the parameters whose categorical summary should be
presented. To analyze all, provide \code{show_bins = "ALL"} (Default), to analyze none, provide \code{show_bins = ""}.}

\item{...}{not used.}
}
\description{
EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options
}
\details{
\itemize{
\item Supplementary Exposure table with binning of desired analysis values.
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
\item \code{ANL01FL} is not relevant subset
}
}
\section{Functions}{
\itemize{
\item \code{ext01_2_lyt}: \code{ext01_2} Layout

\item \code{ext01_2_pre}: \code{ext01_2} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  ext01_2_pre()

ext01_2_main(db)
syn_test_data() \%>\%
  ext01_2_pre()
}
