% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_relevel_params}
\alias{fct_relevel_params}
\title{Reorder PARAM and PARAMCD Levels Simultaneously}
\usage{
fct_relevel_params(df, paramcd_levels)
}
\arguments{
\item{df}{data.frame with PARAM and PARAMCD variables}

\item{paramcd_levels}{ordered levels of PARAMCD}
}
\description{
Reorder PARAM and PARAMCD Levels Simultaneously
}
\examples{

df <- data.frame(PARAMCD = factor(c("A", "B", "C")), PARAM = factor(paste("letter", LETTERS[1:3])))

fct_relevel_params(df, paramcd_levels = c("B", "A", "C"))

fct_relevel_params(df, paramcd_levels = c("letter B", "letter A", "LETTER C"))
fct_relevel_params(df, paramcd_levels = c("letter B", "letter A"))

}
