% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{filter_adcm_anl01fl}
\alias{filter_adcm_anl01fl}
\title{Filter \code{adcm} for \code{ANL01FL} and  \code{ATIREL}}
\usage{
filter_adcm_anl01fl(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
Filter \code{adcm} for \code{ANL01FL} and  \code{ATIREL}
}
\details{
filter with \code{ANL01FL} (instead of \code{SAFFL } which is external to \code{chevron}).
}
