% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{filter_adcm_cmt01}
\alias{filter_adcm_cmt01}
\title{Filter \code{adcm} for \code{ANL01FL} and  \code{SAFFL}}
\usage{
filter_adcm_cmt01(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
Filter \code{adcm} for \code{ANL01FL} and  \code{SAFFL}
}
\details{
filter with \code{ANL01FL} instead of \code{SAFFL } which is external to \code{chevron}.
}
