% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_manipulation.R
\name{filter_adex_drug}
\alias{filter_adex_drug}
\title{Filter \code{adex} for \code{PARCAT1}}
\usage{
filter_adex_drug(x)
}
\arguments{
\item{x}{(\code{dm})}
}
\value{

}
\description{
Filter \code{adex} for \code{PARCAT1}
}
