% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{filter_adlb_anl01fl}
\alias{filter_adlb_anl01fl}
\title{Filter \code{adlb} for \code{ANL01FL}}
\usage{
filter_adlb_anl01fl(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}
}
\description{
Filter \code{adlb} for \code{ANL01FL}
}
