% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{filter_egt02}
\alias{filter_egt02}
\title{Filter post-baseline values in \code{adeg}}
\usage{
filter_egt02(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
Filter post-baseline values in \code{adeg}
}
