% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date}
\alias{format_date}
\title{Formatting of date}
\usage{
format_date(date_format = "\%d\%b\%Y")
}
\arguments{
\item{date_format}{(\code{string}) the output format.}
}
\value{
a \code{function} converting a date into \code{string}.
}
\description{
Formatting of date
}
