% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fuse_sequentially}
\alias{fuse_sequentially}
\title{Fuse list elements}
\usage{
fuse_sequentially(x, y)
}
\arguments{
\item{x}{(\code{list}) to fuse.}

\item{y}{(\code{list}) to fuse. Elements with names already existing in \code{x} are discarded.}
}
\description{
Fuse list elements
}
\keyword{internal}
