% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_adam_datasets}
\alias{get_adam_datasets}
\alias{get_adam_datasets,chevron_tlg-method}
\title{Retrieve Names of Data Sets Associated with the Pipeline}
\usage{
get_adam_datasets(object, ...)

\S4method{get_adam_datasets}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Names of Data Sets Associated with the Pipeline

get_adam_datasets
}
