% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ctc}
\alias{get_ctc}
\title{Get ctc grade data by version}
\usage{
get_ctc(version = 4)
}
\arguments{
\item{version}{(\code{number}) version of ctc grading.}
}
\description{
Get ctc grade data by version
}
