% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_datasets}
\alias{get_datasets}
\alias{get_datasets,tlg_pipeline_s4-method}
\title{Retrieve Names of Data Sets Associated with the Pipeline}
\usage{
get_datasets(object, ...)

\S4method{get_datasets}{tlg_pipeline_s4}(object)
}
\arguments{
\item{object}{(\code{tlg_pipeline_s4}) input.}

\item{...}{not used.}
}
\description{
Retrieve Names of Data Sets Associated with the Pipeline

get_datasets
}
