% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_db_data}
\alias{get_db_data}
\title{Get Data from a DB}
\usage{
get_db_data(db, ...)
}
\arguments{
\item{db}{(named \code{list}) to extract elements form.}

\item{...}{not used.}
}
\value{
named list of data sets.
}
\description{
Get Data from a DB
}
\examples{
\dontrun{
get_db_data(list(iris = iris, mtcars = mtcars, CO2 = CO2))
get_db_data(list(iris = iris, mtcars = mtcars, CO2 = CO2), "iris")
get_db_data(list(iris = iris, mtcars = mtcars, CO2 = CO2), "iris", "CO2")
}
}
