% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gdsr_lab}
\alias{get_gdsr_lab}
\title{Get the label of an adam data set variable}
\usage{
get_gdsr_lab(x, var)
}
\arguments{
\item{x}{(\code{character}) the name of a data set.}

\item{var}{(\code{character}) the name of the variable.}
}
\value{
the label associated with \code{var} for the desired adam dataset.
}
\description{
Get the label of an adam data set variable
}
\examples{
get_gdsr_lab("adlb", "AVAL")
}
