% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\name{get_grade_rule}
\alias{get_grade_rule}
\title{Get grade rule}
\usage{
get_grade_rule(direction = "high", missing = "incl")
}
\arguments{
\item{direction}{(\code{string}) of abnormality direction.}

\item{missing}{(\code{string}) method to deal with missing}
}
\description{
Get grade rule
}
