% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labels}
\alias{get_labels}
\title{Get Labels}
\usage{
get_labels(df, x)
}
\arguments{
\item{df}{(\code{data.frame}).}

\item{x}{(\code{character}) the names of column in the \code{df} data frame.}
}
\value{
the name of each element in the \code{x} vector where missing names are replaced with the value of the label
attribute for the corresponding column in the \code{df} data frame.
}
\description{
Get Labels
}
\keyword{internal}
