% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{get_lyt}
\alias{get_lyt}
\alias{get_lyt,chevron_tlg-method}
\title{Retrieve Layout}
\usage{
get_lyt(object, ...)

\S4method{get_lyt}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Layout
}
