% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_main}
\alias{get_main}
\alias{get_main,tlg_pipeline_S4-method}
\title{Retrieve Main Function}
\usage{
get_main(object, ...)

\S4method{get_main}{tlg_pipeline_S4}(object)
}
\arguments{
\item{object}{(\code{tlg_pipeline_S4}) input.}

\item{...}{not used.}
}
\description{
Retrieve Main Function

get_main
}
