% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_postprocess}
\alias{get_postprocess}
\alias{get_postprocess,tlg_pipeline_s4-method}
\title{Retrieve Post Processing Function}
\usage{
get_postprocess(object, ...)

\S4method{get_postprocess}{tlg_pipeline_s4}(object)
}
\arguments{
\item{object}{(\code{tlg_pipeline_s4}) input.}

\item{...}{not used.}
}
\description{
Retrieve Post Processing Function

get_postprocess
}
