% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{get_preprocess}
\alias{get_preprocess}
\alias{get_preprocess,chevron_tlg-method}
\title{Retrieve pre-processing Function}
\usage{
get_preprocess(object)

\S4method{get_preprocess}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}
}
\description{
Retrieve pre-processing Function
}
