% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rlp_match.R
\name{get_rlp_match}
\alias{get_rlp_match}
\title{Get data for a single match on Rugby League Project}
\usage{
get_rlp_match(rlp_match_id)
}
\arguments{
\item{rlp_match_id}{An integer that represents a match ID found on Rugby League Project.}
}
\value{
A list containing match, team and player information.
}
\description{
Get data for a single match on Rugby League Project
}
\examples{
get_rlp_match(39999)
}
