% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rlp_seasons.R
\name{get_rlp_seasons}
\alias{get_rlp_seasons}
\title{Get fixture and result data from Rugby League Project}
\usage{
get_rlp_seasons(rlp_competition_code, years)
}
\arguments{
\item{rlp_competition_code}{A string representing the competition, as used on Rugby League Project.}

\item{years}{A vector of integers that represent the first year of the seasons wanted.}
}
\value{
A \code{tibble} with each row aligning to a match
}
\description{
Get fixture and result data from Rugby League Project
}
\examples{
get_rlp_seasons("nrl", 2019)
}
