% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_section_div}
\alias{get_section_div}
\title{Get Section dividers}
\usage{
get_section_div()
}
\value{
(\code{character}) value with section dividers at corresponding section.
}
\description{
Get Section dividers
}
