% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tc_seasons.R
\name{get_tc_seasons}
\alias{get_tc_seasons}
\title{Get NRL fixture and result data from The Cruncher}
\usage{
get_tc_seasons(years)
}
\arguments{
\item{years}{A vector of integers that represent the year of the seasons wanted (2017- available)}
}
\value{
A \code{tibble} with each row aligning to a match
}
\description{
Get NRL fixture and result data from The Cruncher
}
\examples{
get_tc_seasons(2017:2023)
}
