% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_x_hjust}
\alias{get_x_hjust}
\title{Get a harmonious value of horizontal justification for x axis}
\usage{
get_x_hjust(x)
}
\arguments{
\item{x}{(\code{numeric}) angle between -90 and 90 degree.}
}
\description{
Get a harmonious value of horizontal justification for x axis
}
\keyword{internal}
