% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_x_vjust}
\alias{get_x_vjust}
\title{Get a harmonious value of vertical justification for x axis}
\usage{
get_x_vjust(x)
}
\arguments{
\item{x}{(\code{numeric}) angle between -90 and 90 degree.}
}
\description{
Get a harmonious value of vertical justification for x axis
}
\keyword{internal}
