% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_list}
\alias{gg_list}
\title{List of \code{gg} object}
\usage{
gg_list(...)
}
\arguments{
\item{...}{(\code{ggplot}) objects.}
}
\value{
a \code{gg_list} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
