% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_theme_chevron}
\alias{gg_theme_chevron}
\title{Theme for Chevron Plot}
\usage{
gg_theme_chevron(
  grid_y = TRUE,
  grid_x = FALSE,
  legend_position = "top",
  text_axis_x_rot = TRUE
)
}
\arguments{
\item{grid_y}{(\code{flag}) should horizontal grid be displayed.}

\item{grid_x}{(\code{flag}) should vertical grid be displayed.}

\item{legend_position}{(\code{string}) the position of the legend.}

\item{text_axis_x_rot}{(\code{flag}) should the text of the x axis be rotated.}
}
\value{
a \code{theme} object.
}
\description{
Theme for Chevron Plot
}
