% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_g-S4methods.R
\name{graph}
\alias{graph}
\alias{graph,chevron_tlg-method}
\alias{graph<-}
\alias{graph<-,chevron_g-method}
\title{Graph}
\usage{
graph(x)

\S4method{graph}{chevron_tlg}(x)

graph(x) <- value

\S4method{graph}{chevron_g}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_g}) input.}

\item{value}{(\code{function}) returning a \code{ggplot}. Typically one of the \verb{_plot} function of \code{chevron}.}
}
\description{
retrieve or set \code{graph} function.
}
