% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grob_list}
\alias{grob_list}
\title{List of \code{grob} object}
\usage{
grob_list(...)
}
\arguments{
\item{...}{(\code{grob}) objects.}
}
\value{
a \code{grob_list} object.
}
\description{
List of \code{grob} object
}
