% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_explicit_na.R
\name{h_as_factor}
\alias{h_as_factor}
\title{Helper Coercion Function to Factor with explicit missing levels}
\usage{
h_as_factor(x, na_label)
}
\arguments{
\item{x}{(\code{character} or \code{logical}) input to be turned into factor with explicit missing level.}

\item{na_label}{(\code{character}) the label to encode missing levels.}
}
\value{

}
\description{
Helper Coercion Function to Factor with explicit missing levels
}
\details{
This function preserves labels.
}
