% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\name{h_chevron_tlg}
\alias{h_chevron_tlg}
\title{\code{chevron_tlg} constructor helper}
\usage{
h_chevron_tlg(
  main = function(adam_db, lyt_ls, ...) adam_db,
  lyt = list(function(...) basic_table()),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_
)
}
\description{
handle missing arguments for the \code{chevron_tlg} constructor.
}
