% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_pad_or_round}
\alias{h_pad_or_round}
\title{Padding or Rounding Numbers}
\usage{
h_pad_or_round(n, digits = NA, ...)
}
\arguments{
\item{digits}{(\code{integer}) number of digits.}

\item{x}{(\code{numeric}) to modify.}
}
\description{
Padding or Rounding Numbers
}
\examples{
h_pad_or_round(123.1234, 10)
h_pad_or_round(123.1234, 1)
h_pad_or_round(123, 1)
h_pad_or_round(123, 3)
h_pad_or_round(123, 0)
}
