% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_pad_or_round_dev}
\alias{h_pad_or_round_dev}
\title{Padding or Rounding Numbers and Deviation}
\usage{
h_pad_or_round_dev(x, digits = NA, ...)
}
\arguments{
\item{x}{(\code{numeric}) representing the main value and its error.}

\item{digits}{(\code{integer}) number of digits.}

\item{...}{not used.}
}
\description{
Padding or Rounding Numbers and Deviation
}
\examples{
h_pad_or_round_dev(c(1.11111, 0.5), 3)
}
