% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_pad_or_round_pct}
\alias{h_pad_or_round_pct}
\title{Padding or Rounding Numbers Error as Percentage}
\usage{
h_pad_or_round_pct(x, digits = NA, ...)
}
\arguments{
\item{x}{(\code{numeric}) representing the main value and its error.}

\item{digits}{(\code{integer}) number of digits.}
}
\description{
Padding or Rounding Numbers Error as Percentage
}
\examples{
h_pad_or_round_pct(c(1.11111, 0.5), 3)
}
