% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_pad_or_round_sep}
\alias{h_pad_or_round_sep}
\title{Padding or Rounding Numbers with Separator}
\usage{
h_pad_or_round_sep(x, digits = NA, sep = " - ", ...)
}
\arguments{
\item{x}{(\code{numeric}) to format.}

\item{digits}{(\code{integer}) number of digits.}

\item{sep}{(\code{string}) separating the formatted values.}
}
\description{
Padding or Rounding Numbers with Separator
}
\examples{
h_pad_or_round_sep(c(1.11111, 2.22222), 3)
}
