% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_relabel}
\alias{h_relabel}
\title{h_relabel}
\usage{
h_relabel(data, lab)
}
\arguments{
\item{data}{any object that can take a label attribute.}

\item{lab}{(\code{character}) to be used as label.}
}
\value{

}
\description{
Utility function to relabel column in \code{dplyr} context.
}
