% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{h_unwrap_layout}
\alias{h_unwrap_layout}
\title{Helper Function Extracting Layout Functions}
\usage{
h_unwrap_layout(x, pattern)
}
\description{
Helper Function Extracting Layout Functions
}
\keyword{internal}
