% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifneeded_convert_day2month}
\alias{ifneeded_convert_day2month}
\title{Helper function to convert days to months if needed}
\usage{
ifneeded_convert_day2month(data)
}
\description{
Helper function to convert days to months if needed
}
\keyword{internal}
