% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{ifneeded_split_col}
\alias{ifneeded_split_col}
\title{Helper function to add a column split if specified}
\usage{
ifneeded_split_col(lyt, var, overall_col_lbl = NULL)
}
\arguments{
\item{lyt}{(\code{rtables}) object.}

\item{var}{(\code{string}) the name of the variable initiating a new column split.}

\item{overall_col_lbl}{(\code{string}) the label for the overall column.}
}
\value{
\code{rtables} object.
}
\description{
Helper function to add a column split if specified
}
\keyword{internal}
