% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifneeded_split_row}
\alias{ifneeded_split_row}
\title{Helper function to add a row split if specified}
\usage{
ifneeded_split_row(lyt, var, lbl_var)
}
\arguments{
\item{lyt}{(\code{rtables}) object.}

\item{var}{(\code{string}) the name of the variable initiating a new row split.}

\item{lbl_var}{(\code{string})the label of the variable \code{var}.}
}
\value{
\code{rtables} lyt object.
}
\description{
Helper function to add a row split if specified
}
\keyword{internal}
