% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmg01.R
\docType{data}
\name{kmg01_1_main}
\alias{kmg01_1_main}
\alias{kmg01_1_pre}
\alias{kmg01_1_post}
\alias{kmg01_1}
\title{\code{KMG01} Kaplan-Meier Plot 1.}
\format{
An object of class \code{chevron_g} of length 1.
}
\usage{
kmg01_1_main(
  adam_db,
  dataset = "adtte",
  arm_var = "ARM",
  x_name = "Time (Days)",
  y_name = "Survival Probability",
  show_statis = TRUE,
  show_censor = TRUE,
  pval_method = "wald",
  ties = "exact",
  conf_level = 0.95,
  legend_pos = "bottom",
  position_coxph = c(0, 0.05),
  position_surv_med = c(0.9, 0.9),
  line_col = as.list(nestcolor::color_palette()),
  ...
)

kmg01_1_pre(adam_db, dataset, paramcd = "OS", ...)

kmg01_1_post(tlg, ...)

kmg01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{x_name}{(\code{string}) the name of the x-axis.}

\item{y_name}{(\code{string}) the name of the x-axis.}

\item{show_statis}{(\code{flag}) should the summary statistic table be displayed.}

\item{show_censor}{(\code{flag}) should the censor flag be displayed.}

\item{pval_method}{(\code{string}) should the censor flag be displayed.}

\item{ties}{(\code{string}) should the censor flag be displayed.}

\item{conf_level}{(\code{number}) should the censor flag be displayed.}

\item{legend_pos}{(\code{string}) the position of the legend.}

\item{line_col}{(\code{list}) describing the colors to use for the lines or a named \code{list} associating values of \code{arm_var}
with color names.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
\code{KMG01} Kaplan-Meier Plot 1.
}
\details{
\itemize{
\item No overall value.
}
}
\section{Functions}{
\itemize{
\item \code{kmg01_1_main()}: Main TLG Function

\item \code{kmg01_1_pre()}: Preprocessing

\item \code{kmg01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the table specified by \code{dataset} with the columns specified by  \code{arm_var}.
}
}
\examples{
library(dm)
library(dplyr)

col <- c(
  "A: Drug X" = "black",
  "B: Placebo" = "blue",
  "C: Combination" = "gray"
)

run(kmg01_1, syn_data, dataset = "adtte", line_col = col)
}
\keyword{datasets}
