% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{layout}
\alias{layout}
\alias{lyt,chevron_tlg-method}
\alias{lyt<-}
\alias{lyt<-,chevron_tlg-method}
\title{Retrieve Layout}
\usage{
lyt(x, ...)

\S4method{lyt}{chevron_tlg}(x)

lyt(x, ...) <- value

\S4method{lyt}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Layout

Set Layout Function
}
