% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt01.R
\docType{data}
\name{lbt01_1}
\alias{lbt01_1}
\title{\code{LBT01_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
lbt01_1
}
\description{
\code{LBT01_1}
}
\seealso{
\code{\link[=lbt01_1_main]{lbt01_1_main()}}
}
\keyword{datasets}
