% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt04.R
\docType{data}
\name{lbt04_1_main}
\alias{lbt04_1_main}
\alias{lbt04_1_lyt}
\alias{lbt04_1_pre}
\alias{lbt04_1_check}
\alias{lbt04_1_post}
\alias{lbt04_1}
\title{\code{LBT04} Table 1 (Default) Laboratory Abnormalities Not Present at Baseline.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt04_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_param = "Laboratory Test",
  lbl_anrind = "Direction of Abnormality",
  deco = std_deco("LBT04"),
  ...
)

lbt04_1_lyt(arm_var, lbl_param, lbl_anrind, deco, ...)

lbt04_1_pre(adam_db, ...)

lbt04_1_check(adam_db, req_tables = c("adsl", "adlb"), arm_var = "ACTARM", ...)

lbt04_1_post(tlg, ...)

lbt04_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{character}) variable used for column splitting}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}

\item{lbl_anrind}{(\code{character}) label of the \code{ANRIND} variable.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
The \code{LBT04} table provides an
overview of laboratory abnormalities not present at baseline.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Lab test results with missing \code{ANRIND} values are excluded.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{lbt04_1_main()}: Main TLG function

\item \code{lbt04_1_lyt()}: Layout

\item \code{lbt04_1_pre()}: Preprocessing

\item \code{lbt04_1_check()}: Checks

\item \code{lbt04_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns \code{"PARCAT1"}, \code{"PARCAT2"}, \code{"PARAM"}, \code{"ANRIND"},
and column specified by \code{arm_var}.
}
}
\examples{
run(lbt04_1, syn_data)
}
\keyword{datasets}
