% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt04.R
\name{lbt04_lyt}
\alias{lbt04_lyt}
\title{\code{lbt04} Layout}
\usage{
lbt04_lyt(
  arm_var,
  lbl_overall,
  lbl_param,
  lbl_abn_var,
  var_parcat,
  var_param,
  row_split_var,
  row_split_lbl,
  analysis_abn_var,
  variables,
  page_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{lbl_abn_var}{(\code{string}) label of the \code{analysis_abn_var} variable.}

\item{row_split_var}{(\code{character}) additional row split variables.}

\item{variables}{(\code{list}) see \link[tern:count_abnormal]{tern::count_abnormal}}

\item{page_var}{(\code{string}) variable name prior to which the row split is by page.}
}
\description{
\code{lbt04} Layout
}
\keyword{internal}
