% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt05.R
\docType{data}
\name{lbt05_1_main}
\alias{lbt05_1_main}
\alias{lbt05_1_lyt}
\alias{lbt05_1_pre}
\alias{lbt05_1_check}
\alias{lbt05_1_post}
\alias{lbt05_1}
\title{\code{LBT05} Table 1 (Default) Laboratory Abnormalities with Single and Replicated Marked.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt05_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  lbl_param = "Laboratory Test",
  lbl_anrind = "Direction of Abnormality",
  deco = std_deco("LBT05"),
  ...
)

lbt05_1_lyt(arm_var, lbl_overall, lbl_param, lbl_anrind, map, deco)

lbt05_1_pre(adam_db, arm_var = "ACTARM", ...)

lbt05_1_check(adam_db, req_tables, arm_var)

lbt05_1_post(tlg, prune_0 = FALSE, ...)

lbt05_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}

\item{lbl_anrind}{(\code{character}) label of the \code{ANRIND} variable.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{map}{(\code{data.frame}) mapping of \code{PARAM}s to directions of abnormality.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
\code{LBT05} Table 1 (Default) Laboratory Abnormalities with Single and Replicated Marked.
}
\details{
\itemize{
\item Does not remove rows with zero counts by default.
\item Lab test results with missing \code{AVAL} values are excluded.
\item Split columns by arm, typically \code{ACTARM}.
}
}
\section{Functions}{
\itemize{
\item \code{lbt05_1_main()}: Main TLG function

\item \code{lbt05_1_lyt()}: Layout

\item \code{lbt05_1_pre()}: Preprocessing

\item \code{lbt05_1_check()}: Checks

\item \code{lbt05_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns \code{"ONTRTFL"}, \code{"PARCAT2"}, \code{"PARAM"}, \code{"ANRIND"},
\code{"AVALCAT1"}, and column specified by \code{arm_var}.
}
}
\examples{
run(lbt05_1, syn_data)
}
\keyword{datasets}
