% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt05.R
\name{lbt05_lyt}
\alias{lbt05_lyt}
\title{\code{lbt05} Layout}
\usage{
lbt05_lyt(arm_var, lbl_overall, lbl_param, lbl_anrind, map)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{lbl_anrind}{(\code{string}) label of the \code{ANRIND} variable.}

\item{map}{(\code{data.frame}) mapping of \code{PARAM}s to directions of abnormality.}
}
\description{
\code{lbt05} Layout
}
\keyword{internal}
