% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt06.R
\name{lbt06_lyt}
\alias{lbt06_lyt}
\title{\code{lbt06} Layout}
\usage{
lbt06_lyt(
  arm_var,
  visitvar,
  anrind_var,
  bnrind_var,
  lbl_param,
  lbl_visit,
  lbl_anrind,
  lbl_bnrind,
  page_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{visitvar}{(\code{string}) typically \code{"AVISIT"} or user-defined visit incorporating \code{"ATPT"}.}

\item{anrind_var}{(\code{string}) the variable for analysis reference range indicator.}

\item{bnrind_var}{(\code{string}) the variable for baseline reference range indicator.}

\item{lbl_param}{(\code{string}) text label of the \code{PARAM} variable.}

\item{lbl_visit}{(\code{string}) text label of the \code{AVISIT} variable.}

\item{lbl_anrind}{(\code{string}) text label of the \code{ANRIND} variable.}

\item{lbl_bnrind}{(\code{string}) text label of the \code{BNRIND} variable.}

\item{page_var}{(\code{string}) variable name prior to which the row split is by page.}

\item{param}{(\code{string}) the variable for parameter code.}
}
\description{
\code{lbt06} Layout
}
\keyword{internal}
