% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt07.R
\docType{data}
\name{lbt07_1_main}
\alias{lbt07_1_main}
\alias{lbt07_1_lyt}
\alias{lbt07_1_pre}
\alias{lbt07_1_check}
\alias{lbt07_1_post}
\alias{lbt07_1}
\title{\code{LBT07} Table 1 (Default) Laboratory Test Results and Change from Baseline by Visit.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt07_1_main(
  adam_db,
  armvar = "ACTARM",
  lbl_overall = NULL,
  grade_var = c("PARAM", "GRADE_DIR", "GRADE_ANL"),
  deco = std_deco("LBT07"),
  lbl_grade_var = c("Parameter", "Direction of Abnormality", "Toxicity Grade"),
  ...
)

lbt07_1_lyt(
  armvar,
  lbl_overall,
  lbl_gradedir,
  lbl_param,
  grade_var,
  lbl_grade_var,
  deco,
  map,
  ...
)

lbt07_1_pre(adam_db, ...)

lbt07_1_check(adam_db, req_tables = c("adsl", "adlb"), armvar = "ACTARM", ...)

lbt07_1_post(tlg, prune_0 = TRUE, ...)

lbt07_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{grade_var}{(\code{character}) \code{PARAM} and variables derived from the standard lab grade variable \code{ATOXGR}:
\itemize{
\item A grade direction variable (\code{GRADE_DIR}) is required in order to obtain
the correct denominators when building the layout as it is used to define row splitting.
\item A toxicity grade variable (e.g. \code{GRADE_ANL}) where all negative values from
\code{ATOXGR} are replaced by their absolute values.
}}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{lbl_grade_var}{(\code{character}) label of the variables in \code{grade_var}. If \code{NULL}, uses the label
attribute of the columns selected in \code{grade_var}.}

\item{...}{not used.}

\item{lbl_gradedir}{(\code{character}) label of the \code{GRADE_DIR} variable.}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}

\item{map}{(\code{data.frame}) mapping of \code{PARAM}s to directions of abnormality.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
The \code{LBT07} table provides an
overview of the analysis values and its change from baseline of each respective arm over the course of the trial.
}
\details{
\itemize{
\item Split columns by arm, typically \code{ACTARM}.
}
}
\section{Functions}{
\itemize{
\item \code{lbt07_1_main()}: Main TLG function

\item \code{lbt07_1_lyt()}: Layout

\item \code{lbt07_1_pre()}: Preprocessing

\item \code{lbt07_1_check()}: Checks

\item \code{lbt07_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns \code{"USUBJID"}, \code{"ATOXGR"},
\code{"ONTRTFL"} and column specified by \code{armvar}.
}
}
\examples{
run(lbt07_1, syn_data)
}
\keyword{datasets}
