% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt07.R
\name{lbt07_lyt}
\alias{lbt07_lyt}
\title{\code{lbt07} Layout}
\usage{
lbt07_lyt(
  arm_var,
  param_var,
  grad_dir_var,
  grad_anl_var,
  lbl_param_var,
  lbl_grad_dir_var,
  lbl_overall,
  map
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{param_var}{(\code{string}) the name of the column storing the parameters name.}

\item{grad_dir_var}{(\code{string}) the name of the column storing the grade direction variable which  is required in
order to obtain the correct denominators when building the layout as it is used to define row splitting.}

\item{grad_anl_var}{(\code{string})  the name of the column storing toxicity grade variable where all negative values from
\code{ATOXGR} are replaced by their absolute values.}

\item{lbl_param_var}{(\code{string}) label of the \code{param_var} variable.}

\item{lbl_grad_dir_var}{(\code{string}) label for the \code{grad_dir_var} variable.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{map}{(\code{data.frame}) mapping of \code{PARAM}s to directions of abnormality.}
}
\description{
\code{lbt07} Layout
}
\keyword{internal}
