% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt08.R
\docType{data}
\name{lbt08_1_main}
\alias{lbt08_1_main}
\alias{lbt08_1_lyt}
\alias{lbt08_1_pre}
\alias{lbt08_1_post}
\alias{lbt08_1}
\title{\code{LBT08} Table 1 Laboratory Events (Worsened from Baseline) by Highest NCI CTCAE Grade Post-Baseline}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt08_1_main(
  adam_db,
  armvar = "ARMCD",
  lbl_param = "Laboratory Test",
  lbl_grade = "Direction of Abnormality",
  deco = std_deco("LBT08"),
  ...
)

lbt08_1_lyt(armvar, lbl_param, lbl_grade, deco, ...)

lbt08_1_pre(adam_db, ...)

lbt08_1_post(tlg, ...)

lbt08_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable describing arm splits.}

\item{lbl_param}{(\code{character}) label of the \code{PARAMCD} variable.}

\item{lbl_grade}{(\code{character}) label of the \code{GRADDR} variable.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{LBT08} Table 1 Laboratory Events (Worsened from Baseline) by Highest NCI CTCAE Grade Post-Baseline
}
\details{
\itemize{
\item Split columns by arm, typically \code{ARMCD}.
\item Baseline grade variable described by \code{BTOXGR}.
\item Patient counts of Laboratory Events (Worsen by Baseline) by Highest Grade
Post-Baseline flags described by \code{WGRLOFL} ("LOW") and \code{WGRHIFL}("HIGH").
}
}
\section{Functions}{
\itemize{
\item \code{lbt08_1_main()}: Main TLG function

\item \code{lbt08_1_lyt()}: Layout

\item \code{lbt08_1_pre()}: Preprocessing

\item \code{lbt08_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns \code{"ONTRTFL"},
\code{"PARAMCD"}, \code{"ATOXGR"}, \code{WGRLOFL}, \code{WGRHIFL}, and column specified by \code{armvar}.
}
}
\examples{
run(lbt08_1, syn_data)
}
\keyword{datasets}
