% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt14.R
\docType{data}
\name{lbt14_1_main}
\alias{lbt14_1_main}
\alias{lbt14_1_lyt}
\alias{lbt14_1_pre}
\alias{lbt14_1_check}
\alias{lbt14_1_post}
\alias{lbt14_1}
\title{\code{LBT14} Table 1 (Low Direction) Laboratory Test Results Shift Table – Highest NCI-CTCAE Grade Post-Baseline by
Baseline Grade.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt14_1_main(
  adam_db,
  armvar = "ACTARM",
  gr_missing = "incl",
  title = "LBT14 - Low Direction",
  deco = std_deco("LBT14"),
  ...
)

lbt14_1_lyt(armvar, gr_missing, title, deco, ...)

lbt14_1_pre(adam_db, gr_missing = "incl", ...)

lbt14_1_check(adam_db, req_tables = c("adsl", "adlb"), armvar = "ACTARM", ...)

lbt14_1_post(tlg, prune_0 = TRUE, ...)

lbt14_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{gr_missing}{how missing baseline grades should be handled. Defaults to \code{"incl"} to include the \code{"Missing"}
level. Other options are \code{"excl"} to exclude patients with missing baseline grades and \code{"gr_0"} to convert missing
baseline grades to grade 0.}

\item{title}{table title to display. Defaults to \code{"LBT14 - Low Direction"}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
\code{LBT14} Table 1 (Low Direction) Laboratory Test Results Shift Table – Highest NCI-CTCAE Grade Post-Baseline by
Baseline Grade.
}
\details{
\itemize{
\item Only the worst grade recorded for each patient is included in the table.
\item If no missing baseline lab results, the "Missing" level of \code{BTOXGR} is excluded.
\item Grades 0, 1, 2, 3, and 4 are counted as \code{"Not Low"}.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
}
}
\section{Functions}{
\itemize{
\item \code{lbt14_1_main()}: Main TLG function

\item \code{lbt14_1_lyt()}: Layout

\item \code{lbt14_1_pre()}: Preprocessing

\item \code{lbt14_1_check()}: Checks

\item \code{lbt14_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns \code{"USUBJID"}, \code{"PARAM"}, \code{"BTOXGR"}, \code{"ATOXGR"},
and the column specified by \code{armvar}.
}
}
\examples{
run(lbt14_1, syn_data)
}
\keyword{datasets}
