% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt14.R
\name{lbt14_lyt}
\alias{lbt14_lyt}
\title{\code{lbt14} Layout}
\usage{
lbt14_lyt(arm_var, lbl_overall)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
\code{lbt14} Layout
}
\keyword{internal}
