% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listing_format_chevron}
\alias{listing_format_chevron}
\title{Format for Chevron Listings}
\usage{
listing_format_chevron()
}
\value{
a \code{list} of \code{fmt_config}.
}
\description{
Format for Chevron Listings
}
