% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lvls}
\alias{lvls}
\title{Obtain levels from vector}
\usage{
lvls(x)
}
\arguments{
\item{x}{(\code{character}) or (\code{factor}) object to obtain levels.}
}
\description{
Obtain levels from vector
}
\details{
For factors, the levels will be returned. For characters, the sorted unique values will be returned.
}
