% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_t-S4methods.R
\name{lyt}
\alias{lyt}
\alias{lyt,chevron_t-method}
\alias{lyt<-}
\alias{lyt<-,chevron_t-method}
\title{Layout}
\usage{
lyt(x)

\S4method{lyt}{chevron_t}(x)

lyt(x) <- value

\S4method{lyt}{chevron_t}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_t}) input.}

\item{value}{(\code{function}, \verb{list of functions}, \code{PreDataTableLayouts} or \verb{list of PreDataTableLayouts}) typically one
of the \verb{_lyt} function of \code{chevron}. If a \code{function} or a \verb{list of functions} is provided, they must have the \code{...}
formal argument.}
}
\description{
retrieve or set \code{lyt} function.
}
