% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{main}
\alias{main}
\alias{main<-}
\alias{main<-,chevron_tlg-method}
\title{Retrieve Main}
\usage{
main(x, ...)

main(x, ...) <- value

\S4method{main}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Main

Set Main Function
}
