% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\name{make_lyt_fun}
\alias{make_lyt_fun}
\title{Layout function constructor}
\usage{
make_lyt_fun(lyt)
}
\arguments{
\item{lyt}{(\code{function}, \verb{list of functions}, \code{call} or \code{PreDataTableLayouts}) typically one of the \verb{_lyt} function
of \code{chevron}.}
}
\description{
Layout function constructor
}
\examples{
\dontrun{
my_call <- quote(rtables::basic_table() \%>\% rtables::split_cols_by("ARM"))
lyt1 <- make_lyt_fun(my_call)

my_obj <- rtables::basic_table() \%>\% rtables::split_cols_by("ARM")
lyt2 <- make_lyt_fun(my_obj)

my_fun <- function(arm = "ARM", ...) rtables::basic_table() \%>\% rtables::split_cols_by(arm)
lyt3 <- make_lyt_fun(my_fun)
}
}
\keyword{internal}
