% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\name{make_lyt_ls}
\alias{make_lyt_ls}
\title{Layout List Constructor}
\usage{
make_lyt_ls(lyt)
}
\arguments{
\item{lyt}{(a single \code{function} returning \code{PreDataTableLayouts} object or \code{PreDataTableLayouts} object or \code{list} of
either \code{functions} or \code{PreDataTableLayouts} type of elements) typically one of the \verb{_lyt} function of \code{chevron}.
Functions passed to \code{lyt}, whether as a single \code{function} or as a \verb{list of functions}, must have the \code{...} formal
argument.}
}
\description{
Layout List Constructor
}
\examples{
\dontrun{
my_obj <- rtables::basic_table() \%>\% rtables::split_cols_by("ARM")
lyt2 <- make_lyt_ls(my_obj)

my_fun <- function(arm = "ARM", ...) rtables::basic_table() \%>\% rtables::split_cols_by(arm)
lyt3 <- make_lyt_ls(my_fun)
}
}
\keyword{internal}
