% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mht01.R
\name{mht01_1_main}
\alias{mht01_1_main}
\alias{mht01_1_lyt}
\alias{mht01_1_pre}
\title{\code{MHT01} Table 1 (Default) Medical History Table 1.}
\usage{
mht01_1_main(
  adam_db,
  armvar = .study$planarm,
  lbl_overall = .study$lbl_overall,
  lbl_mhbodsys = var_labels_for(adam_db$admh, "MHBODSYS"),
  lbl_mhdecod = var_labels_for(adam_db$admh, "MHDECOD"),
  prune_0 = TRUE,
  deco = std_deco("MHT01"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

mht01_1_lyt(
  armvar = .study$planarm,
  lbl_overall = .study$lbl_overall,
  lbl_mhbodsys = "MedDRA System Organ Class",
  lbl_mhdecod = "MedDRA preferred Term",
  deco = std_deco("MHT01"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

mht01_1_pre(adam_db, ...)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_mhbodsys}{(\code{character}) text label for \code{MHBODSYS}.}

\item{lbl_mhdecod}{(\code{character}) text label for \code{MHDECOD}.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}
}
\description{
The \code{MHT01} table provides an overview of the subjects medical
history by SOC and Preferred Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Order by body system alphabetically and within body system and medical condition by decreasing total number of
patients with the specific condition.
}
}
\section{Functions}{
\itemize{
\item \code{mht01_1_lyt}: \code{mht01_1} Layout

\item \code{mht01_1_pre}: \code{mht01_1} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  mht01_1_pre()

mht01_1_main(adam_db = db) \%>\% head(15)
mht01_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("MHT01")
)
syn_test_data() \%>\%
  mht01_1_pre()
}
