% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mht01.R
\name{mht01_lyt}
\alias{mht01_lyt}
\title{\code{mht01} Layout}
\usage{
mht01_lyt(arm_var, lbl_overall, lbl_mhbodsys, lbl_mhdecod)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_mhbodsys}{(\code{string}) label associated with \code{"MHBODSYS"}.}

\item{lbl_mhdecod}{(\code{string}) label associated with \code{"MHDECOD"}.}
}
\description{
\code{mht01} Layout
}
\keyword{internal}
