% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\docType{data}
\name{missing_rule}
\alias{missing_rule}
\title{Missing rule}
\format{
An object of class \code{rule} (inherits from \code{character}) of length 2.
}
\usage{
missing_rule
}
\description{
Missing rule
}
\keyword{datasets}
