% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mla_dir}
\alias{mla_dir}
\title{MLA Grade Direction Data}
\format{
An object of class \code{data.frame} with 76 rows and 2 columns.
}
\usage{
mla_dir
}
\description{
MLA Grade Direction Data
}
\keyword{datasets}
