% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mla_grid}
\alias{mla_grid}
\title{mla grade data}
\format{
An object of class \code{data.frame} with 76 rows and 2 columns.
}
\usage{
mla_grid
}
\description{
mla grade data
}
\keyword{datasets}
