% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng01.R
\name{mng01_1}
\alias{mng01_1}
\title{\code{MNG01} Graph 1 (Default) Mean Plot 1.}
\usage{
mng01_1(
  adam_db,
  dataset = "adlb",
  xval = "AVISIT",
  yval = "AVAL",
  armvar = .study$actualarm,
  center_fun = c("mean", "median"),
  interval_fun = c("mean_ci", "mean_sei", "mean_sdi", "median_ci", "quantiles",
    "range"),
  jitter = TRUE,
  show_n = FALSE,
  show_h_grid = .study$show_h_grid,
  show_v_grid = .study$show_v_grid,
  legend_pos = .study$legend_pos,
  .study = list(actualarm = "ARM", show_h_grid = TRUE, show_v_grid = FALSE, legend_pos
    = "top")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{dataset}{(\code{character}) the name of the data set to be analyzed.}

\item{xval}{(\code{character}) the name of the variable to be represented on the x-axis.}

\item{yval}{(\code{character}) the name of the variable to be represented on the y-axis.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{center_fun}{(\code{character}) the function to compute the estimate value. One of \code{mean} or \code{median}.}

\item{interval_fun}{(\code{character}) the function defining the crossbar range. One of \code{mean_ci}, \code{mean_sei}, \code{mean_sdi},
\code{median_ci}, \code{quantiles} or \code{range}.}

\item{jitter}{(\code{logical}) should data point be slightly spread on the x-axis.}

\item{show_n}{(\code{logical}) should the number of observation be displayed int the table.}

\item{show_h_grid}{(\code{logical}) should horizontal grid be displayed.}

\item{show_v_grid}{(\code{logical}) should vertical grid be displayed.}

\item{legend_pos}{(\code{character}) the position of the legend. One of \code{top}, \code{bottom}, \code{right}, \code{left} or \code{none}.}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{

}
\description{
Overview of a summary statistics across time and arm for a selected parameter and data set.
}
\details{
\itemize{
\item No overall value.
}
}
\examples{
library(dm)
library(dplyr)

db <- chevron::syn_test_data() \%>\%
  dm_zoom_to("adlb") \%>\%
  filter(PARAM == "Immunoglobulin A Measurement") \%>\%
  dm_update_zoomed()

db \%>\% mng01_1(
  center_fun = "mean",
  interval_fun = "mean_sei",
  legend_pos = "bottom",
  show_n = TRUE,
  jitter = FALSE
)
}
