% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{mutate_adsl_gp}
\alias{mutate_adsl_gp}
\title{Categorize Reason for Discontinuation from Study.}
\usage{
mutate_adsl_gp(
  adam_db,
  reason = .study$disc_reason_var,
  .study = list(disc_reason_var = "DCSREAS")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{reason}{(\code{character}) the variable name for variable with the reason for discontinuation.}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
Categorize Reason for Discontinuation from Study.
}
