% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{mutate_cmt01a}
\alias{mutate_cmt01a}
\title{Coerce \code{CMSEQ} to factor in \code{adcm}}
\usage{
mutate_cmt01a(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
Coerce \code{CMSEQ} to factor in \code{adcm}
}
