% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{mutate_egt02}
\alias{mutate_egt02}
\title{Categorize \code{adeg} values}
\usage{
mutate_egt02(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
Categorize \code{adeg} values
}
