% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{mutate_vst02}
\alias{mutate_vst02}
\title{Categorize \code{advs} values}
\usage{
mutate_vst02(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
Categorize \code{advs} values
}
