% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_list}
\alias{name_list}
\title{Create list with names}
\usage{
name_list(obj, names)
}
\arguments{
\item{obj}{Object content.}

\item{names}{(\code{character}) names of the object.}
}
\description{
Create list with names
}
