% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\docType{data}
\name{nocoding}
\alias{nocoding}
\title{No Coding Available rule}
\format{
An object of class \code{rule} (inherits from \code{character}) of length 2.
}
\usage{
nocoding
}
\description{
No Coding Available rule
}
\keyword{datasets}
