% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\docType{data}
\name{outcome_rule}
\alias{outcome_rule}
\title{Outcome Rule}
\format{
An object of class \code{rule} (inherits from \code{character}) of length 6.
}
\usage{
outcome_rule
}
\description{
Outcome Rule
}
\keyword{datasets}
