% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdt01.R
\docType{data}
\name{pdt01_1_main}
\alias{pdt01_1_main}
\alias{pdt01_1_lyt}
\alias{pdt01_1_pre}
\alias{pdt01_1_post}
\alias{pdt01_1}
\title{\code{pdt01_1} Table 1 (Default) Major Protocol Deviations.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
pdt01_1_main(
  adam_db,
  armvar = "ARM",
  dvcode_var = "DVDECOD",
  lbl_dvcode_var = "Protocol Deviation Coded Term",
  dvterm_var = "DVTERM",
  lbl_dvterm_var = "Category",
  lbl_overall = NULL,
  deco = std_deco("pdt01_1"),
  ...
)

pdt01_1_lyt(
  armvar,
  lbl_overall,
  dvcode_var,
  lbl_dvcode_var,
  dvterm_var,
  lbl_dvterm_var,
  deco,
  ...
)

pdt01_1_pre(adam_db, dvcode_var = "DVDECOD", dvterm_var = "DVTERM", ...)

pdt01_1_post(
  tlg,
  prune_0 = TRUE,
  dvcode_var = "DVDECOD",
  dvterm_var = "DVTERM",
  ...
)

pdt01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{dvcode_var}{(\code{character}) the variable defining the protocol deviation coded term. By default \code{DVDECOD}.}

\item{lbl_dvcode_var}{(\code{character}) label for the variable defining the protocol deviation coded term.}

\item{dvterm_var}{(\code{character}) the variable defining the protocol deviation term. By default \code{DVTERM}.}

\item{lbl_dvterm_var}{(\code{character}) label for the variable defining the protocol deviation term.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
A major protocol deviations
table with the number of subjects and the total number of treatments by medication class sorted alphabetically and
medication name sorted by frequencies.
}
\details{
\itemize{
\item Data should be filtered for major protocol deviations. \code{(DVCAT == "MAJOR")}.
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by medication class alphabetically and within medication class by decreasing total number of patients with
the specific medication.
}
}
\section{Functions}{
\itemize{
\item \code{pdt01_1_main()}: Main TLG function

\item \code{pdt01_1_lyt()}: Layout

\item \code{pdt01_1_pre()}: Preprocessing

\item \code{pdt01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{addv} table with the columns specified in \code{dvcode_var} and \code{dvterm_var} as well
as \code{"DVSEQ"}.
}
}
\examples{
library(magrittr)
library(dm)

db <- syn_data \%>\%
  dm_zoom_to("addv") \%>\%
  filter(.data$DVCAT == "MAJOR") \%>\%
  dm_update_zoomed()

run(pdt01_1, db)
}
\keyword{datasets}
