% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdt01.R
\name{pdt01_lyt}
\alias{pdt01_lyt}
\title{\code{pdt01} Layout}
\usage{
pdt01_lyt(
  arm_var,
  lbl_overall,
  dvcode_var,
  lbl_dvcode_var,
  dvterm_var,
  lbl_dvterm_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{dvcode_var}{(\code{string}) the variable defining the protocol deviation coded term. By default \code{DVDECOD}.}

\item{lbl_dvcode_var}{(\code{string}) label for the variable defining the protocol deviation coded term.}

\item{dvterm_var}{(\code{string}) the variable defining the protocol deviation term. By default \code{DVTERM}.}

\item{lbl_dvterm_var}{(\code{string}) label for the variable defining the protocol deviation term.}
}
\description{
\code{pdt01} Layout
}
\keyword{internal}
