% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdt02.R
\docType{data}
\name{pdt02_1_main}
\alias{pdt02_1_main}
\alias{pdt02_1_lyt}
\alias{pdt02_1_pre}
\alias{pdt02_1_post}
\alias{pdt02_1}
\title{\code{pdt02_1} Table 1 (Default) Major Protocol Deviations Related to Epidemic/Pandemic.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
pdt02_1_main(
  adam_db,
  armvar = "ARM",
  dvreas_var = "DVREAS",
  lbl_dvreas_var = "Primary Reason",
  dvterm_var = "DVTERM",
  lbl_dvterm_var = "Description",
  lbl_overall = NULL,
  deco = std_deco("pdt02_1"),
  ...
)

pdt02_1_lyt(
  armvar,
  lbl_overall,
  dvreas_var,
  lbl_dvreas_var,
  dvterm_var,
  lbl_dvterm_var,
  deco,
  ...
)

pdt02_1_pre(adam_db, dvreas_var = "DVREAS", dvterm_var = "DVTERM", ...)

pdt02_1_post(
  tlg,
  prune_0 = TRUE,
  dvreas_var = "DVREAS",
  dvterm_var = "DVTERM",
  ...
)

pdt02_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{dvreas_var}{(\code{character}) the variable defining the reason for deviation. By default \code{DVREAS}.}

\item{lbl_dvreas_var}{(\code{character}) label for the variable defining the reason for deviation.}

\item{dvterm_var}{(\code{character}) the variable defining the protocol deviation term. By default \code{DVTERM}.}

\item{lbl_dvterm_var}{(\code{character}) label for the variable defining the protocol deviation term.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
A major protocol deviations
table with the number of subjects and the total number of Major Protocol Deviations Related
to Epidemic/Pandemic sorted alphabetically and deviations name sorted by frequencies.
}
\details{
\itemize{
\item Data should be filtered for major protocol deviations related to epidemic/pandemic.
\code{(AEPRELFL == "Y" & DVCAT == "MAJOR")}.
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by deviation reason alphabetically and within deviation reason by decreasing total number of patients with
the specific deviation term.
}
}
\section{Functions}{
\itemize{
\item \code{pdt02_1_main()}: Main TLG function

\item \code{pdt02_1_lyt()}: Layout

\item \code{pdt02_1_pre()}: Preprocessing

\item \code{pdt02_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{addv} table with the columns specified in \code{dvreas_var} and \code{dvterm_var} as well
as \code{"DVSEQ"}.
}
}
\examples{
run(pdt02_1, syn_data)
}
\keyword{datasets}
