% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdt02.R
\name{pdt02_lyt}
\alias{pdt02_lyt}
\title{\code{pdt02} Layout}
\usage{
pdt02_lyt(
  arm_var,
  lbl_overall,
  lbl_dvreas_var,
  lbl_dvterm_var,
  dvreas_var,
  dvterm_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_dvreas_var}{(\code{string}) label for the variable defining the reason for deviation.}

\item{lbl_dvterm_var}{(\code{string}) label for the variable defining the protocol deviation term.}

\item{dvreas_var}{(\code{string}) the variable defining the reason for deviation. By default \code{DVREAS}.}

\item{dvterm_var}{(\code{string}) the variable defining the protocol deviation term. By default \code{DVTERM}.}
}
\description{
\code{pdt02} Layout
}
\keyword{internal}
