% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_wider_labels}
\alias{pivot_wider_labels}
\title{Pivot wider a data frame while preserving labels.}
\usage{
pivot_wider_labels(
  df,
  names_from = "PARAMCD",
  labels_from = "PARAM",
  values_from = "AVAL",
  keep = "USUBJID"
)
}
\arguments{
\item{df}{(\code{data.frame}) to pivot.}

\item{names_from}{(\code{string}) the name of a column in \code{df} that will be used to name the columns of the pivoted data
frame.}

\item{labels_from}{(\code{string}) the name of a column in \code{df} that will be used to label the columns of the pivoted
data frame.}

\item{values_from}{(\code{string}) the name of a column in \code{df}containing the values to be pivoted.}

\item{keep}{(\verb{vector of strings}) containing the name of the columns to be conserved in the pivoted data frame.
Typically a unique identifier for the pivoted observations, such as \code{USUBJID}.}
}
\value{
(\code{data.frame})
}
\description{
Pivot wider a data frame while preserving labels.
}
\examples{
library(scda)
adsub <- synthetic_cdisc_data("rcd_2021_03_22")$adsub
pivot_wider_labels(adsub, "PARAMCD", "PARAM", "AVAL", c("USUBJID", "SUBJID"))
}
