% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess,chevron_tlg-method}
\alias{preprocess<-}
\alias{preprocess<-,chevron_tlg-method}
\title{Retrieve Preprocess}
\usage{
preprocess(x, ...)

\S4method{preprocess}{chevron_tlg}(x)

preprocess(x, ...) <- value

\S4method{preprocess}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Preprocess

Set Preprocess Function
}
