% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{preprocess_data}
\alias{preprocess_data}
\title{pre-process \code{ADaM} data for a TLG function}
\usage{
preprocess_data(adam_db, tlgfname, pmap = std_pmap, .study)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{tlgfname}{(\code{character}) function name of TLG function}

\item{pmap}{(\code{data.frame}) preprocessing map, see \code{std_pmap} for the required format}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
pre-process \code{ADaM} data for a TLG function
}
\examples{
\dontrun{
library(magrittr)
db <- syn_test_data()

db \%>\%
  preprocess_data("aet02_2")
}
}
