% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prune_except}
\alias{prune_except}
\title{Prune table except specified levels}
\usage{
prune_except(keep)
}
\arguments{
\item{keep}{(\code{character}) levels to keep.}
}
\value{
A pruning \code{function}.
}
\description{
Prune table except specified levels
}
\keyword{internal}
