% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_study_object.R
\name{read_study_object}
\alias{read_study_object}
\title{Reading Study Object from \code{yaml} file}
\usage{
read_study_object(file)
}
\arguments{
\item{file}{(\code{string}) path to a \code{yaml} file containing the \code{.study} object specification.}
}
\value{
list fitting the criteria of a \code{.study} object
}
\description{
Reading Study Object from \code{yaml} file
}
