% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_manipulation.R
\name{remove_adex_aval}
\alias{remove_adex_aval}
\title{Remove specific \code{AVAL} data from \code{adex}}
\usage{
remove_adex_aval(
  adam_db,
  show_stats = .study$show_cont_stats,
  .study = list(show_cont_stats = c("ALL"))
)
}
\arguments{
\item{show_stats}{(\verb{vector of strings}) providing the name of the parameters whose statistical summary should be
presented. To analyze all, provide \code{show_stats = "ALL"} (Default), to analyze none, provide \code{paramvar = ""}.}
}
\description{
Useful if some continuous \code{AVAL} data corresponding to specific \code{PARAM} should not be displayed.
}
