% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{reorder_adex_params}
\alias{reorder_adex_params}
\title{Reorder \code{PARAM} and \code{PARAMCD} levels}
\usage{
reorder_adex_params(
  adam_db,
  paramcd_order = .study$paramcd_order,
  .study = list(paramcd_order = c("TNDOSE", "DOSE", "NDOSE", "TDOSE"))
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{paramcd_order}{(\verb{vector of character}) providing the \code{PARAMCD} values in the desired order.}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{
a \code{dm} object.
}
\description{
Reorder \code{PARAM} and \code{PARAMCD} levels
}
