% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{reorder_dtht01}
\alias{reorder_dtht01}
\title{Reorder levels of \code{DTHCAT} in \code{adsl}}
\usage{
reorder_dtht01(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}
}
\description{
The \code{OTHER} level is pushed last so that it appears last, and then the \code{DTHCAUS} related to \code{OTHER} appear underneath
as intended.
}
