% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.gg_list.ggplot}
\alias{as.gg_list.ggplot}
\alias{rl_list}
\alias{as.rl_list}
\alias{as.rl_list.list}
\alias{as.rl_list.listing_df}
\title{List of \code{rlistings} object}
\usage{
\method{as.gg_list}{ggplot}(obj)

rl_list(...)

as.rl_list(obj)

\method{as.rl_list}{list}(obj)

\method{as.rl_list}{listing_df}(obj)
}
\arguments{
\item{obj}{(\code{rlisting} or \code{list} of \code{rlistings})}

\item{...}{(\code{rlistings}) objects.}
}
\value{
a \code{rl_list} object.

a \code{rl_list} object.
}
\description{
List of \code{rlistings} object

Convert Object to List of \code{rl_list}.
}
