% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt01.R
\docType{data}
\name{rmpt01_1_main}
\alias{rmpt01_1_main}
\alias{split_if_not_null}
\alias{rmpt01_1_lyt}
\alias{rmpt01_1_pre}
\alias{rmpt01_1_check}
\alias{rmpt01_1_post}
\alias{rmpt01_1}
\title{\code{RMPT01} Table 1 (Default) Duration of Exposure for Risk Management Plan Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt01_1_main(
  adam_db,
  anl_vars = c("AVALCAT1", "AVAL"),
  lbl_vars = c("Patients", "Person time*"),
  parcat = NULL,
  lbl_parcat = "Actual Treatment",
  deco = std_deco("RMPT01"),
  ...
)

split_if_not_null(lyt, parcat, lbl_parcat)

rmpt01_1_lyt(anl_vars, lbl_vars, parcat, lbl_parcat, deco)

rmpt01_1_pre(adam_db, anl_vars = c("AVALCAT1", "AVAL"), ...)

rmpt01_1_check(adam_db, anl_vars, req_tables = c("adsl", "adex"))

rmpt01_1_post(tlg, prune_0 = FALSE, ...)

rmpt01_1
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{anl_vars}{(\code{character}) the names of variables to be analyzed.}

\item{lbl_vars}{(\code{character}) the labels of the patient toll and time variables.}

\item{parcat}{(\code{string}) the name of the variable initiating a new row split.}

\item{lbl_parcat}{(\code{string}) the label of the variable \code{parcat}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{RMPT01} table provides an overview of duration of exposure for SE patients
}
\details{
\itemize{
\item Person time is the sum of exposure across all patients.
\item Summary statistics are by default based on the number of patients in the corresponding \code{N} row
(number of non-missing values).
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt01_1_main()}: Main TLG function

\item \code{split_if_not_null()}: Helper function to add a row split if specified

\item \code{rmpt01_1_lyt()}: Layout

\item \code{rmpt01_1_pre()}: Preprocessing

\item \code{rmpt01_1_check()}: Checks

\item \code{rmpt01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adex} table with \code{"PARAMCD"} and the columns specified by \code{anl_vars} which
is denoted as \code{c("AVALCAT1", "AVAL")} by default.
}
}
\examples{
run(rmpt01_1, syn_data)

run(rmpt01_1, syn_data, parcat = "PARCAT2")
}
\keyword{datasets}
\keyword{internal}
