% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt03.R
\name{rmpt03_lyt}
\alias{rmpt03_lyt}
\title{\code{rmpt03} Layout}
\usage{
rmpt03_lyt(summaryvars, lbl_summaryvars, show_tot, split_var)
}
\arguments{
\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adsl}
table of \code{adam_db} is used as label.}

\item{lbl_summaryvars}{(\code{character}) label associated with the analyzed variables.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{character}) recursive split row variable.}
}
\description{
\code{rmpt03} Layout
}
\keyword{internal}
