% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt04.R
\name{rmpt04_lyt}
\alias{rmpt04_lyt}
\title{\code{rmpt04} Layout}
\usage{
rmpt04_lyt(summaryvars, lbl_summaryvars, show_tot, split_var)
}
\arguments{
\item{summaryvars}{(\code{character}) variables to be analyzed. The label attribute of the corresponding columns in \code{adsl}
table of \code{adam_db} is used as label.}

\item{lbl_summaryvars}{(\code{string}) label associated with the analyzed variable.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{character}) recursive split row variable.}
}
\description{
\code{rmpt04} Layout
}
\keyword{internal}
