% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt06.R
\name{rmpt06_lyt}
\alias{rmpt06_lyt}
\title{\code{rmpt06} Layout}
\usage{
rmpt06_lyt(
  arm_var,
  lbl_overall,
  method,
  conf_level,
  show_diff,
  ref_group,
  method_diff,
  conf_level_diff,
  grade_groups
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
\code{rmpt06} Layout
}
\keyword{internal}
