% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspt01.R
\docType{data}
\name{rspt01_1_main}
\alias{rspt01_1_main}
\alias{rspt01_1_lyt}
\alias{rspt01_1_pre}
\alias{rspt01_1_post}
\alias{rspt01_1}
\title{\code{RSPT01} Table 1 (Default) Demographics and Baseline Characteristics Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rspt01_1_main(
  adam_db,
  dataset = "adrs",
  arm_var = "ARM",
  ref_group = NULL,
  odds_ratio = TRUE,
  strat_analysis = FALSE,
  conf_level = 0.95,
  methods = list(),
  deco = std_deco("RSPT01"),
  ...
)

rspt01_1_lyt(
  arm_var,
  ref_group,
  odds_ratio,
  strat_analysis,
  strata,
  conf_level,
  methods,
  deco
)

rspt01_1_pre(
  adam_db,
  arm_var = "ARM",
  dataset = "adrs",
  responder = c("CR", "PR"),
  ...
)

rspt01_1_post(tlg, prune_0 = TRUE, ...)

rspt01_1
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}

\item{summaryvars_lbls}{(\code{character}) labels corresponding to the analyzed variables.}
}
\description{
For each variable, summary statistics are
by default based on the number of patients in the corresponding \code{n} row.
}
\section{Functions}{
\itemize{
\item \code{rspt01_1_main()}: Main TLG function

\item \code{rspt01_1_lyt()}: Layout

\item \code{rspt01_1_pre()}: Preprocessing

\item \code{rspt01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item 
}
}
\examples{
rspt01_1_pre(syn_data)
library(dplyr)
library(dunlin)

syn_data2 <- log_filter(syn_data, PARAMCD == "BESRSPI", "adrs")
run(rspt01_1, syn_data2)
}
\keyword{datasets}
