% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspt01.R
\name{rspt01_lyt}
\alias{rspt01_lyt}
\title{\code{rspt01} Layout}
\usage{
rspt01_lyt(
  arm_var,
  lbl_overall,
  rsp_var,
  ref_group,
  odds_ratio,
  perform_analysis,
  strata,
  conf_level,
  methods
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
\code{rspt01} Layout
}
\keyword{internal}
