% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtht01.R
\name{s_summary_na}
\alias{s_summary_na}
\title{Summary factor allowing NA}
\usage{
s_summary_na(
  x,
  labelstr,
  denom = c("n", "N_row", "N_col"),
  .N_row,
  .N_col,
  ...
)
}
\arguments{
\item{x}{(\code{factor}) input.}

\item{denom}{(\code{string}) denominator choice.}

\item{.N_row}{(\code{integer}) number of rows in row-splitted dataset.}

\item{.N_col}{(\code{integer}) number of rows in column-splitted dataset.}

\item{...}{Not used}
}
\description{
Summary factor allowing NA
}
\keyword{internal}
