% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_output.R
\name{sample_output}
\alias{sample_output}
\title{Sample function of Chevron}
\usage{
sample_output(sd)
}
\arguments{
\item{sd}{(\code{list}) cdisc_data}
}
\value{
(\code{list}) of tables
}
\description{
Returns a list of predefined table output.
}
\examples{
library(scda)

sd <- synthetic_cdisc_data("rcd_2021_03_22")
sample_output(sd)

}
