% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_study_object.R
\name{sample_study_object}
\alias{sample_study_object}
\title{Extraction of .study object default values}
\usage{
sample_study_object()
}
\value{
(\code{list})
}
\description{
Sample the default value of the \code{.study} object defined in the \code{tlg-functions} and returns an existing default value.
}
\examples{
lst <- sample_study_object()
dput(lst)
}
