% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{script}
\alias{script}
\alias{script_funs}
\alias{script_funs,chevron_tlg-method}
\alias{script_funs,chevron_simple-method}
\title{Create Script for \code{TLG} Generation}
\usage{
script_funs(x, adam_db, args, name = deparse(substitute(x)))

\S4method{script_funs}{chevron_tlg}(x, adam_db, args, name = deparse(substitute(x)))

\S4method{script_funs}{chevron_simple}(x, adam_db, args, name = deparse(substitute(x)))
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{adam_db}{(\code{string}) the name of the dataset.}

\item{args}{(\code{string}) the name of argument list.}

\item{name}{(\code{string}) name of the template.}
}
\value{
\code{character} that can be integrated into an executable script.
}
\description{
Create Script for \code{TLG} Generation
}
\examples{
script_funs(aet04, adam_db = "syn_data", args = "args")
}
