% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_decoration}
\alias{set_decoration}
\title{Temporary function to add decorator to a table}
\usage{
set_decoration(x, deco)
}
\arguments{
\item{x}{(\code{rtables})}

\item{deco}{(\code{list}) typically generated with \code{std_deco()}}
}
\value{

}
\description{
Temporary function to add decorator to a table
}
