% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{set_lyt}
\alias{set_lyt}
\alias{set_lyt,chevron_tlg-method}
\title{Set Layout Function}
\usage{
set_lyt(object, lyt, ...)

\S4method{set_lyt}{chevron_tlg}(object, lyt)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{lyt}{(\code{function}, \verb{list of functions}, \code{call} or \code{PreDataTableLayouts}) typically one of the \verb{_lyt} function
of \code{chevron}.}
}
\description{
Set Layout Function
}
