% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_section_div}
\alias{set_section_div}
\title{Set Section Dividers}
\usage{
set_section_div(x)
}
\arguments{
\item{x}{(\code{integerish}) value of at which the section divider should be added.}
}
\value{
invisible \code{NULL}. Set the \code{chevron.section_div} option.
}
\description{
Set Section Dividers
}
\details{
Section dividers are empty lines between sections in tables.
E.g. if 1 is used then for the first row split an empty line is added.
Currently it only works for \code{aet02}, \code{cmt01a} and \code{mht01} template.
}
