% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smart_prune}
\alias{smart_prune}
\title{Prune Table up to an \code{ElementaryTable}}
\usage{
smart_prune(tlg)
}
\arguments{
\item{tlg}{(\code{TableTree}) object.}
}
\value{
pruned \code{TableTree}.
}
\description{
Avoid returning \code{NULL} when the \code{table} is empty.
}
