% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_str_time}
\alias{sort_str_time}
\title{Format Date and Keep Chronological Order in Levels}
\usage{
sort_str_time(x, format = "\%d\%b\%Y", to_upper = TRUE)
}
\arguments{
\item{x}{(\code{character}, \code{factor}, \code{Date} or \code{POSIXct}) to convert.}

\item{format}{(\code{string}) see \code{strptime}.}

\item{to_upper}{(\code{flag}) whether to convert the final result to upper case.}
}
\description{
Format Date and Keep Chronological Order in Levels
}
\examples{
sort_str_time("2020-01-01")
sort_str_time(as.Date("2020-07-01"))
}
