% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_strp_time}
\alias{sort_strp_time}
\title{Format Date and Keep Chronomoligal Order in Levels}
\usage{
sort_strp_time(x, format = "\%d\%b\%Y", to_upper = TRUE)
}
\arguments{
\item{x}{(\code{character}, \code{factor} or \code{POSIXct}) to convert.}

\item{format}{(\code{string}) see \code{strptime}.}

\item{to_upper}{(\code{flag}) whether to convert the final result to upper case.}
}
\description{
Format Date and Keep Chronomoligal Order in Levels
}
\keyword{internal}
