% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{split_rows_by_recurive}
\alias{split_rows_by_recurive}
\title{Count or summarize by groups}
\usage{
split_rows_by_recurive(lyt, split_var, ...)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts}) \code{rtable} layout.}

\item{split_var}{(\code{character}) variables to split rows by.}

\item{...}{Further arguments for \code{split_rows_by}}
}
\description{
Count or summarize by groups
}
\keyword{internal}
