% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_manipulation.R
\name{std_filter}
\alias{std_filter}
\title{Retrieve Standard Subsetting for Templates}
\usage{
std_filter(id)
}
\description{
Retrieve Standard Subsetting for Templates
}
\examples{
std_filter("aet02_1")
}
