% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{std_filter_fun}
\alias{std_filter_fun}
\title{Retrieve Standard Subsetting for Templates}
\usage{
std_filter_fun(idt)
}
\arguments{
\item{idt}{(\code{character}) the id of the table.}
}
\description{
Retrieve Standard Subsetting for Templates
}
\examples{
std_filter_fun("aet02_1")
}
