% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_manipulation.R
\name{std_mutate}
\alias{std_mutate}
\title{Retrieve Standard Mutation for Templates}
\usage{
std_mutate(id)
}
\description{
Retrieve Standard Mutation for Templates
}
\examples{
std_mutate("aet02_1")
}
