% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{std_mutate_fun}
\alias{std_mutate_fun}
\title{Retrieve Standard Mutation for Templates}
\usage{
std_mutate_fun(tlgfname, pmap = std_pmap())
}
\arguments{
\item{idt}{(\code{character}) the id of the table.}
}
\description{
Retrieve Standard Mutation for Templates
}
\examples{
std_mutate_fun("aet02_1")
}
