% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\docType{data}
\name{std_pmap}
\alias{std_pmap}
\title{Standard Preprocessing Map}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 4 columns.
}
\usage{
std_pmap
}
\description{
The pre-processing map contains information how the \code{ADaM} data needs to be pre-processed for each function.
}
\examples{
std_pmap
}
\keyword{datasets}
