% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_null.R
\name{std_postprocess}
\alias{std_postprocess}
\title{Standard Post Processing}
\usage{
std_postprocess(tlg, ind = 2L, ...)
}
\arguments{
\item{tlg}{to convert to null report.}

\item{ind}{(\code{integer}) indentation for the outputs of class \code{VTableTree}.}

\item{...}{not used.}
}
\value{
a post processed \code{tlg} or a \code{NULL} report.
}
\description{
Standard Post Processing
}
