% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_null.R
\name{std_postprocessing}
\alias{std_postprocessing}
\title{Standard Post Processing}
\usage{
std_postprocessing(tlg, ...)
}
\arguments{
\item{tlg}{to post process.}

\item{...}{additional arguments passed to \link{report_null}.}
}
\value{
a processed \code{tlg} or a null report.
}
\description{
Standard Post Processing
}
\examples{
library(rtables)
std_postprocessing(build_table(basic_table() |> analyze("Species"), iris), ind = 10L)

}
