% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_manipulation.R
\name{std_preproc_map}
\alias{std_preproc_map}
\title{Standard Preprocessing Map}
\usage{
std_preproc_map()
}
\description{
The preprocessing map contains information how the ADaM data needs to be preprocessed for each function.
}
\examples{
std_preproc_map()
}
