% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_study_object}
\alias{std_study_object}
\title{Standard Study Object}
\format{
a \code{list} of 19 elements that can be used by \code{tlg_functions}.
}
\usage{
std_study_object
}
\description{
Standard Study Object
}
\keyword{datasets}
