% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtht01.R
\name{summarize_vars_allow_na}
\alias{summarize_vars_allow_na}
\title{Summarize variables allow na}
\usage{
summarize_vars_allow_na(
  lyt,
  vars,
  var_labels = vars,
  nested = TRUE,
  ...,
  show_labels = "default",
  table_names = vars,
  section_div = NA_character_,
  .stats = c("n", "count_fraction"),
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL,
  inclNAs = TRUE
)
}
\description{
Summarize variables allow na
}
\keyword{internal}
