% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{syn_test_data}
\alias{syn_test_data}
\alias{syn_data}
\title{Retrieve Synthetic Test Data Used For Examples}
\format{
An object of class \code{dm} of length 19.
}
\usage{
syn_test_data()

syn_data
}
\description{
Retrieve Synthetic Test Data Used For Examples

example data generated with syn_test_data()
}
\keyword{datasets}
