% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4class.R
\name{tlg_pipeline_S4}
\alias{tlg_pipeline_S4}
\title{\code{tlg_pipeline_S4} constructor}
\usage{
tlg_pipeline_S4(
  main = function(adam_db, ...) adam_db,
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_
)
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{_main} function of \code{chevron}.}

\item{preprocess}{(\code{function}) returning a pre processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post processed \code{tlg}.}

\item{adam_datasets}{(\code{character}) representing the name of the table from an \code{AdAM} data set required for \code{tlg}
creation.}
}
\description{
\code{tlg_pipeline_S4} constructor
}
\examples{
x <- tlg_pipeline_S4(aet01_1_main, aet01_1_pre, adam_datasets = c("adsl", "adae"))
}
