% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttet01.R
\name{ttet01_lyt}
\alias{ttet01_lyt}
\title{ttet01 Layout}
\usage{
ttet01_lyt(
  arm_var,
  ref_group,
  summarize_event,
  perform_analysis,
  strata,
  pval_method,
  conf_level,
  conf_type,
  quantiles,
  ties,
  timeunit,
  timepoint,
  method
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{timeunit}{(string) time unit get from AVALU, by default is "Months"}
}
\description{
ttet01 Layout
}
