% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{unwrap_layout}
\alias{unwrap_layout}
\title{Extracting Layout Function.}
\usage{
unwrap_layout(x, pattern = "_lyt$")
}
\arguments{
\item{x}{(\code{function}) containing a call to a layout function.}

\item{pattern}{(\code{string}) identifying layout functions}
}
\value{
invisible \code{NULL} and print the content of the layout functions found in the body of \code{x}.
}
\description{
Extracting Layout Function.
}
\examples{
unwrap_layout(aet01_main)

}
\keyword{internal}
