% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_labels_for}
\alias{var_labels_for}
\title{Retrieve labels for certain variables}
\usage{
var_labels_for(df, vars)
}
\arguments{
\item{df}{(\code{data.frame}) containing columns with label attribute.}

\item{vars}{(\code{character}) variable names in \code{df}.}
}
\description{
Retrieve labels for certain variables
}
\details{
The labels will be returned if the column has \code{label} attribute, otherwise the column name will be returned.
Any values between brackets {} will be replaced with \code{dunlin::render_safe}.
}
