% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst01.R
\name{vst01_1_main}
\alias{vst01_1_main}
\alias{vst01_1_lyt}
\alias{vst01_1_pre}
\title{\code{VST01} Table 1 (Default) Vital Sign Results and change from Baseline By Visit Table 1.}
\usage{
vst01_1_main(
  adam_db,
  armvar = .study$actualarm,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = var_labels_for(adam_db$advs, summaryvars),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("VST01"),
  .study = list(actualarm = "ACTARM", evo_vars = c("AVAL", "CHG"))
)

vst01_1_lyt(
  armvar = .study$actualarm,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = .study$evo_vars_lbls,
  visitvar = .study$visitvar,
  lbl_avisit = "",
  lbl_param = "",
  deco = std_deco("VST01"),
  .study = list(actualarm = "ACTARM", evo_vars = c("AVAL", "CHG"), evo_vars_lbls =
    c("Analysis \\nValue", "Change from \\nBaseline"), visitvar = "AVISIT")
)

vst01_1_pre(adam_db, ...)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting.}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_avisit}{(\code{character}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}

\item{...}{not used.}
}
\description{
The \code{VST01} table provides an
overview of the Vital Sign values and its change from baseline of each respective arm
over the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{vst01_1_lyt()}: \code{vst01_1} Layout

\item \code{vst01_1_pre()}: \code{vst01_1} Preprocessing

}}
\examples{
library(dm)

db <- syn_test_data() \%>\%
  vst01_1_pre()

vst01_1_main(db)
vst01_1_main(db, summaryvars_lbls = c("Value at Visit", "Change from Baseline"))
syn_test_data() \%>\%
  vst01_1_pre()
}
