% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst01.R
\name{vst01_lyt}
\alias{vst01_lyt}
\title{\code{vst01} Layout}
\usage{
vst01_lyt(
  arm_var,
  summaryvars,
  summaryvars_lbls,
  visitvar,
  lbl_avisit,
  lbl_param
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\code{character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\code{character}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{lbl_avisit}{(\code{string}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}
}
\description{
\code{vst01} Layout
}
\keyword{internal}
