% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst02.R
\name{vst02_2_main}
\alias{vst02_2_main}
\alias{vst02_2_lyt}
\alias{vst02_2_pre}
\title{\code{VST02} Table 2 (Supplementary) Vital Sign Abnormalities Table 2.}
\usage{
vst02_2_main(
  adam_db,
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("VST02_2"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

vst02_2_lyt(
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("VST02_2"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

vst02_2_pre(adam_db, ...)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{character}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{character}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}
}
\description{
Assessments Outside Normal Limits Among Subject
Without Abnormality at Baseline.
}
\details{
\itemize{
\item 
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{vst02_2_lyt}: \code{vst02_2} Layout

\item \code{vst02_2_pre}: \code{vst02_2} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  dm_zoom_to("advs") \%>\%
  filter(PARAM \%in\% c("Temperature")) \%>\%
  dm_update_zoomed() \%>\%
  vst02_2_pre()

vst02_2_main(adam_db = db)
vst02_2_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
syn_test_data() \%>\%
  vst02_2_pre()
}
