% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_rules.R
\docType{data}
\name{yes_no_rule}
\alias{yes_no_rule}
\title{Yes/No rule in title case}
\format{
An object of class \code{rule} (inherits from \code{character}) of length 8.
}
\usage{
yes_no_rule
}
\description{
Yes/No rule in title case
}
\keyword{datasets}
