% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ael01_nollt.R
\docType{data}
\name{ael01_nollt_1_main}
\alias{ael01_nollt_1_main}
\alias{ael01_nollt_1_pre}
\alias{ael01_nollt_1_check}
\alias{ael01_nollt_1_post}
\alias{ael01_nollt_1}
\title{\code{AEL01_NOLLT} Listing 1 (Default) Glossary of Preferred Terms and Investigator-Specified Terms.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
ael01_nollt_1_main(
  adam_db,
  dataset = "adae",
  key_cols = c("AEBODSYS", "AEDECOD"),
  disp_cols = "AETERM",
  deco = std_deco("AEL01_NOLLT"),
  ...
)

ael01_nollt_1_pre(
  adam_db,
  dataset = "adae",
  key_cols = c("AEBODSYS", "AEDECOD"),
  disp_cols = "AETERM",
  new_lbls = list(AETERM = "Investigator-Specified\\n Adverse Event Term"),
  ...
)

ael01_nollt_1_check(adam_db, dataset, vars)

ael01_nollt_1_post(tlg, ...)

ael01_nollt_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{key_cols}{(\code{character}) names of columns that should be treated as key columns when rendering the listing.
Key columns allow you to group repeat occurrences.}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{new_lbls}{(named \code{list} of \code{strings}) list where names correspond to variable names in \code{dataset} and
strings to the new labels to apply to the named variables. Set to \code{NULL} to use default labels.}

\item{vars}{(\code{character}) variables to be included in the listing.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{AEL01_NOLLT} Listing 1 (Default) Glossary of Preferred Terms and Investigator-Specified Terms.
}
\details{
\itemize{
\item Removes duplicate rows.
\item By default, uses dataset \code{adae}, sorting by key columns \code{AEBODSYS} and \code{AEDECOD}.
\item If using with a dataset other than \code{adae}, be sure to specify the desired labels for variables in
\code{key_cols} and \code{disp_cols}, and pre-process missing data.
}
}
\section{Functions}{
\itemize{
\item \code{ael01_nollt_1_main()}: Main TLG function

\item \code{ael01_nollt_1_pre()}: Preprocessing

\item \code{ael01_nollt_1_check()}: Checks

\item \code{ael01_nollt_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the \code{dataset} table with columns specified by \code{key_cols} and \code{disp_cols}.
}
}
\examples{
run(ael01_nollt_1, syn_data)
}
\keyword{datasets}
