% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\name{aet01_1}
\alias{aet01_1}
\title{AET01 Table}
\usage{
aet01_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET01"),
  .study = list(armvar = "ACTARM", totals_for = c("DTHFL", "DCSREAS"), lbl_overall =
    NULL)
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
AET01 Table
}
\examples{
# TODO: add examples
}
