% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\docType{data}
\name{aet01_1_main}
\alias{aet01_1_main}
\alias{aet01_1_lyt}
\alias{aet01_1_pre}
\alias{aet01_1_check}
\alias{aet01_1}
\title{\code{AET01} Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet01_1_main(
  adam_db,
  armvar = "ACTARM",
  lbl_overall = NULL,
  prune_0 = FALSE,
  deco = std_deco("AET01"),
  safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL",
    "RELWD", "RELDSM", "SEV"),
  lbl_safety_var = NULL,
  ...
)

aet01_1_lyt(armvar, lbl_overall, deco, safety_var, lbl_safety_var, ...)

aet01_1_pre(adam_db, ...)

aet01_1_check(
  adam_db,
  req_tables = c("adsl", "adae"),
  armvar = "ACTARM",
  safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL",
    "RELWD", "RELDSM", "SEV"),
  ...
)

aet01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{safety_var}{(\code{character}) the safety variables to be summarized.}

\item{lbl_safety_var}{(\code{character}) the labels of the safety variables to be summarized.}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}
}
\description{
\code{AET01} Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.
}
\details{
\itemize{
\item Does not remove rows with zero counts by default.
}
}
\section{Functions}{
\itemize{
\item \code{aet01_1_main()}: Main TLG function

\item \code{aet01_1_lyt()}: Layout

\item \code{aet01_1_pre()}: Preprocessing

\item \code{aet01_1_check()}: Checks

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adsl} table with the \code{"DTHFL"} and \code{"DCSREAS"} columns.
\item \code{adam_db} object must contain an \code{adae} table with the columns passed to \code{safety_var}.
}
}
\examples{
run(aet01_1, syn_test_data(), armvar = "ARM")
}
\keyword{datasets}
