% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\name{aet01_2}
\alias{aet01_2}
\alias{aet01_2_lyt}
\title{\code{AET01} Table 2 (Supplementary) Overview of Deaths and Adverse Events Summary Table 2.}
\usage{
aet01_2(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET01"),
  safety_var = .study$safety_var,
  lbl_safety_var = var_labels_for(adam_db$adae, .study$safety_var),
  .study = list(armvar = "ARM", lbl_overall = NULL, safety_var = c("FATAL", "SER",
    "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL", "RELWD", "RELDSM", "CTC35"))
)

aet01_2_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET01"),
  safety_var = .study$safety_var,
  lbl_safety_var = .study$lbl_safety_var,
  .study = list(armvar = "ARM", lbl_overall = NULL, safety_var = c("FATAL", "SER",
    "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL", "RELWD", "RELDSM", "CTC35", "CTC45",
    "SEV", "SMQ01", "SMQ02", "CQ01"), lbl_safety_var = c("FATAL", "SER", "SERWD",
    "SERDSM", "RELSER", "WD", "DSM", "REL", "RELWD", "RELDSM", "CTC35", "CTC45", "SEV",
    "SMQ01", "SMQ02", "CQ01"))
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{safety_var}{(\code{character}) the safety variables to be summarized.}

\item{lbl_safety_var}{(\code{character}) the labels of the safety variables to be summarized.}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{



}
\description{
Overview of death and details of adverse events.
}
\section{Functions}{
\itemize{
\item \code{aet01_2_lyt}: 
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  preprocess_data("aet01_2")

aet01_2(db, armvar = "ARM")

aet01_2_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("AET01")
)
}
