% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\docType{data}
\name{aet01_2_main}
\alias{aet01_2_main}
\alias{aet01_2_lyt}
\alias{aet01_2_pre}
\alias{aet01_2_check}
\alias{aet01_2}
\title{\code{AET01} Table 2 (Supplementary) Overview of Deaths and Adverse Events Summary Table 2. Overview of death and summary
of adverse events with medical concepts.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet01_2_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("AET01"),
  safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL",
    "RELWD", "RELDSM", "SEV"),
  lbl_safety_var = NULL,
  medconcept_var = c("SMQ01", "SMQ02", "CQ01"),
  lbl_medconcept_var = NULL,
  ...
)

aet01_2_lyt(
  arm_var,
  lbl_overall,
  deco,
  safety_var,
  lbl_safety_var,
  medconcept_var,
  lbl_medconcept_var,
  ...
)

aet01_2_pre(adam_db, ...)

aet01_2_check(
  adam_db,
  req_tables = c("adsl", "adae"),
  arm_var = "ACTARM",
  safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL",
    "RELWD", "RELDSM", "SEV"),
  medconcept_var = c("SMQ01", "SMQ02", "CQ01"),
  ...
)

aet01_2
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{safety_var}{(\code{character}) the safety variables to be summarized.}

\item{lbl_safety_var}{(\code{character}) the labels of the safety variables to be summarized.}

\item{medconcept_var}{(\code{character}) the medical concept variables to be summarized.}

\item{lbl_medconcept_var}{(\code{character}) the label of the medical concept variables to be summarized.}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}
}
\description{
\code{AET01} Table 2 (Supplementary) Overview of Deaths and Adverse Events Summary Table 2. Overview of death and summary
of adverse events with medical concepts.
}
\details{
\itemize{
\item Does not remove rows with zero counts by default.
}
}
\section{Functions}{
\itemize{
\item \code{aet01_2_main()}: Main TLG function

\item \code{aet01_2_lyt()}: Layout

\item \code{aet01_2_pre()}: Preprocessing

\item \code{aet01_2_check()}: Checks

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adsl} table with the \code{"DTHFL"} and \code{"DCSREAS"} columns.
\item \code{adam_db} object must contain an \code{adae} table with the columns passed to \code{safety_var}.
}
}
\examples{
run(aet01_2, syn_data)
}
\keyword{datasets}
