% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01_aesi.R
\docType{data}
\name{aet01_aesi_1_main}
\alias{aet01_aesi_1_main}
\alias{aet01_aesi_1_lyt}
\alias{aet01_aesi_1_pre}
\alias{aet01_aesi_1_check}
\alias{aet01_aesi_post}
\alias{aet01_aesi_1}
\title{\code{AET01_AESI} Table 1 (Default) Adverse Event of Special Interest Summary Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet01_aesi_1_main(
  adam_db,
  arm_var = "ACTARM",
  aesi_vars = list(NA_character_),
  deco = std_deco("AET01_AESI"),
  ...
)

aet01_aesi_1_lyt(arm_var, aesi_vars, deco, lbl_aesi_vars)

aet01_aesi_1_pre(
  adam_db,
  req_tables = c("adsl", "adae"),
  arm_var = "ACTARM",
  ...
)

aet01_aesi_1_check(adam_db, req_tables = c("adsl", "adae"), arm_var = "ACTARM")

aet01_aesi_post(tlg, prune_0 = FALSE, deco = std_deco("AET01_AESI"), ...)

aet01_aesi_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{aesi_vars}{(\code{list}) the AESI variables to be included in the summary. Defaults to \code{NA}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{lbl_aesi_vars}{(\code{character}) the labels of the AESI variables to be summarized.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
\code{AET01_AESI} Table 1 (Default) Adverse Event of Special Interest Summary Table.
}
\details{
\itemize{
\item Does not remove rows with zero counts by default.
}
}
\section{Functions}{
\itemize{
\item \code{aet01_aesi_1_main()}: Main TLG function

\item \code{aet01_aesi_1_lyt()}: Layout

\item \code{aet01_aesi_1_pre()}: Preprocessing

\item \code{aet01_aesi_1_check()}: Checks

\item \code{aet01_aesi_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adae} table with columns \code{"AEOUT"}, \code{"AEACN"}, \code{"AECONTRT"}, \code{"AESER"},
\code{"AREL"}, and the column specified by \code{arm_var}.
\item \code{aesi_vars} may contain any/all of the following variables to display: \code{"ALLRESWD"}, \code{"ALLRESDSM"},
\code{"ALLRESCONTRT"}, \code{"NOTRESWD"}, \code{"NOTRESDSM"}, \code{"NOTRESCONTRT"}, \code{"SERWD"}, \code{"SERDSM"}, \code{"SERCONTRT"},
\code{"RELWD"}, \code{"RELDSM"}, \code{"RELCONTRT"}, \code{"RELSER"}.
\item \code{aesi_vars} variable prefixes are defined as follows:
\itemize{
\item \code{"ALLRES"} = "all non-fatal adverse events resolved"
\item \code{"NOTRES"} = "at least one unresolved or ongoing non-fatal adverse event"
\item \code{"SER"} = "serious adverse event"
\item \code{"REL"} = "related adverse event"
}
\item \code{aesi_vars} variable suffixes are defined as follows:
\itemize{
\item \code{"WD"} = "patients with study drug withdrawn"
\item \code{"DSM"} = "patients with dose modified/interrupted"
\item \code{"CONTRT"} = "patients with treatment received"
}
\item Several \code{aesi_vars} can be added to the table at once:
\itemize{
\item \code{aesi_vars = "ALL"} will include all possible \code{aesi_vars}.
\item Including \code{"ALL_XXX"} in \code{aesi_vars} where \code{XXX} is one of the prefixes listed above will include all
\code{aesi_vars} with that prefix.
}
}
}
\examples{
run(aet01_aesi_1, syn_data)
}
\keyword{datasets}
