% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet01.R
\name{aet01_lyt}
\alias{aet01_lyt}
\title{aet01 Layout}
\usage{
aet01_lyt(
  arm_var,
  lbl_overall,
  safety_var,
  lbl_safety_var,
  medconcept_var,
  lbl_medconcept_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{safety_var}{(\code{character}) the safety variables to be summarized.}

\item{lbl_safety_var}{(\code{character}) the labels of the safety variables to be summarized.}
}
\description{
aet01 Layout
}
\keyword{internal}
