% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_1}
\alias{aet02_1}
\title{Adverse Events by System Organ Class and Preferred Term Table}
\usage{
aet02_1(
  adsl,
  adae,
  armvar = .study$armvar,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
Adverse Events by System Organ Class and Preferred Term Table
}
\examples{

library(scda)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae

aet02_1(adsl, adae)

adae <- adae \%>\% var_relabel(AEBODSYS = "Medra System Organ Class")
aet02_1(adsl, adae)
}
