% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\docType{data}
\name{aet02_1_main}
\alias{aet02_1_main}
\alias{aet02_1_lyt}
\alias{aet02_1_pre}
\alias{aet02_1_check}
\alias{aet02_1}
\title{\code{AET02_1} Pipeline}
\format{
a \code{tlg_pipeline_S4} object with the following slots:
\itemize{
\item \code{main} the \code{chevron::aet02_1_main} function.
\item \code{preprocess} the  \code{chevron::aet02_1_pre} function.
\item \code{postprocess} the identity function.
\item \code{check} no checks.
\item \code{adam_datasets} \code{"adsl"} and \code{"adae"}.
}
}
\usage{
aet02_1_main(
  adam_db,
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

aet02_1_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  lbl_aebodsys = "Body System or Organ Class",
  lbl_aedecod = "Dictionary-Derived Term",
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

aet02_1_pre(adam_db, ...)

aet02_1_check(
  adam_db,
  req_tables = c("adsl", "adae"),
  armvar = .study$actualarm,
  .study = list(actualarm = "ACTARM"),
  ...
)

aet02_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}
}
\description{
\code{AET02_1} Pipeline of the class \code{tlg_pipeline_S4}
}
\details{
\itemize{
\item Numbers represent absolute numbers of subject and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Dictionary-Derived Code (\code{AEDECOD}) by highest overall frequencies.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_1_main}: \code{aet02_1} main table

\code{AET02} Table 1 (Default) Adverse Events by System Organ Class and Preferred Term Table 1

The \code{AET02} table provides an overview of the number of subjects experiencing adverse events and the number of advert
events categorized by Body System and Dictionary-Derived Term.

\item \code{aet02_1_lyt}: \code{aet02_1} Layout

\item \code{aet02_1_pre}: \code{aet02_1} Preprocessing

\item \code{aet02_1_check}: \code{aet02_1} Checks
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  aet02_1_pre()

aet02_1_main(adam_db = db) \%>\% head(15)

# alternatively adam_db also accepts a names list
aet02_1_main(adam_db = list(adsl = db$adsl, adae = db$adae)) \%>\% head()

aet02_1_main(db, lbl_overall = "All Patients") \%>\% head()

db_m <- db \%>\%
  dm_zoom_to(adae) \%>\%
  mutate(AEBODSYS = formatters::with_label(AEBODSYS, "Medra System Organ Class")) \%>\%
  dm_update_zoomed()

aet02_1_main(db_m) \%>\% head()
aet02_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("AET02")
)
syn_test_data() \%>\%
  aet02_1_pre()
}
\keyword{datasets}
