% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\docType{data}
\name{aet02_1_main}
\alias{aet02_1_main}
\alias{aet02_1_lyt}
\alias{aet02_1_pre}
\alias{aet02_1_check}
\alias{aet02_1}
\title{\code{AET02} Table 1 (Default) Adverse Events by System Organ Class and Preferred Term Table 1}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet02_1_main(
  adam_db,
  lyt_ls = list(aet02_1_lyt),
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

aet02_1_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  lbl_aebodsys = "MedDRA System Organ Class",
  lbl_aedecod = "MedDRA Preferred Term",
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

aet02_1_pre(adam_db, ...)

aet02_1_check(
  adam_db,
  req_tables = c("adsl", "adae"),
  armvar = .study$actualarm,
  .study = list(actualarm = "ACTARM"),
  ...
)

aet02_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_ls}{(\verb{list of functions}) creating \verb{Pre-data Table Layout}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{req_tables}{(\code{character}) names of the required tables.}
}
\description{
The \code{AET02} table provides an overview of the number of subjects experiencing adverse events and the number of advert
events categorized by Body System and Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of subject and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Dictionary-Derived Code (\code{AEDECOD}) by highest overall frequencies.
\item Missing values in \code{AEBODSYS}, and \code{AEDECOD} are labeled by \verb{No Coding available}.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_1_main()}: Main TLG function

\item \code{aet02_1_lyt()}: Layout

\item \code{aet02_1_pre()}: Preprocessing

\item \code{aet02_1_check()}: Checks

}}
\examples{
run(aet02_1, syn_test_data())
}
\keyword{datasets}
