% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_1}
\alias{aet02_1}
\alias{aet02_1_lyt}
\title{AET02 Table 1 (Default) Adverse Events by System Organ Class and Preferred Term Table 1}
\usage{
aet02_1(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet02_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "Body System or Organ Class",
  lbl_AEDECOD = "Dictionary-Derived Term",
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the adam Datasets.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AEBODSYS}{(\code{string}) text label for AEBODSYS.}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}
}
\value{

}
\description{
The AET02 table provides an overview of the number of patients experiencing advert events and the number of advert
events categorized by Body System and Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Dictionary-Derived Code (\code{AEDECOD}) by highest overall frequencies.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_1_lyt}: 
}}

\examples{
library(dm)
library(rtables)

db <- syn_test_data() \%>\%
   dm_select_tbl(adsl, adae)

db <- db \%>\%
  (std_filter("aet02_1"))() \%>\%
  (std_mutate("aet02_1"))()

aet02_1(adam_db = db) \%>\% head(15)

# Additional Examples
db_s <- db \%>\%
  dm_filter(adsl, SEX == "F")

aet02_1(adam_db = db_s) \%>\% head()

# alternatively adam_db also accepts a names list
aet02_1(adam_db = list(adsl = db$adsl, adae = db$adae)) \%>\% head()

aet02_1(db, lbl_overall = "All Patients") \%>\% head()

db_m <- db \%>\%
  dm_zoom_to(adae) \%>\%
  mutate(AEBODSYS = with_label(AEBODSYS, "Medra System Organ Class")) \%>\%
  dm_update_zoomed()

aet02_1(db_m) \%>\% head()

aet02_1_lyt(
 armvar = "ACTARM",
 lbl_overall = "",
 deco = std_deco("AET02")
)
}
