% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_2}
\alias{aet02_2}
\alias{aet02_2_lyt}
\title{\code{AET02} Table 2 (Supplementary) Adverse Events by System Organ Class, High Level Term and Preferred Term Table 2}
\usage{
aet02_2(
  adam_db,
  armvar = .study$actarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(actarm = "ACTARM", lbl_overall = NULL)
)

aet02_2_lyt(
  armvar = .study$actarm,
  lbl_overall = .study$lbl_overall,
  lbl_aebodsys = "AEBODSYS",
  lbl_aehlt = "AEHLT",
  lbl_aedecod = "AEDECOD",
  deco = std_deco("AET02"),
  .study = list(actarm = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aehlt}{(\code{character}) text label for \code{AEHLT}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}
}
\description{
The \code{AET02_2} table provides an overview of the number of patients experiencing adverse events and the number of
adverse events categorized by Body System, High Level Term and Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Body System or Organ Class, High Level Term and Dictionary-Derived Term hierarchically by highest overall
frequencies.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_2_lyt}: 
}}

\examples{

library(dm)
library(rtables)

db <- syn_test_data() \%>\%
  preprocess_data("aet02_2")


aet02_2(db) \%>\% head(15)

# Additional Examples
aet02_2(db, lbl_overall = "All Patients") \%>\% head()

db_m <- db \%>\%
  dm_zoom_to(adae) \%>\%
  mutate(AEBODSYS = with_label(AEBODSYS, "MedDRA System Organ Class")) \%>\%
  dm_update_zoomed()

aet02_2(db_m) \%>\% head()
aet02_2_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  lbl_aebodsys = "Body System or Organ Class",
  lbl_aehlt = "High Level Term",
  lbl_aedecod = "Dictionary-Derived Term",
  deco = std_deco("AET02")
)
}
