% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_2}
\alias{aet02_2}
\title{Adverse Events by System Organ Class and Dictionary-Derived Term Table}
\usage{
aet02_2(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{remove 0 count rows}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
The AET02_2 table provides an overview of the number of patients experiencing advert events and the number of advert
events categorized by Body System, High Level Term and Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Body System or Organ Class, High Level Term and Dictionary-Derived Term hierarchically by highest overall frequencies.
}
}
\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

aet02_2(adsl, adae)
aet02_2(adsl, adae, lbl_overall = "All Patients")

adae <- adae \%>\% var_relabel(AEBODSYS = "Medra System Organ Class")
aet02_2(adsl, adae)

}
