% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\docType{data}
\name{aet02_3_lyt}
\alias{aet02_3_lyt}
\alias{aet02_3_pre}
\alias{aet02_3}
\title{\code{AET02} Table 3 (Supplementary) Adverse Events by Dictionary-Derived Term Table 3.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet02_3_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  lbl_aedecod = "MedDRA Preferred Term",
  deco = std_deco("AET02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

aet02_3_pre(adam_db, ...)

aet02_3
}
\arguments{
\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}

\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}
}
\description{
The \code{AET02_3} table provides an overview of the number of patients experiencing adverse events and the number of
adverse events categorized by Dictionary-Derived Term.
}
\section{Functions}{
\itemize{
\item \code{aet02_3_lyt()}: Layout

\item \code{aet02_3_pre()}: Preprocessing

}}
\examples{
aet02_3_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  lbl_aedecod = "Dictionary-Derived Term",
  deco = std_deco("AET02")
)
aet02_3_pre(syn_test_data())
}
\keyword{datasets}
