% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\docType{data}
\name{aet02_3_main}
\alias{aet02_3_main}
\alias{aet02_3_lyt}
\alias{aet02_3_pre}
\alias{aet02_3_post}
\alias{aet02_3}
\title{\code{AET02} Table 3 (Supplementary) Adverse Events by Dictionary-Derived Term Table 3.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet02_3_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  lbl_aedecod = "MedDRA Preferred Term",
  deco = std_deco("AET02"),
  ...
)

aet02_3_lyt(arm_var, lbl_overall, lbl_aedecod, deco)

aet02_3_pre(adam_db, ...)

aet02_3_post(tlg, prune_0 = TRUE, ...)

aet02_3
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{AET02_3} table provides an overview of the number of patients experiencing adverse events and the number of
adverse events categorized by Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Dictionary-Derived Code by highest overall frequencies.
\item Missing values of \code{AEDECOD} in \code{aead} are labeled by \verb{No Coding available}.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_3_main()}: Main TLG function

\item \code{aet02_3_lyt()}: Layout

\item \code{aet02_3_pre()}: Preprocessing

\item \code{aet02_3_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adae} table with the column \code{"AEDECOD"}.
}
}
\examples{
run(aet02_3, syn_data)
}
\keyword{datasets}
