% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet02.R
\name{aet02_3}
\alias{aet02_3}
\alias{aet02_3_lyt}
\title{AET02 Table 3 (Supplementary) Adverse Events by Dictionary-Derived Term Table 3}
\usage{
aet02_3(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet02_3_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEDECOD = "AEDECOD",
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the adam Datasets.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}
}
\value{

}
\description{
The AET02_3 table provides an overview of the number of patients experiencing advert events and the number of advert
events categorized by Dictionary-Derived Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort Dictionary-Derived Code by highest overall frequencies.
}
}
\section{Functions}{
\itemize{
\item \code{aet02_3_lyt}: 
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
   dm_select_tbl(adsl, adae)

db <- db \%>\%
  (std_filter("aet02_3"))() \%>\%
  (std_mutate("aet02_3"))()

aet02_3(adam_db = db) \%>\% head()

aet02_3(db, lbl_overall = "All Patients") \%>\% head()

aet02_3_lyt(
 armvar = "ACTARM",
 lbl_overall = "",
 lbl_AEDECOD = "Dictionary-Derived Term",
 deco = std_deco("AET02")
)
}
