% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\name{aet03_1}
\alias{aet03_1}
\title{Advert Events by Greatest Intensity}
\usage{
aet03_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{remove 0 count rows}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
An adverse events table categorized by System Organ Class, Dictionary-Derived Term  and Greatest intensity
}
\details{
\itemize{
\item Default Adverse Events by Greatest Intensity table
\item Numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm
\item Does not include a total column by default
\item Sort by Body System or Organ Class (SOC) and Dictionary-Derived Term (PT)
}
}
\examples{

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

aet03_1(adsl, adae)
aet02_1(adsl, adae, lbl_overall = "All Patients")

}
