% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\docType{data}
\name{aet03_1}
\alias{aet03_1}
\title{\code{AET03_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet03_1
}
\description{
\code{AET03_1}
}
\seealso{
\code{\link[=aet03_1_main]{aet03_1_main()}}
}
\keyword{datasets}
