% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\name{aet03_1}
\alias{aet03_1}
\alias{aet03_1_lyt}
\title{AET03 Table 1 (Default) Advert Events by Greatest Intensity Table 1}
\usage{
aet03_1(
  adam_db,
  armvar = .study$armvar,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet03_1_lyt(
  armvar = .study$armvar,
  lbl_AEBODSYS = "",
  lbl_AEDECOD = "",
  severity_grade = .study$severity_grade,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL, severity_grade = c("MILD",
    "MODERATE", "SEVERE", "LIFE THREATENING"))
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the adam Datasets.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AEBODSYS}{(\code{string}) text label for AEBODSYS.}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}

\item{severity_grade}{(\verb{vector of strings}) describing the severity levels present in the data set.}
}
\value{

}
\description{
An adverse events table categorized by System Organ Class, Dictionary-Derived Term  and Greatest intensity.
}
\details{
\itemize{
\item Default Adverse Events by Greatest Intensity table
\item Numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm
\item Does not include a total column by default
\item Sort by Body System or Organ Class (SOC) and Dictionary-Derived Term (PT)
}
}
\section{Functions}{
\itemize{
\item \code{aet03_1_lyt}: 
}}

\examples{

library(dm)
library(rtables)

db <- syn_test_data() \%>\%
   dm_select_tbl(adsl, adae)

db <- db \%>\%
  (std_filter_fun("aet03_1"))() \%>\%
  (std_mutate_fun("aet03_1"))()

aet03_1(db)
aet03_1(db, lbl_overall = "All Patients")

aet03_1_lyt(armvar = "ACTARM",
 lbl_AEBODSYS = "Body System or Organ Class",
 lbl_AEDECOD = "Dictionary-Derived Term",
 lbl_overall = NULL,
 deco = std_deco("AET03"))
}
