% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet03.R
\docType{data}
\name{aet03_1_main}
\alias{aet03_1_main}
\alias{aet03_1_lyt}
\alias{aet03_1_pre}
\alias{aet03_1}
\title{\code{AET03} Table 1 (Default) Advert Events by Greatest Intensity Table 1.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet03_1_main(
  adam_db,
  lyt_fun = aet03_1_lyt,
  armvar = .study$actualarm,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

aet03_1_lyt(
  armvar = .study$actualarm,
  lbl_aebodsys = "",
  lbl_aedecod = "",
  severity_grade = .study$severity_grade,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, severity_grade = c("MILD",
    "MODERATE", "SEVERE", "LIFE THREATENING")),
  ...
)

aet03_1_pre(adam_db, ...)

aet03_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_fun}{(\code{function} or \verb{list of functions}) creating \verb{Pre-data Table Layout}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}

\item{severity_grade}{(\verb{vector of character}) describing the severity levels present in the dataset.}
}
\description{
An adverse events table categorized by System
Organ Class, Dictionary-Derived Term  and Greatest intensity.
}
\details{
\itemize{
\item Default Adverse Events by Greatest Intensity table.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by Body System or Organ Class (\code{SOC}) and Dictionary-Derived Term (\code{PT}).
}
}
\section{Functions}{
\itemize{
\item \code{aet03_1_main()}: Main TLG function

\item \code{aet03_1_lyt()}: Layout

\item \code{aet03_1_pre()}: Preprocessing

}}
\examples{
run(aet03_1, syn_test_data())
}
\keyword{datasets}
