% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\docType{data}
\name{aet04_1}
\alias{aet04_1}
\title{\code{AET04_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
aet04_1
}
\description{
\code{AET04_1}
}
\seealso{
\code{\link[=aet04_1_main]{aet04_1_main()}}
}
\keyword{datasets}
