% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\docType{data}
\name{aet04_1_main}
\alias{aet04_1_main}
\alias{aet04_1_lyt}
\alias{aet04_1_pre}
\alias{aet04_1}
\title{\code{AET04} Table 1 (Default) Adverse Events by Highest \code{NCI} \code{CTACAE} \code{AE} Grade Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet04_1_main(
  adam_db,
  armvar = "ACTARM",
  group_grades = NULL,
  lbl_overall = NULL,
  prune_0 = TRUE,
  deco = std_deco("AET04"),
  ...
)

aet04_1_lyt(
  armvar,
  lbl_aebodsys,
  lbl_aedecod,
  group_grades,
  lbl_overall,
  deco,
  ...
)

aet04_1_pre(adam_db, ...)

aet04_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{group_grades}{(\code{list}) putting in correspondence severity levels and labels.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}
}
\description{
The \code{AET04} table provides an
overview of adverse event with the highest \code{NCI} \code{CTCAE} grade per individual.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Events with missing grading values are excluded.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sort Body System or Organ Class and Dictionary-Derived Term by highest overall frequencies. Analysis Toxicity
Grade is sorted by severity.
}
}
\section{Functions}{
\itemize{
\item \code{aet04_1_main()}: Main TLG function

\item \code{aet04_1_lyt()}: Layout

\item \code{aet04_1_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adae} table with the columns \code{"AETOXGR"}, \code{"AEBODSYS"} and \code{"AEDECOD"}.
}
}
\examples{
group_grades <- list(
  "Any Grade" = c("1", "2", "3", "4", "5"),
  "Grade 1-2" = c("1", "2"),
  "Grade 3-5" = c("3", "4", "5")
)

run(aet04_1, syn_test_data(), group_grades = group_grades)
}
\keyword{datasets}
