% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\name{aet04_1}
\alias{aet04_1}
\alias{aet04_1_lyt}
\title{AET04 Table 1 (Default) Adverse Events by Highest NCI CTACAE AE Grade Table 1}
\usage{
aet04_1(
  adam_db,
  armvar = .study$armvar,
  group_grades = .study$group_grades,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET04"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL, group_grades = list(`Any Grade`
    = c("1", "2", "3", "4", "5"), `Grade 1-2` = c("1", "2"), `Grade 3-4` = c("3", "4"),
    `Grade 5` = c("5")))
)

aet04_1_lyt(
  armvar = .study$armvar,
  lbl_AEBODSYS = "AEBODSYS",
  lbl_AEDECOD = "AEDECOD",
  group_grades = .study$group_grades,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET04"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL, group_grades = list(`Any Grade`
    = c("1", "2", "3", "4", "5"), `Grade 1-2` = c("1", "2"), `Grade 3-4` = c("3", "4"),
    `Grade 5` = c("5")))
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{group_grades}{(\code{list}) putting in correspondence severity levels and labels.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AEBODSYS}{(\code{string}) text label for AEBODSYS.}

\item{lbl_AEDECOD}{(\code{string}) text label for AEDECOD.}
}
\description{
The AET04 table provides an overview of adverse event with the highest NCI CTCAE grade per individual.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sort Body System or Organ Class and Dictionary-Derived Term by highest overall frequencies. Analysis Toxicity
Grade is sorted by severity.
}
}
\section{Functions}{
\itemize{
\item \code{aet04_1_lyt}: 
}}

\examples{
library(dm)
library(rtables)

db <- syn_test_data() \%>\%
   preprocess_data("aet03_1")

aet04_1(db)
aet04_1(db, prune_0 = FALSE)
aet04_1(db, lbl_overall = "All Patients")

aet04_1(db, group_grades = list(
  "Any Grade" = c("1", "2", "3", "4", "5"),
  "Grade 1-2" = c("1", "2"),
  "Grade 3-5" = c("3", "4", "5")
))


aet04_1_lyt(armvar = "ACTARM")
}
