% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\docType{data}
\name{aet04_1_main}
\alias{aet04_1_main}
\alias{aet04_1_lyt}
\alias{aet04_1_pre}
\alias{aet04_1_post}
\alias{aet04_1}
\title{\code{AET04} Table 1 (Default) Adverse Events by Highest \code{NCI} \code{CTACAE} \code{AE} Grade Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet04_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  lbl_aebodsys = "MedDRA System Organ Class",
  lbl_aedecod = "MedDRA Preferred Term",
  grade_groups = NULL,
  deco = std_deco("AET04"),
  ...
)

aet04_1_lyt(
  arm_var,
  lbl_overall,
  lbl_aebodsys,
  lbl_aedecod,
  toxicity_grade,
  grade_groups,
  deco
)

aet04_1_pre(adam_db, ...)

aet04_1_post(tlg, prune_0 = TRUE, ...)

aet04_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aebodsys}{(\code{string}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{string}) text label for \code{AEDECOD}.}

\item{grade_groups}{(\code{list}) putting in correspondence toxicity grades and labels.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{toxicity_grade}{(\code{character}) putting in correspondence toxicity levels.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{AET04} table provides an
overview of adverse event with the highest \code{NCI} \code{CTCAE} grade per individual.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Events with missing grading values are excluded.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sort Body System or Organ Class and Dictionary-Derived Term by highest overall frequencies. Analysis Toxicity
Grade is sorted by severity.
}
}
\section{Functions}{
\itemize{
\item \code{aet04_1_main()}: Main TLG function

\item \code{aet04_1_lyt()}: Layout

\item \code{aet04_1_pre()}: Preprocessing

\item \code{aet04_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adae} table with the columns \code{"AEBODSYS"}, \code{"AEDECOD"} and \code{"ATOXGR"}.
}
}
\examples{
grade_groups <- list(
  "Grade 1-2" = c("1", "2"),
  "Grade 3-4" = c("3", "4"),
  "Grade 5" = c("5")
)

grade_groups <- list(
  "Grade 1-2" = c("1", "2"),
  "Grade 3-5" = c("3", "4", "5")
)

run(aet04_1, syn_data, grade_groups = grade_groups)
}
\keyword{datasets}
