% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet04.R
\name{aet04_1}
\alias{aet04_1}
\alias{aet04_1_lyt}
\alias{aet04_1_pre}
\title{\code{AET04} Table 1 (Default) Adverse Events by Highest NCI CTACAE AE Grade Table 1}
\usage{
aet04_1(
  adam_db,
  armvar = .study$actualarm,
  group_grades = .study$group_grades,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET04"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, group_grades = NULL)
)

aet04_1_lyt(
  armvar = .study$actualarm,
  lbl_aebodsys = "AEBODSYS",
  lbl_aedecod = "AEDECOD",
  group_grades = .study$group_grades,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET04"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, group_grades = NULL)
)

aet04_1_pre(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{group_grades}{(\code{list}) putting in correspondence severity levels and labels.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_aebodsys}{(\code{character}) text label for \code{AEBODSYS}.}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}
}
\description{
The \code{AET04} table provides an overview of adverse event with the highest NCI CTCAE grade per individual.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sort Body System or Organ Class and Dictionary-Derived Term by highest overall frequencies. Analysis Toxicity
Grade is sorted by severity.
}
}
\section{Functions}{
\itemize{
\item \code{aet04_1_lyt}: \code{aet04_1} Layout

\item \code{aet04_1_pre}: \code{aet04_1} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  aet04_1_pre()

aet04_1(db)

aet04_1(
  db,
  prune_0 = FALSE,
  lbl_overall = "All Patients",
  group_grades = list(
    "Any Grade" = c("1", "2", "3", "4", "5"),
    "Grade 1-2" = c("1", "2"),
    "Grade 3-5" = c("3", "4", "5")
  )
)
aet04_1_lyt(armvar = "ACTARM")
syn_test_data() \%>\%
  aet04_1_pre()
}
