% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet05_all.R
\docType{data}
\name{aet05_all_main}
\alias{aet05_all_main}
\alias{aet05_all_pre}
\alias{aet05_all}
\title{\code{AET05_ALL} Table 1 (Default) Adverse Event Rate Adjusted for Patient-Years at Risk - All Occurrences.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
aet05_all_main(adam_db, arm_var = "ACTARM", ...)

aet05_all_pre(adam_db, ...)

aet05_all
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) the arm variable used for arm splitting.}

\item{...}{Further arguments passed to \code{tern::control_incidence_rate()}.}
}
\description{
The \code{AET05_ALL} table produces the standard adverse event rate adjusted for patient-years at risk summary
considering all occurrences.
}
\details{
\itemize{
\item Total patient-years at risk is the sum over all patients of the time intervals (in years).
\item Split columns by arm, typically \code{ACTARM}.
\item The table allows confidence level to be adjusted, default is 2-sided 5\%.
\item Keep zero count rows by default.
}
}
\section{Functions}{
\itemize{
\item \code{aet05_all_main()}: Main TLG function

\item \code{aet05_all_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adaette} table with the columns \code{"PARAMCD"}, \code{"AVAL"}, and \code{"CNSR"}.
}
}
\examples{
run(aet05_all, syn_data, time_unit_output = 100)

run(aet05_all, syn_data, conf_level = 0.90, conf_type = "exact", time_unit_output = 100)
}
\keyword{datasets}
