% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet05.R
\name{aet05_lyt}
\alias{aet05_lyt}
\title{\code{aet05} Layout}
\usage{
aet05_lyt(arm_var, param_var, vars, n_events, control)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{param_var}{(\code{string}) variable for parameter code.}

\item{vars}{(\code{string}) variable for the primary analysis variable to be iterated over.}

\item{n_events}{(\code{string}) variable to count the number of events observed.}

\item{control}{(\code{list}) parameters for estimation details, specified by using the helper function
control_incidence_rate().}
}
\description{
\code{aet05} Layout
}
\keyword{internal}
