% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aet10.R
\name{aet10_lyt}
\alias{aet10_lyt}
\title{aet10 Layout}
\usage{
aet10_lyt(arm_var, lbl_overall, lbl_aedecod)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aedecod}{(\code{character}) text label for \code{AEDECOD}.}
}
\description{
aet10 Layout
}
\keyword{internal}
