% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{args}
\alias{args}
\title{Get Arguments List}
\usage{
args(x, simplify = FALSE)
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{simplify}{(\code{flag}) whether to simplify the output.}
}
\description{
Get Arguments List
}
