% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_args}
\alias{assert_args}
\title{Assert that names are valid arguments for the \code{run} method of a \code{chevron_tlg} object.}
\usage{
assert_args(object, args_names)
}
\arguments{
\item{object}{(\code{chevron_tlg}) defining the valid arguments.}

\item{args_names}{(\code{character}) to assert.}
}
\value{
invisible \code{NULL} if all \code{args_names} are valid argument for the \code{chevron_tlg} object or an warning otherwise.
}
\description{
Assert that names are valid arguments for the \code{run} method of a \code{chevron_tlg} object.
}
\examples{
\dontrun{
assert_args(aet01_1, "lbl_overall")
assert_args(
  aet01_1,
  c("lbl_", "lbl_overall", "armvar", "_var", "xxxx")
)
}
}
\keyword{internal}
