% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_colnames}
\alias{assert_colnames}
\title{Check if strings are column names of a \code{data.frame}}
\usage{
assert_colnames(df, x, null_ok = TRUE)
}
\arguments{
\item{df}{(\code{data.frame}) input to check for the presence of column names.}

\item{x}{(vector of \code{character}) the names of the columns to be checked.}

\item{null_ok}{(\code{logical}) can \code{x} be NULL.}
}
\description{
Provides a clearer error message in the case of missing variable.
}
\examples{
\dontrun{
assert_colnames(mtcars, c("speed", "seats"), null_ok = TRUE)

my_colnames <- NULL
assert_colnames(mtcars, my_colnames, null_ok = FALSE)
}
}
