% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_one_tablenames}
\alias{assert_one_tablenames}
\title{Assert that at least one name is among table names of a \code{list} of \code{data.frame}.}
\usage{
assert_one_tablenames(db, tab, null_ok = TRUE, qualifier = NULL)
}
\arguments{
\item{db}{(\code{list} of \code{data.frame}) input to check for the presence or tables.}

\item{tab}{(\code{character}) the names of the tables to be checked.}

\item{null_ok}{(\code{flag}) can \code{x} be NULL.}

\item{qualifier}{(\code{string}) to be returned if the check fails.}
}
\description{
Assert that at least one name is among table names of a \code{list} of \code{data.frame}.
}
\examples{
\dontrun{

lsd <- list(
  mtcars = mtcars,
  iris = iris
)

assert_one_tablenames(lsd, c("mtcars", "x", "y"), qualifier = "first test:")
}
}
\keyword{internal}
