% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_valid_var}
\alias{assert_valid_var}
\title{Check whether var is valid}
\usage{
assert_valid_var(x, label, na_ok, empty_ok, ...)
}
\arguments{
\item{x}{value of col_split variable}

\item{label}{(\code{string}) hints.}

\item{na_ok}{(\code{flag}) whether NA value is allowed}

\item{empty_ok}{(\code{flag}) whether length 0 value is allowed.}

\item{...}{Not used.}
}
\description{
Check whether var is valid
}
\details{
This function checks the variable values are valid or not.
}
