% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bol_YN}
\alias{bol_YN}
\title{Convert Y N values used in CDISC to R boolean object}
\usage{
bol_YN(x)
}
\arguments{
\item{x}{character vector with Y and N}
}
\description{
Empty strings are mapped to \code{NA}
}
\examples{
\dontrun{
chevron:::bol_YN(c("Y", "Y", "N", "", NA))
}
}
