% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfbt01.R
\docType{data}
\name{cfbt01_main}
\alias{cfbt01_main}
\alias{cfbt01_pre}
\alias{cfbt01_post}
\alias{cfbt01}
\title{\code{CFBT01} Change from Baseline By Visit Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
cfbt01_main(
  adam_db,
  dataset,
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  page_by = TRUE,
  ...
)

cfbt01_pre(adam_db, dataset, ...)

cfbt01_post(tlg, prune_0 = TRUE, ...)

cfbt01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{row_split_var}{(\code{character}) row split variable other than \code{PARAMCD}.}

\item{summaryvars}{(\code{character}) variables to be analyzed. The label attribute of the corresponding column in
table of \code{adam_db} is used as label.}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed}

\item{precision}{(named \code{list} of \code{integer}) where names are values found in the \code{PARAMCD} column and the the values
indicate the number of digits that should be represented for \code{min}, \code{max} and \code{median}. \code{Mean} and \code{sd} are
represented with one more decimal of precision.}

\item{default_precision}{(\code{integer}) the default number of digits.}

\item{page_by}{(\code{flag}) indicator whether the parameter row split is by page.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{CFBT01} table provides an
overview of the actual values and its change from baseline of each respective arm
over the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{cfbt01_main()}: Main TLG function

\item \code{cfbt01_pre()}: Preprocessing

\item \code{cfbt01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain table named as \code{dataset} with the columns specified in \code{summaryvars}.
}
}
\examples{
library(dunlin)
proc_data <- log_filter(
  syn_data,
  PARAMCD \%in\% c("DIABP", "SYSBP"), "advs"
)
run(cfbt01, proc_data, dataset = "advs")
}
\keyword{datasets}
