% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfbt01.R
\name{cfbt01_lyt}
\alias{cfbt01_lyt}
\title{\code{cfbt01} Layout}
\usage{
cfbt01_lyt(
  arm_var,
  summaryvars,
  summaryvars_lbls,
  row_split_var,
  row_split_lbl,
  visitvar,
  lbl_avisit,
  lbl_param,
  precision,
  default_precision,
  page_by
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\code{character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\code{character}) the label of the variables to be analyzed.}

\item{row_split_var}{(\code{character}) row split variable other than \code{PARAMCD}.}

\item{row_split_lbl}{(\code{character}) label of further row splits.}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{lbl_avisit}{(\code{string}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{precision}{(named \code{list} of \code{integer}) where names are values found in the \code{PARAMCD} column and the the values
indicate the number of digits that should be represented for \code{min}, \code{max} and \code{median}. \code{Mean} and \code{sd} are
represented with one more decimal of precision.}

\item{default_precision}{(\code{integer}) the default number of digits.}

\item{page_by}{(\code{flag}) indicator whether the parameter row split is by page.}
}
\description{
\code{cfbt01} Layout
}
\keyword{internal}
