% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{chevron-package}
\alias{chevron}
\alias{chevron-package}
\title{\code{chevron} package}
\description{
Standard Table, Listings, and Graphs (TLG) library used in clinical trials.
}
\author{
\strong{Maintainer}: Benoit Falquet \email{benoit.falquet@roche.com}

Authors:
\itemize{
  \item Liming Li \email{liming.li@roche.com}
  \item Xiaoli Duan \email{xiaoli.duan@roche.com}
}

Other contributors:
\itemize{
  \item Tim Barnett \email{timothy.barnett@roche.com} [contributor]
  \item Tian Fang \email{tian.fang@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
