% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{chevron-package}
\alias{chevron}
\alias{chevron-package}
\title{\code{chevron} package}
\description{
Provide standard tables, listings, and graphs (TLG) libraries used in clinical trials. This package implements a structure to reformat the data with 'dunlin', create reporting tables using 'rtables' and 'tern' with standardized input arguments to enable quick generation of standard outputs. In addition, it also provides comprehensive data checks and script generation functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/chevron/}
  \item \url{https://github.com/insightsengineering/chevron/}
  \item Report bugs at \url{https://github.com/insightsengineering/chevron/issues}
}

}
\author{
\strong{Maintainer}: Liming Li \email{liming.li@roche.com}

Authors:
\itemize{
  \item Benoit Falquet \email{benoit.falquet@roche.com}
  \item Xiaoli Duan \email{xiaoli.duan@roche.com}
}

Other contributors:
\itemize{
  \item Adrian Waddell \email{waddell.adrian@gene.com} [contributor]
  \item Chenkai Lv \email{chenkai.lv@roche.com} [contributor]
  \item Pawel Rucki \email{pawel.rucki@roche.com} [contributor]
  \item Tim Barnett \email{timothy.barnett@roche.com} [contributor]
  \item Tian Fang \email{tian.fang@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
