% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{chevron-package}
\alias{chevron}
\alias{chevron-package}
\title{Chevron Package}
\description{
Standard Table, Listings, and Graphs (TLG) library used in clinical trials.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Adrian Waddell \email{adrian.waddell@roche.com}
  \item Benoit Falquet \email{benoit.falquet@roche.com}
}

Other contributors:
\itemize{
  \item Roche [copyright holder, funder]
  \item Gabriel Becker \email{gabembecker@gmail.com} [contributor]
  \item Tim Barnett \email{timothy.barnett@roche.com} [contributor]
}

}
