% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{chevron-package}
\alias{chevron}
\alias{chevron-package}
\title{\code{chevron} package}
\description{
Standard Table, Listings, and Graphs (TLG) library used in clinical trials.
}
\author{
\strong{Maintainer}: Benoit Falquet \email{benoit.falquet@roche.com}

Authors:
\itemize{
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Heng Wang \email{wang.heng@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
