% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R, R/aet04.R
\docType{class}
\name{chevron_tlg-class}
\alias{chevron_tlg-class}
\alias{.chevron_tlg}
\alias{chevron_tlg}
\alias{aet04_2}
\title{\code{chevron_tlg} class}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
chevron_tlg(
  main = function(adam_db, ...) adam_db,
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_
)

aet04_2
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{_main} function of \code{chevron}.}

\item{preprocess}{(\code{function}) returning a pre-processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post-processed \code{tlg}.}

\item{adam_datasets}{(\code{character}) representing the name of the table from an \code{AdAM} dataset required for \code{tlg}
creation.}
}
\description{
The \code{chevron_tlg} class associates a \code{preprocess} function, a main \code{tlg} function and \code{AdAM} tables names and
optionally a \code{postprocess} function.
}
\section{Functions}{
\itemize{
\item \code{chevron_tlg}: Default Constructor
}}

\section{Slots}{

\describe{
\item{\code{main}}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{*_main} function from \code{chevron}.}

\item{\code{preprocess}}{(\code{function}) returning a pre-processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{*_pre} function from \code{chevron}.}

\item{\code{postprocess}}{(\code{function}) returning a post-processed \code{tlg}.}

\item{\code{adam_datasets}}{(\code{character}) representing the name of the tables from an \code{AdAM} dataset required for \code{tlg}
creation.}
}}

\note{
To ensure the correct execution of the workflow additional validation criteria are:
\itemize{
\item the first argument of the \code{main} function must be \code{adam_db}, the input \code{dm} object to pre-process.
\item the first argument of the \code{preprocess} function must be \code{adam_db}, the input \code{dm} object to create \code{tlg} output.
\item the first argument of the \code{postprocess} function must be \code{tlg}, the input \code{TableTree} object to post-process.
}
}
\examples{
x <- chevron_tlg(aet01_1_main, aet01_1_pre, adam_datasets = c("adsl", "adae"))
}
\seealso{
\code{\link[=aet04_2_main]{aet04_2_main()}}
}
\keyword{datasets}
