% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4class.R, R/aet01.R, R/aet02.R,
%   R/aet03.R, R/aet04.R, R/cmt01a.R, R/cmt02_pt.R, R/dmt01.R, R/dst01.R,
%   R/dtht01.R, R/egt01.R, R/egt02.R, R/ext01.R, R/lbt01.R, R/mht01.R,
%   R/vst01.R, R/vst02.R
\docType{class}
\name{chevron_tlg-class}
\alias{chevron_tlg-class}
\alias{.chevron_tlg}
\alias{chevron_tlg}
\alias{aet01_1}
\alias{aet01_2}
\alias{aet02_1}
\alias{aet02_2}
\alias{aet02_3}
\alias{aet03_1}
\alias{aet04_1}
\alias{aet04_2}
\alias{cmt01a_1}
\alias{cmt01a_2}
\alias{cmt01a_3}
\alias{cmt02_pt_1}
\alias{dmt01_1}
\alias{dst01_1}
\alias{dst01_2}
\alias{dst01_3}
\alias{dtht01_1}
\alias{egt01_1}
\alias{egt02_1}
\alias{egt02_2}
\alias{ext01_1}
\alias{ext01_2}
\alias{lbt01_1}
\alias{mht01_1}
\alias{vst01_1}
\alias{vst02_1}
\alias{vst02_2}
\title{\code{chevron_tlg} Class}
\format{
An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.

An object of class \code{chevron_tlg} of length 1.
}
\usage{
chevron_tlg(
  main = function(adam_db, ...) adam_db,
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_
)

aet01_1

aet01_2

aet02_1

aet02_2

aet02_3

aet03_1

aet04_1

aet04_2

cmt01a_1

cmt01a_2

cmt01a_3

cmt02_pt_1

dmt01_1

dst01_1

dst01_2

dst01_3

dtht01_1

egt01_1

egt02_1

egt02_2

ext01_1

ext01_2

lbt01_1

mht01_1

vst01_1

vst02_1

vst02_2
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{_main} function of \code{chevron}.}

\item{preprocess}{(\code{function}) returning a pre processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post processed \code{tlg}.}

\item{adam_datasets}{(\code{character}) representing the name of the table from an \code{AdAM} data set required for \code{tlg}
creation.}
}
\description{
The \code{chevron_tlg} class associates a \code{preprocess} function, a main \code{tlg} function and \code{AdAM} tables names and
optionally a \code{postprocess} function.
}
\section{Functions}{
\itemize{
\item \code{chevron_tlg}: Default Constructor
}}

\section{Slots}{

\describe{
\item{\code{main}}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{*_main} function from \code{chevron}.}

\item{\code{preprocess}}{(\code{function}) returning a pre processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{*_pre} function from \code{chevron}.}

\item{\code{postprocess}}{(\code{function}) returning a post processed \code{tlg}.}

\item{\code{adam_datasets}}{(\code{character}) representing the name of the tables from an \code{AdAM} data set required for \code{tlg}
creation.}
}}

\note{
To ensure the correct execution of the pipeline workflow additional validation criteria are:
\itemize{
\item the first argument of the \code{main} function must be \code{adam_db}, the input \code{dm} object to preprocess.
\item the first argument of the \code{preprocess} function must be \code{adam_db}, the input \code{dm} object to create \code{tlg} output.
\item the first argument of the \code{postprocess} function must be \code{tlg}, the input \code{TableTree} object to post process.
}
}
\examples{
x <- chevron_tlg(aet01_1_main, aet01_1_pre, adam_datasets = c("adsl", "adae"))
}
\seealso{
\code{\link[=aet01_1_main]{aet01_1_main()}}

\code{\link[=aet01_2_main]{aet01_2_main()}}

\code{\link[=aet02_1_main]{aet02_1_main()}}

\code{\link[=aet02_2_main]{aet02_2_main()}}

\code{\link[=aet02_3_main]{aet02_3_main()}}

\code{\link[=aet03_1_main]{aet03_1_main()}}

\code{\link[=aet04_1_main]{aet04_1_main()}}

\code{\link[=aet04_2_main]{aet04_2_main()}}

\code{\link[=cmt01a_1_main]{cmt01a_1_main()}}

\code{\link[=cmt01a_2_main]{cmt01a_2_main()}}

\code{\link[=cmt01a_3_main]{cmt01a_3_main()}}

\code{\link[=cmt02_pt_1_main]{cmt02_pt_1_main()}}

\code{\link[=dmt01_1_main]{dmt01_1_main()}}

\code{\link[=dst01_1_main]{dst01_1_main()}}

\code{\link[=dst01_2_main]{dst01_2_main()}}

\code{\link[=dst01_3_main]{dst01_3_main()}}

\code{\link[=dtht01_1_main]{dtht01_1_main()}}

\code{\link[=egt01_1_main]{egt01_1_main()}}

\code{\link[=egt02_1_main]{egt02_1_main()}}

\code{\link[=egt02_2_main]{egt02_2_main()}}

\code{\link[=ext01_1_main]{ext01_1_main()}}

\code{\link[=ext01_2_main]{ext01_2_main()}}

\code{\link[=lbt01_1_main]{lbt01_1_main()}}

\code{\link[=mht01_1_main]{mht01_1_main()}}

\code{\link[=vst01_1_main]{vst01_1_main()}}

\code{\link[=vst02_1_main]{vst02_1_main()}}

\code{\link[=vst02_2_main]{vst02_2_main()}}
}
\keyword{datasets}
