% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\docType{class}
\name{chevron_tlg-class}
\alias{chevron_tlg-class}
\alias{.chevron_tlg}
\alias{chevron_tlg}
\title{\code{chevron_tlg} class}
\usage{
chevron_tlg(
  main = function(adam_db, ...) adam_db,
  lyt = list(function(...) basic_table()),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = report_null,
  adam_datasets = NA_character_
)
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{_main} function of \code{chevron}.}

\item{lyt}{(\code{function}, \verb{list of functions}, \code{PreDataTableLayouts} or \verb{list of PreDataTableLayouts}) typically one of the \verb{_lyt} function
of \code{chevron}.}

\item{preprocess}{(\code{function}) returning a pre-processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post-processed \code{tlg}.}

\item{adam_datasets}{(\code{character}) representing the name of the table from an \code{ADaM} dataset required for \code{tlg}
creation.}
}
\description{
The \code{chevron_tlg} class associates a \code{preprocess} function, a main \code{tlg} function and \code{AdAM} tables names and
optionally a \code{postprocess} function.
}
\section{Functions}{
\itemize{
\item \code{chevron_tlg()}: Default Constructor

}}
\section{Slots}{

\describe{
\item{\code{main}}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{*_main} function from \code{chevron}.}

\item{\code{lyt}}{(\verb{list of function}).  Typically one of the \verb{*_lyt} function from \code{chevron}wrapped in a \code{list}.}

\item{\code{preprocess}}{(\code{function}) returning a pre-processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{*_pre} function from \code{chevron}.}

\item{\code{postprocess}}{(\code{function}) returning a post-processed \code{tlg}.}

\item{\code{adam_datasets}}{(\code{character}) representing the name of the tables from an \code{AdAM} dataset required for \code{tlg}
creation.}
}}

\note{
To ensure the correct execution of the workflow additional validation criteria are:
\itemize{
\item the first argument of the \code{main} function must be \code{adam_db}, the input \code{dm} object to pre-process.
\item the first argument of the \code{preprocess} function must be \code{adam_db}, the input \code{dm} object to create \code{tlg} output.
\item the first argument of the \code{postprocess} function must be \code{tlg}, the input \code{TableTree} object to post-process.
}
}
\examples{
x <- chevron_tlg(aet01_1_main, aet01_1_lyt, aet01_1_pre, adam_datasets = c("adsl", "adae"))
}
