% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4class.R
\docType{class}
\name{chevron_tlg-class}
\alias{chevron_tlg-class}
\alias{.chevron_tlg}
\alias{chevron_tlg}
\alias{chevron_t-class}
\alias{.chevron_t}
\alias{chevron_table}
\alias{chevron_l-class}
\alias{.chevron_l}
\alias{chevron_listing}
\alias{chevron_g-class}
\alias{.chevron_g}
\alias{chevron_graph}
\alias{chevron_t}
\alias{chevron_l}
\alias{chevron_g}
\title{\code{chevron_tlg} class}
\usage{
chevron_t(
  main = function(adam_db, lyt_ls, ...) build_table(lyt_ls[[1]](), adam_db[[1]]),
  lyt = list(function(...) basic_table()),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = report_null,
  adam_datasets = NA_character_,
  ...
)

chevron_l(
  main = function(adam_db, ...) data.frame(),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_,
  ...
)

chevron_g(
  main = function(adam_db, ...) ggplot2::ggplot(),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_,
  ...
)

chevron_tlg(
  main,
  preprocess,
  postprocess,
  adam_datasets,
  type = c(NA, "table", "listing", "graph"),
  ...
)
}
\arguments{
\item{main}{(\code{function}) returning a \code{tlg}, with \code{adam_db} as first argument and \code{...} as last argument. Typically
one of the \verb{_main} function of \code{chevron}.}

\item{lyt}{(a single \code{function} returning \code{PreDataTableLayouts} object or \code{PreDataTableLayouts} object or \code{list} of
either \code{functions} or \code{PreDataTableLayouts} type of elements) typically one of the \verb{_lyt} function of \code{chevron}.
Functions passed to \code{lyt}, whether as a single \code{function} or as a \verb{list of functions}, must have the \code{...} formal
argument.}

\item{preprocess}{(\code{function}) returning a pre-processed \code{dm} object, with \code{adam_db} as first argument and \code{...} as
last argument. Typically one of the \verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post-processed \code{tlg}, with \code{tlg} as first argument.}

\item{adam_datasets}{(\code{character}) representing the names of the tables from an \code{ADaM} dataset required for \code{tlg}
creation.}

\item{...}{used to pass additional class specific argument. see \link{chevron_t}}

\item{type}{(\code{string}) indicating the subclass.}
}
\description{
The \code{chevron_tlg} class associates a \code{preprocess} function, a main \code{tlg} function and \code{AdAM} tables names and
optionally a \code{postprocess} function.

\code{chevron_t}, a subclass of \link{chevron_tlg} with specific validation criteria to handle graph creation

\code{chevron_l}, a subclass of \link{chevron_tlg} with specific validation criteria to handle listing creation

\code{chevron_g}, a subclass of \link{chevron_tlg} with specific validation criteria to handle graph creation
}
\section{Functions}{
\itemize{
\item \code{chevron_tlg()}: Default Constructor

}}
\section{Slots}{

\describe{
\item{\code{main}}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{*_main} function from \code{chevron}.}

\item{\code{preprocess}}{(\code{function}) returning a pre-processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{*_pre} function from \code{chevron}.}

\item{\code{postprocess}}{(\code{function}) returning a post-processed \code{tlg}.}

\item{\code{adam_datasets}}{(\code{character}) representing the name of the tables from an \code{AdAM} dataset required for \code{tlg}
creation.}

\item{\code{lyt}}{(\verb{list of function}).  Typically one of the \verb{*_lyt} function from \code{chevron} wrapped in a \code{list}.}
}}

\note{
To ensure the correct execution of the workflow additional validation criteria are:
\itemize{
\item the first argument of the \code{main} function must be \code{adam_db}, the input \code{dm} object to pre-process. In addition, it
must possess the \code{...} argument.
\item the \code{lyt} function must posses the \code{...} argument.
\item the first argument of the \code{preprocess} function must be \code{adam_db}, the input \code{dm} object to create \code{tlg} output.
In addition, it must possess the \code{...} argument.
\item the first argument of the \code{postprocess} function must be \code{tlg}, the input \code{TableTree} object to post-process.
}
}
\examples{
chevron_t_obj <- chevron_t()
chevron_t_obj <- chevron_t(postprocess = function(tlg, indent, ...) {
  rtables::table_inset(tlg) <- indent
  tlg
})

chevron_l_obj <- chevron_l()

chevron_g_obj <- chevron_g()
chevron_g_obj <- chevron_g(postprocess = function(tlg, title, ...) tlg + ggplot2::labs(main = title))

x <- chevron_tlg(aet01_1_main, aet01_1_lyt, aet01_1_pre, adam_datasets = c("adsl", "adae"), type = "table")
}
