% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cml02a_gl.R
\name{cml02a_gl_main}
\alias{cml02a_gl_main}
\alias{cml02a_gl_pre}
\alias{cml02a_gl}
\title{\code{CML02A_GL} Listing 1 (Default) Concomitant Medication Class Level 2, Preferred Name, and Investigator-Specified
Terms.}
\usage{
cml02a_gl_main(
  adam_db,
  dataset = "adcm",
  key_cols = c("ATC2", "CMDECOD"),
  disp_cols = c("ATC2", "CMDECOD", "CMTRT"),
  split_into_pages_by_var = NULL,
  unique_rows = TRUE,
  ...
)

cml02a_gl_pre(
  adam_db,
  dataset = "adcm",
  disp_cols = c("ATC2", "CMDECOD", "CMTRT"),
  ...
)

cml02a_gl
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{key_cols}{(\code{character}) names of columns that should be treated as key columns when rendering the listing.
Key columns allow you to group repeat occurrences.}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{split_into_pages_by_var}{(\code{character} or \code{NULL}) the name of the variable to split the listing by.}

\item{unique_rows}{(\code{flag}) whether to keep only unique rows in listing.}

\item{...}{not used.}
}
\value{
the main function returns an \code{rlistings} or a \code{list} object.

the preprocessing function returns a \code{list} of \code{data.frame}.
}
\description{
\code{CML02A_GL} Listing 1 (Default) Concomitant Medication Class Level 2, Preferred Name, and Investigator-Specified
Terms.
}
\section{Functions}{
\itemize{
\item \code{cml02a_gl_main()}: Main TLG function

\item \code{cml02a_gl_pre()}: Preprocessing

}}
\examples{
run(cml02a_gl, syn_data)
}
