% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cml02a_gl.R
\docType{data}
\name{cml02a_gl_main}
\alias{cml02a_gl_main}
\alias{cml02a_gl_pre}
\alias{cml02a_gl_post}
\alias{cml02a_gl}
\title{\code{CML02A_GL} Listing 1 (Default) Concomitant Medication Class Level 2, Preferred Name, and Investigator-Specified
Terms.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
cml02a_gl_main(
  adam_db,
  dataset = "adcm",
  key_cols = c("ATC2", "CMDECOD", "CMTRT"),
  disp_cols = names(adam_db[[dataset]]),
  default_formatting = list(all = fmt_config(align = "left"), numeric = fmt_config(align
    = "center")),
  col_formatting = NULL,
  unique_rows = TRUE,
  ...
)

cml02a_gl_pre(
  adam_db,
  dataset = "adcm",
  key_cols = c("ATC2", "CMDECOD", "CMTRT"),
  ...
)

cml02a_gl_post(tlg, ...)

cml02a_gl
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{key_cols}{(\code{character}) names of columns that should be treated as key columns when rendering the listing.
Key columns allow you to group repeat occurrences.}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{default_formatting}{(\code{list}) the default format of the listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{col_formatting}{(\code{list}) the format of specific listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{unique_rows}{(\code{flag}) whether to keep only unique rows in listing.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{CML02A_GL} Listing 1 (Default) Concomitant Medication Class Level 2, Preferred Name, and Investigator-Specified
Terms.
}
\details{
\itemize{
\item Removes duplicate rows.
\item By default, uses dataset \code{adae}, sorting by key columns \code{AEBODSYS} and \code{AEDECOD}.
\item If using with a dataset other than \code{adae}, be sure to specify the desired labels for variables in
\code{key_cols} and \code{disp_cols}, and pre-process missing data.
}
}
\section{Functions}{
\itemize{
\item \code{cml02a_gl_main()}: Main TLG function

\item \code{cml02a_gl_pre()}: Preprocessing

\item \code{cml02a_gl_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the \code{dataset} table with columns specified by \code{key_cols} and \code{disp_cols}.
}
}
\examples{
run(cml02a_gl, syn_data)
}
\keyword{datasets}
