% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\docType{data}
\name{cmt01a_main}
\alias{cmt01a_main}
\alias{cmt01a_pre}
\alias{cmt01a_post}
\alias{cmt01a}
\title{\code{CMT01A} Concomitant Medication by Medication Class and Preferred Name.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
cmt01a_main(
  adam_db,
  arm_var = "ARM",
  incl_n_treatment = TRUE,
  row_split_var = "ATC2",
  medname_var = "CMDECOD",
  lbl_overall = NULL,
  ...
)

cmt01a_pre(adam_db, ...)

cmt01a_post(
  tlg,
  prune_0 = TRUE,
  sort_by_freq = FALSE,
  row_split_var = "ATC2",
  medname_var = "CMDECOD",
  ...
)

cmt01a
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{incl_n_treatment}{(\code{flag}) include total number of treatments per medication.}

\item{row_split_var}{(\code{character}) the variable defining the medication category. By default \code{ATC2}.}

\item{medname_var}{(\code{string}) variable name of medical treatment name.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}

\item{sort_by_freq}{(\code{flag}) whether to sort medication class by frequency.}
}
\description{
A concomitant medication
table with the number of subjects and the total number of treatments by medication class.
}
\details{
\itemize{
\item Data should be filtered for concomitant medication. \code{(ATIREL == "CONCOMITANT")}.
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by medication class alphabetically and within medication class by decreasing total number of patients with
the specific medication.
}
}
\section{Functions}{
\itemize{
\item \code{cmt01a_main()}: Main TLG function

\item \code{cmt01a_pre()}: Preprocessing

\item \code{cmt01a_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adcm} table with the columns specified in \code{row_split_var} and \code{medname_var}
as well as \code{"CMSEQ"}.
}
}
\examples{
library(dplyr)

proc_data <- syn_data
proc_data$adcm <- proc_data$adcm \%>\%
  filter(ATIREL == "CONCOMITANT")

run(cmt01a, proc_data)
}
\keyword{datasets}
