% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\docType{data}
\name{cmt01a_2_main}
\alias{cmt01a_2_main}
\alias{cmt01a_2_pre}
\alias{cmt01a_2}
\title{\code{CMT01A} Table 2 (Supplementary) Concomitant Medication by Medication Class and Preferred Name (Classes sorted by
frequency).}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
cmt01a_2_main(
  adam_db,
  lyt_ls = list(cmt01a_1_lyt),
  armvar = "ARM",
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  lbl_overall = NULL,
  prune_0 = TRUE,
  deco = std_deco("CMT01A"),
  ...
)

cmt01a_2_pre(adam_db, medcat_var = "ATC2", medname_var = "CMDECOD", ...)

cmt01a_2
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_ls}{(\verb{list of functions} returning \code{PreDataTableLayout} object. Typically one of the \verb{_lyt} function of
\code{chevron} stored in a \code{list}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{medcat_var}{(\code{character}) the variable defining the medication category. By default \code{ATC2}.}

\item{lbl_medcat_var}{(\code{character}) label for the variable defining the medication category.}

\item{medname_var}{(\code{character}) the variable defining the medication name. By default \code{CMDECOD}.}

\item{lbl_medname_var}{(\code{character}) label for the variable defining the medication name.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}
}
\description{
A concomitant medication table with the number of subjects and the total number of treatments by
medication class and medication name sorted by frequencies.
}
\details{
\itemize{
\item Data should be filtered for concomitant medication. \code{(ATIREL == "CONCOMITANT")}.
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by medication class frequency and within medication class by decreasing total number of patients with
the specific medication.
}
}
\section{Functions}{
\itemize{
\item \code{cmt01a_2_main()}: Main TLG function

\item \code{cmt01a_2_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adcm} table with the columns specified in \code{medcat_var} and \code{medname_var} as well
as \code{"CMSEQ"}.
}
}
\examples{
library(magrittr)
library(dm)

db <- syn_test_data() \%>\%
  dm_zoom_to("adcm") \%>\%
  filter(.data$ATIREL == "CONCOMITANT") \%>\%
  dm_update_zoomed()

run(cmt01a_2, db)
}
\keyword{datasets}
