% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\name{cmt01a_3}
\alias{cmt01a_3}
\alias{cmt01a_1_lyt}
\title{\code{CMT01A} Table 3 (Supplementary) Concomitant Medication by Medication Class and Preferred Name (Total number of
treatments per medication class suppressed).}
\usage{
cmt01a_3(
  adam_db,
  armvar = .study$armvar,
  medcat_var = "ATC2",
  medname_var = "CMDECOD",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("CMT01A"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

cmt01a_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Level 2 Text",
  medname_var = "CMDECOD",
  lbl_medname_var = "Standardized Medication Name",
  deco = std_deco("CMT01A"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{medcat_var}{(\code{character}) the variable defining the medication category. By default \code{ATC2}.}

\item{medname_var}{(\code{character}) the variable defining the medication name. By default \code{CMDECOD}.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_cmdecod}{(\code{character}) text label for \code{CMDECOD}.}
}
\description{
The \code{CMT01A}
}
\details{
\itemize{
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by medication class alphabetically and within medication class by decreasing total number of patients with
the specific medication.
}
}
\section{Functions}{
\itemize{
\item \code{cmt01a_1_lyt}: 
}}

\examples{
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
  preprocess_data("cmt01a_1")

cmt01a_3(adam_db = db)

cmt01a_3_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("CMT01A")
)
}
