% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\name{cmt01a_3}
\alias{cmt01a_3}
\alias{cmt01a_3_lyt}
\alias{cmt01a_3_pre}
\title{\code{CMT01A} Table 3 (Supplementary) Concomitant Medication by Medication Class and Preferred Name (Total number of
treatments per medication class suppressed).}
\usage{
cmt01a_3(
  adam_db,
  armvar = .study$planarm,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("CMT01A"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

cmt01a_3_lyt(
  armvar = .study$planarm,
  lbl_overall = .study$lbl_overall,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  deco = std_deco("CMT01A"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

cmt01a_3_pre(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{medcat_var}{(\code{character}) the variable defining the medication category. By default \code{ATC2}.}

\item{lbl_medcat_var}{(\code{character}) the label for the medication category.}

\item{medname_var}{(\code{character}) the variable defining the medication name. By default \code{CMDECOD}.}

\item{lbl_medname_var}{(\code{character}) the label for the medication name.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
A concomitant medication table with the number of subjects and the total number of treatments by medication class
sorted alphabetically and medication name sorted by frequencies presented without the total number of treatments per
medication.
}
\details{
\itemize{
\item Data should be filtered for concomitant medication. \code{(ATIREL == "CONCOMITANT")}.
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by medication class alphabetically and within medication class by decreasing total number of patients with
the specific medication.
}
}
\section{Functions}{
\itemize{
\item \code{cmt01a_3_lyt}: 

\item \code{cmt01a_3_pre}: 
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  dm_zoom_to("adcm") \%>\%
  filter(ATIREL == "CONCOMITANT") \%>\%
  dm_update_zoomed() \%>\%
  cmt01a_3_pre()

cmt01a_3(adam_db = db)
cmt01a_3_lyt(
  armvar = "ARM",
  lbl_overall = NULL,
  deco = std_deco("CMT01A")
)
syn_test_data() \%>\%
  cmt01a_3_pre()
}
