% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt01a.R
\docType{data}
\name{cmt01a_3_main}
\alias{cmt01a_3_main}
\alias{cmt01a_3_lyt}
\alias{cmt01a_3_pre}
\alias{cmt01a_3}
\title{\code{CMT01A} Table 3 (Supplementary) Concomitant Medication by Medication Class and Preferred Name (Total number of
treatments per medication class suppressed).}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
cmt01a_3_main(
  adam_db,
  armvar = "ARM",
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  lbl_overall = NULL,
  prune_0 = TRUE,
  deco = std_deco("CMT01A"),
  ...
)

cmt01a_3_lyt(
  armvar = "ARM",
  lbl_overall = NULL,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  deco = std_deco("CMT01A"),
  ...
)

cmt01a_3_pre(adam_db, medcat_var = "ATC2", medname_var = "CMDECOD", ...)

cmt01a_3
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{medcat_var}{(\code{character}) the variable defining the medication category. By default \code{ATC2}.}

\item{lbl_medcat_var}{(\code{character}) the label for the medication category.}

\item{medname_var}{(\code{character}) the variable defining the medication name. By default \code{CMDECOD}.}

\item{lbl_medname_var}{(\code{character}) the label for the medication name.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}
}
\description{
A concomitant medication table with the number of subjects and the total
number of treatments by medication class sorted alphabetically and medication name sorted by frequencies presented
without the total number of treatments per medication.
}
\details{
\itemize{
\item Data should be filtered for concomitant medication. \code{(ATIREL == "CONCOMITANT")}.
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Sort by medication class alphabetically and within medication class by decreasing total number of patients with
the specific medication.
}
}
\section{Functions}{
\itemize{
\item \code{cmt01a_3_main()}: Main TLG function

\item \code{cmt01a_3_lyt()}: Layout

\item \code{cmt01a_3_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adcm} table with the columns specified in \code{medcat_var} and \code{medname_var} as well
as \code{"CMSEQ"}.
}
}
\examples{
library(magrittr)
library(dm)

db <- syn_test_data() \%>\%
  dm_zoom_to("adcm") \%>\%
  filter(.data$ATIREL == "CONCOMITANT") \%>\%
  dm_update_zoomed()

run(cmt01a_3, db)
}
\keyword{datasets}
