% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt02_pt.R
\name{cmt02_pt_1}
\alias{cmt02_pt_1}
\title{\code{CMT02_PT} Table 1 (Default)}
\usage{
cmt02_pt_1(
  adam_db,
  armvar = .study$armvar,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("CMT02_PT"),
  .study = list(armvar = "ARM", lbl_overall = "All patients")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{

}
\description{
\code{CMT02_PT} Table 1 (Default)
}
\details{
\itemize{
\item 
\item 
}
}
\examples{
library(dm)
library(rtables)

db <- syn_test_data() \%>\%
  preprocess_data("cmt02_pt_1")

cmt02_pt_1(adam_db = db)

}
