% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt02_pt.R
\docType{data}
\name{cmt02_pt_1_main}
\alias{cmt02_pt_1_main}
\alias{cmt02_pt_1_lyt}
\alias{cmt02_pt_1_pre}
\alias{cmt02_pt_1_post}
\alias{cmt02_pt_1}
\title{\code{CMT02_PT} Table 1 (Default) Concomitant Medications by Preferred Name.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
cmt02_pt_1_main(
  adam_db,
  armvar = "ARM",
  lbl_overall = NULL,
  deco = std_deco("CMT02_PT"),
  ...
)

cmt02_pt_1_lyt(armvar, lbl_overall, deco, ...)

cmt02_pt_1_pre(adam_db, ...)

cmt02_pt_1_post(tlg, prune_0 = TRUE, ...)

cmt02_pt_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
A concomitant medication table with the
number of subjects and the total number of treatments by medication name sorted by frequencies.
}
\details{
\itemize{
\item Data should be filtered for concomitant medication. \code{(ATIREL == "CONCOMITANT")}.
\item Order by decreasing total number of patients with the specific medication.
\item Does not include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{cmt02_pt_1_main()}: Main TLG function

\item \code{cmt02_pt_1_lyt()}: Layout

\item \code{cmt02_pt_1_pre()}: Preprocessing

\item \code{cmt02_pt_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adcm} table with the columns \code{"CMDECOD"} and \code{"CMSEQ"}.
}
}
\examples{
run(cmt02_pt_1, syn_data)
}
\keyword{datasets}
