% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxt01.R
\name{coxt01_lyt}
\alias{coxt01_lyt}
\title{\code{coxt01} Layout}
\usage{
coxt01_lyt(variables, col_split, lbl_vars, control, ...)
}
\arguments{
\item{variables}{(\code{list}) list of variables in a Cox proportional hazards regression model.}

\item{...}{not used.}
}
\description{
\code{coxt01} Layout
}
\keyword{internal}
