% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxt02.R
\docType{data}
\name{coxt02_1_main}
\alias{coxt02_1_main}
\alias{coxt02_1_lyt}
\alias{coxt02_1_pre}
\alias{coxt02_1_check}
\alias{coxt02_1_post}
\alias{coxt02_1}
\title{\code{COXT02} Table 1 (Default) Multi-Variable Cox Regression Model Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
coxt02_1_main(
  adam_db,
  arm_var = "ARM",
  time_var = "AVAL",
  censor_var = "CNSR",
  covariates = c("SEX", "RACE", "AAGE"),
  strata = NULL,
  lbl_vars = "Effect/Covariate Included in the Model",
  conf_level = 0.95,
  ties = "exact",
  deco = std_deco("COXT02"),
  ...
)

coxt02_1_lyt(variables, lbl_vars, conf_level, ties, deco)

coxt02_1_pre(
  adam_db,
  arm_var = "ARM",
  time_var = "AVAL",
  censor_var = "CNSR",
  covariates = c("SEX", "RACE", "AAGE"),
  strata = NULL,
  ...
)

coxt02_1_check(
  adam_db,
  arm_var,
  time_var,
  censor_var,
  covariates,
  strata,
  req_tables = c("adsl", "adtte")
)

coxt02_1_post(tlg, prune_0 = FALSE, ...)

coxt02_1
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) the arm variable used for arm splitting.}

\item{time_var}{(\code{string}) the time variable in a Cox proportional hazards regression model.}

\item{censor_var}{(\code{string}) the censor variable in a Cox proportional hazards regression model.}

\item{covariates}{(\code{character}) will be fitted and the corresponding effect will be estimated.}

\item{strata}{(\code{character}) will be fitted for the stratified analysis.}

\item{lbl_vars}{(\code{string}) text label for the a Cox regression model variables.}

\item{conf_level}{(\code{string}) confidence level of the interval when fitting a Cox regression model
and estimating hazard ratio to describe the effect size in a survival analysis.}

\item{ties}{(\code{string}) specifies the method for tie handling.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{variables}{(\code{list}) list of variables in a Cox proportional hazards regression model.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{COXT02} table follows the same principles as the general Cox model analysis
and produces the estimates for each of the covariates included in the model
(usually the main effects without interaction terms).
}
\details{
\itemize{
\item The table allows confidence level to be adjusted, default is 2-sided 5\%.
\item The stratified analysis is with DISCRETE tie handling (equivalent to \code{tern::control_coxreg(ties = "exact")} in R).
\item Model includes treatment plus specified covariate(s) as factor(s) or numeric(s),
with \code{"SEX"}, \code{"RACE"} and \code{"AAGE"} as default candidates.
\item The selection of the covariates and whether or not there is a selection process
(vs. a fixed, pre-specified list) needs to be specified in the DAP.
\item For pairwise comparisons using the hazard ratio, the value for the control group is the denominator.
\item Keep zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{coxt02_1_main()}: Main TLG function

\item \code{coxt02_1_lyt()}: Layout

\item \code{coxt02_1_pre()}: Preprocessing

\item \code{coxt02_1_check()}: Checks

\item \code{coxt02_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adtte} table with with \code{"PARAMCD"}, \code{"ARM"},
\code{"AVAL"}, \verb{"CNSR}, and the columns specified by \code{"covariates"} which is denoted as
\code{c("SEX", "RACE", "AAGE")} by default.
}
}
\examples{
library(dplyr)
library(dunlin)

proc_data <- log_filter(syn_data, PARAMCD == "CRSD", "adtte")

run(coxt02_1, proc_data)

run(coxt02_1, proc_data, covariates = c("SEX", "AAGE"), strata = c("RACE"), conf_level = 0.90)
}
\keyword{datasets}
