% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxt02.R
\docType{data}
\name{coxt02_1_main}
\alias{coxt02_1_main}
\alias{coxt02_1_lyt}
\alias{coxt02_1_pre}
\alias{coxt02_1_post}
\alias{coxt02_1}
\title{\code{COXT02} Table 1 (Default) Multi-Variable Cox Regression Model Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
coxt02_1_main(
  adam_db,
  arm_var = "ARM",
  lbl_term = "Effect/Covariate Included in the Model",
  time = "AVAL",
  event = "event",
  covariates = c("AGE", "SEX", "RACE"),
  conf_level = 0.95,
  deco = std_deco("COXT02"),
  ...
)

coxt02_1_lyt(lbl_term, conf_level, deco)

coxt02_1_pre(adam_db, ...)

coxt02_1_post(tlg, prune_0 = FALSE, ...)

coxt02_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}

\item{lbl_}{(\code{string}) text label for ``. If \code{NULL}, the value of the argument defaults to the label of
the corresponding column in the `ad` table.}
}
\description{
The \code{COXT02} table follows the same principles as the general Cox model analysis
and produces the estimates for each of the covariates included in the model
(usually the main effects without interaction terms).
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Order by body system alphabetically and within body system and medical condition by decreasing total number of
patients with the specific condition.
}
}
\section{Functions}{
\itemize{
\item \code{coxt02_1_main()}: Main TLG function

\item \code{coxt02_1_lyt()}: Layout

\item \code{coxt02_1_pre()}: Preprocessing

\item \code{coxt02_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{ad} table with columns \code{""} and \code{""}.
}
}
\examples{
run(coxt02_1, syn_data)
}
\keyword{datasets}
