% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml01.R
\docType{data}
\name{dml01_main}
\alias{dml01_main}
\alias{dml01_pre}
\alias{dml01_post}
\alias{dml01}
\title{\code{DML01} Listing 1 (Default) Demographics and Baseline Characteristics.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
dml01_main(
  adam_db,
  dataset = "adsl",
  key_cols = "ARM",
  disp_cols = names(adam_db[[dataset]]),
  default_formatting = list(all = fmt_config(align = "left"), numeric = fmt_config(align
    = "center")),
  col_formatting = NULL,
  unique_rows = TRUE,
  ...
)

dml01_pre(
  adam_db,
  dataset = "adsl",
  summaryvars = c("AGEGR1", "ETHNIC"),
  key_cols = "ARM",
  ...
)

dml01_post(tlg, ...)

dml01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{key_cols}{(\code{character}) names of columns that should be treated as key columns when rendering the listing.
Key columns allow you to group repeat occurrences.}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{default_formatting}{(\code{list}) the default format of the listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{col_formatting}{(\code{list}) the format of specific listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{unique_rows}{(\code{flag}) whether to keep only unique rows in listing.}

\item{...}{not used.}

\item{summaryvars}{(\code{character}) variables summarized in demographic table. The label attribute of the corresponding
column in \code{adsl} table of \code{adam_db} is used as label.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{DML01} Listing 1 (Default) Demographics and Baseline Characteristics.
}
\section{Functions}{
\itemize{
\item \code{dml01_main()}: Main TLG function

\item \code{dml01_pre()}: Preprocessing

\item \code{dml01_post()}: Postprocessing

}}
\examples{
res <- run(dml01, syn_data)
}
\keyword{datasets}
