% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml02.R
\docType{data}
\name{dml02_main}
\alias{dml02_main}
\alias{dml02_pre}
\alias{dml02_post}
\alias{dml02}
\title{\code{DML02} Listing 1 (Default) Randomization.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
dml02_main(
  adam_db,
  dataset = "adsl",
  arm_var = "ARM",
  disp_cols = names(adam_db[[dataset]]),
  default_formatting = list(all = fmt_config(align = "left"), numeric = fmt_config(align
    = "center")),
  col_formatting = NULL,
  unique_rows = TRUE,
  ...
)

dml02_pre(adam_db, dataset = "adsl", arm_var = "ARM", ...)

dml02_post(tlg, ...)

dml02
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{default_formatting}{(\code{list}) the default format of the listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{col_formatting}{(\code{list}) the format of specific listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{unique_rows}{(\code{flag}) whether to keep only unique rows in listing.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{DML02} Listing 1 (Default) Randomization.
}
\section{Functions}{
\itemize{
\item \code{dml02_main()}: Main TLG function

\item \code{dml02_pre()}: Preprocessing

\item \code{dml02_post()}: Postprocessing

}}
\examples{
res <- run(dml02, syn_data)
}
\keyword{datasets}
