% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\docType{data}
\name{dmt01_main}
\alias{dmt01_main}
\alias{dmt01_pre}
\alias{dmt01_post}
\alias{dmt01}
\title{\code{DMT01} Table 1 (Default) Demographics and Baseline Characteristics Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
dmt01_main(
  adam_db,
  arm_var = "ARM",
  summaryvars = c("AAGE", "AGEGR1", "SEX", "ETHNIC", "RACE"),
  lbl_overall = "All {Patient_label}",
  ...
)

dmt01_pre(adam_db, ...)

dmt01_post(tlg, prune_0 = TRUE, ...)

dmt01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\code{character}) variables summarized in demographic table. The label attribute of the corresponding
column in \code{adsl} table of \code{adam_db} is used as label.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
For each variable, summary statistics are
by default based on the number of patients in the corresponding \code{n} row.
}
\details{
\itemize{
\item Information from \code{ADSUB} are generally included into \code{ADSL} before analysis.
\item Default demographic and characteristics table
\item If not specified otherwise, numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows
\item Split columns by arm (planned or actual / code or description)
\item Include a total column by default
}
}
\section{Functions}{
\itemize{
\item \code{dmt01_main()}: Main TLG function

\item \code{dmt01_pre()}: Preprocessing

\item \code{dmt01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adsl} table with the columns specified in \code{summaryvars}.
}
}
\examples{
run(dmt01, syn_data)
}
\keyword{datasets}
