% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\docType{data}
\name{dmt01_1_main}
\alias{dmt01_1_main}
\alias{dmt01_1_lyt}
\alias{dmt01_1_pre}
\alias{dmt01_1}
\title{\code{DMT01} Table 1 (Default) Demographics and Baseline Characteristics Table 1.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
dmt01_1_main(
  adam_db,
  armvar = .study$planarm,
  summaryvars = .study$demo_vars,
  summaryvars_lbls = .study$demo_vars_lbls,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DMT01"),
  .study = list(planarm = "ARM", demo_vars = c("AGE", "BWGHTSI", "SEX", "COUNTRY",
    "RACE"), demo_vars_lbls = NULL, lbl_overall = "All Patients")
)

dmt01_1_lyt(
  armvar = .study$planarm,
  summaryvars = .study$demo_vars,
  summaryvars_lbls = .study$demo_vars_lbl,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DMT01"),
  .study = list(planarm = "ARM", demo_vars = c("AAGE", "AGEGR1", "SEX", "ETHNIC",
    "RACE"), demo_vars_lbl = c("Age (yr)", "Pooled Age Group 1 (yr)", "SEX", "ETHNIC",
    "RACE"), lbl_overall = "All Patients")
)

dmt01_1_pre(adam_db, ...)

dmt01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) variables summarized in demographic table.
Usually a vector containing the following one or more of the following:
\code{AAGE}, \code{AGEGR1}, \code{SEX}, \code{ETHNIC}, \code{RACE} and by default all of them.}

\item{summaryvars_lbls}{(\verb{vector of character}) labels corresponding to the analyzed variables.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}

\item{demo_vars}{(\verb{vector of strings}) variables summarized in demographic table.}

\item{demo_vars_lbls}{(\verb{vector of strings}) labels corresponding to the analyzed variables.}
}
\description{
For each variable, summary statistics are
by default based on the number of patients in the corresponding \code{n} row.
}
\details{
\itemize{
\item Information from \code{ADSUB} are generally included into \code{ADSL} before analysis.
\item Default demographic and characteristics table
\item If not specified otherwise, numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows
\item Split columns by arm (planned or actual / code or description)
\item Include a total column by default
}
}
\section{Functions}{
\itemize{
\item \code{dmt01_1_main()}: Main TLG function

\item \code{dmt01_1_lyt()}: Layout

\item \code{dmt01_1_pre()}: Preprocessing

}}
\examples{
library(dm)
library(magrittr)

db <- syn_test_data() \%>\%
  dmt01_1_pre()

dmt01_1_main(db, summaryvars = c("AGE", "RACE", "SEX"), lbl_overall = NULL)
dmt01_1_main(db,
  summaryvars = c("AGE", "RACE", "SEX"),
  summaryvars_lbls = c("Age (yr)", "Race", "Sex")
)
dmt01_1_lyt(armvar = "ACTARM")
dmt01_1_pre(syn_test_data())
}
\keyword{datasets}
