% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\name{dmt01_1}
\alias{dmt01_1}
\alias{dmt01_1_lyt}
\title{DMT01 Table 1 (Default) Demographics and Baseline Characteristics Table 1}
\usage{
dmt01_1(
  adam_db,
  armvar = .study$armvar,
  summaryvars = c("AAGE", "AGEGR1", "SEX", "ETHNIC", "RACE", "BWGHTSI"),
  summaryvars_lbls = var_labels_for(adam_db$adsl, summaryvars),
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DMT01"),
  .study = list(armvar = "ARM", lbl_overall = "All Patients")
)

dmt01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$summary_demo,
  summaryvars_lbls = .study$summary_demo_lbl,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DMT01"),
  .study = list(armvar = "ARM", summary_demo = c("AAGE", "AGEGR1", "SEX", "ETHNIC",
    "RACE", "BWGHTSI"), summary_demo_lbl = c("Age (yr)", "Pooled Age Group 1 (yr)",
    "SEX", "ETHNIC", "RACE", "Baseline Weight (kg)"), lbl_overall = "All Patients")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of strings}) variables summarized in demographic table.}

\item{summaryvars_lbls}{(\verb{vector of strings}) labels corresponding to the analyzed variables.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
For each variable, summary statistics are by default based on the number of patients in the corresponding \code{n} row.
}
\details{
\itemize{
\item Information from ADSUB are generally included into ADSL before analysis.
\item Default demographic and characteristics table
\item If not specified otherwise, numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows
\item Split columns by arm (planned or actual / code or description)
\item Include a total column by default
}
}
\section{Functions}{
\itemize{
\item \code{dmt01_1_lyt}: 
}}

\examples{

library(dm)

db <- syn_test_data() \%>\%
   preprocess_data("dmt01_1")

dmt01_1(db, summaryvars = c("AGE", "RACE", "SEX", "BWGHTSI"))
dmt01_1(db, summaryvars = c("AGE", "RACE", "SEX"), lbl_overall = NULL)
dmt01_1(db,
        summaryvars = c("AGE", "RACE", "SEX"),
        summaryvars_lbls = c("Age (yr)", "Race", "Sex"))

dmt01_1_lyt(armvar = "ACTARM")

}
