% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\name{dmt01_1}
\alias{dmt01_1}
\title{Demographics and Baseline Characteristics Table}
\usage{
dmt01_1(
  ad_bl,
  armvar = .study$armvar,
  summaryvars = .study$summaryvars,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DMT01"),
  .study = list(armvar = "ARM", summaryvars = c("AAGE", "AGEGR1", "SEX", "ETHNIC",
    "RACE", "BWGHTSI"), lbl_overall = "All Patients")
)
}
\arguments{
\item{ad_bl}{subject baseline dataset with one row per patient}

\item{armvar}{variable used for column splitting}

\item{summaryvars}{variables summarized in demographic table}

\item{lbl_overall}{label used for overall column}

\item{prune_0}{remove 0 count rows}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
For each variable, summary statistics are by default based on the number of patients in the corresponding \code{n} row.
}
\examples{
library(scda)
adsl <- synthetic_cdisc_data("rcd_2021_03_22")$adsl

dmt01_1(ad_bl = adsl,
 summaryvars = c("AGE", "RACE", "SEX"),
)

dmt01_1(
 ad_bl = adsl,
 armvar = "ARMCD",
 summaryvars = c("AGE", "RACE", "SEX")
)

}
