% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt01.R
\name{dmt01_lyt}
\alias{dmt01_lyt}
\title{dmt01 Layout}
\usage{
dmt01_lyt(arm_var, summaryvars, summaryvars_lbls, lbl_overall)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars_lbls}{(\code{character}) labels corresponding to the analyzed variables.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
dmt01 Layout
}
\keyword{internal}
