% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl01.R
\docType{data}
\name{dsl01_main}
\alias{dsl01_main}
\alias{dsl01_pre}
\alias{dsl01_post}
\alias{dsl01}
\title{\code{DSL01} Listing 1 (Default) Patients with Study Drug Withdrawn Due to Adverse Events.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
dsl01_main(
  adam_db,
  dataset = "adsl",
  disp_cols = names(adam_db[[dataset]]),
  default_formatting = list(all = fmt_config(align = "left"), numeric = fmt_config(align
    = "center")),
  col_formatting = NULL,
  unique_rows = TRUE,
  ...
)

dsl01_pre(adam_db, dataset = "adsl", ...)

dsl01_post(tlg, ...)

dsl01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{default_formatting}{(\code{list}) the default format of the listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{col_formatting}{(\code{list}) the format of specific listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{unique_rows}{(\code{flag}) whether to keep only unique rows in listing.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
\code{DSL01} Listing 1 (Default) Patients with Study Drug Withdrawn Due to Adverse Events.
}
\section{Functions}{
\itemize{
\item \code{dsl01_main()}: Main TLG function

\item \code{dsl01_pre()}: Preprocessing

\item \code{dsl01_post()}: Postprocessing

}}
\examples{
res <- run(dsl01, syn_data)
}
\keyword{datasets}
