% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl02.R
\docType{data}
\name{dsl02_main}
\alias{dsl02_main}
\alias{dsl02_pre}
\alias{dsl02}
\title{\code{dsl02} Listing 1 (Default) Patients Who Discontinued Early from Study.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
dsl02_main(
  adam_db,
  dataset = "adsl",
  arm_var = "ACTARM",
  disp_cols = c("ID", "ASR", arm_var, "TRTSDTM", "EOSDY", "SSAEDY", "EOSRDY", "DCSREAS"),
  ...
)

dsl02_pre(adam_db, dataset = "adsl", arm_var = "ACTARM", ...)

dsl02
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{...}{not used.}

\item{default_formatting}{(\code{list}) the default format of the listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{col_formatting}{(\code{list}) the format of specific listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}
}
\description{
\code{dsl02} Listing 1 (Default) Patients Who Discontinued Early from Study.
}
\section{Functions}{
\itemize{
\item \code{dsl02_main()}: Main TLG function

\item \code{dsl02_pre()}: Preprocessing

}}
\examples{
res <- run(dsl02, syn_data)
}
\keyword{datasets}
