% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_1}
\alias{dst01_1}
\title{Adverse Events by System Organ Class and Preferred Term Table}
\usage{
dst01_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(armvar = "ACTARM", lbl_overall = "All patients")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
Adverse Events by System Organ Class and Preferred Term Table
}
\details{
\itemize{
\item dst01_1 == default
\item is the dataset always adsl? => YES
\item are the summary vars ever different from Completed Study and Discontinued Study (e.g include Ongoing?) => should be an option
\item Always split by ARM ? => should also be possible to be TRT01P/TREATsecp: decision is made outside of the template (above the template)
\item Always a total column? => YES as a standard (should be optional)
\item numeric variables are always count & fraction => YES
\item are fractions always x/N? => N
\item is the grouping into safety/non-safety fixed ? => safety is fixed (rest in non-safety or missing)
\item should 0 counts rows be removed? => YES
\item Should factor level be sorted ? => abc
\item What is the alternative count ? => ?
\item 
\item A disposition table summarizing the reasons for patient withdrawal
\item Split column by ARM
\item Include a total column
\item sort withdrawal reasons by alphabetic order
}
}
\examples{

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

dst01_1(adsl, adae)
dst01_1(adsl, adae, lbl_overall = "")

dst01_2(adsl,adae)

adae <- adae \%>\% var_relabel(AEBODSYS = "Medra System Organ Class")
dst01_1(adsl, adae)

}
