% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_1}
\alias{dst01_1}
\title{Patient Disposition Table}
\usage{
dst01_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{(\code{string}) variable. Usually one of \code{ARM}, \code{ACTARM}, \code{TRT01A}, or \code{TRT01A}.}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{status}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a variable
name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis period.}

\item{reason}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS}, however can
also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
The DST01 Disposition Table provides an overview of patients study completion. For patients who discontinued the
study a reason is provided.
}
\details{
\itemize{
\item Default patient disposition table summarizing the reasons for patients withdrawal
\item Numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows
\item Split columns by arm
\item Include a total column by default
\item Sort withdrawal reasons by alphabetic order
}
}
\examples{

library(scda)
library(dplyr)

sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

dst01_1(adsl, adae)
dst01_1(adsl, adae, lbl_overall = "")

}
