% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\docType{data}
\name{dst01_1_main}
\alias{dst01_1_main}
\alias{dst01_1_lyt}
\alias{dst01_1_pre}
\alias{dst01_1}
\title{\code{DST01_1} Pipeline}
\format{
a \code{tlg_pipeline_S4} object with the following slots:
\itemize{
\item \code{main} the \code{chevron::dst01_1_main} function.
\item \code{preprocess} the  \code{chevron::dst01_1_pre} function.
\item \code{postprocess} the identity function.
\item \code{check} no checks.
\item \code{adam_datasets} \code{"adsl"}.
}
}
\usage{
dst01_1_main(
  adam_db,
  armvar = .study$planarm,
  status_var = .study$status_var,
  disc_reason_var = .study$disc_reason_var,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", disc_reason_var =
    "DCSREAS", status_var = "EOSSTT")
)

dst01_1_lyt(
  armvar = .study$planarm,
  status = .study$status,
  disc_reason_var = .study$disc_reason_var,
  completed_lbl = "COMPLETED",
  ongoing_lbl = "ONGOING",
  discontinued_lbl = "DISCONTINUED",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", status = "EOSSTT",
    disc_reason_var = "DCSREAS")
)

dst01_1_pre(adam_db, ...)

dst01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable. Usually one of \code{ARM}, \code{ACTARM}, \code{TRT01A}, or \code{TRT01A}.}

\item{status_var}{(\code{character}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a
variable name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{disc_reason_var}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS},
however can also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{status}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a variable
name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis period.}

\item{completed_lbl}{(\code{string}) associated with completed study and found in the columns given by \code{status}. By
Default \code{COMPLETED}.}

\item{ongoing_lbl}{(\code{string}) associated with ongoing treatment and found in the columns given by \code{status_treatment}.
By Default `ONGOING.}

\item{discontinued_lbl}{(\code{string}) associated with discontinued study and found in the columns given by \code{status}. By
Default \code{DISCONTINUED}.}

\item{...}{not used.}
}
\description{
\code{DST01_1} Pipeline of the class \code{tlg_pipeline_S4}
}
\details{
\itemize{
\item Default patient disposition table summarizing the reasons for patients withdrawal.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows.
\item Split columns by arm.
\item Include a total column by default.
\item Sort withdrawal reasons by alphabetic order.
}

Since the two parts of the tables are pruned differently, the layout function returns a list of layouts,
which allows the tables to be constructed and pruned separately before binding.
}
\section{Functions}{
\itemize{
\item \code{dst01_1_main}: \code{dst01} main function

DST01 Table 1 (Default) Patient Disposition Table 1. The DST01 Disposition Table provides an overview of patients
study completion. For patients who discontinued the study a reason is provided.

\item \code{dst01_1_lyt}: \code{dst01_1} Layout

\item \code{dst01_1_pre}: \code{dst01_1} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  dst01_1_pre()

dst01_1_main(db)
dst01_1_lyt(
  armvar = "ACTARM",
  status = "EOP01STT",
  disc_reason_var = "DCP01RS"
)
syn_test_data() \%>\%
  dst01_1_pre()
}
\keyword{datasets}
