% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\docType{data}
\name{dst01_1_main}
\alias{dst01_1_main}
\alias{dst01_1_lyt}
\alias{dst01_1_pre}
\alias{dst01_1_post}
\alias{dst01_1}
\title{DST01 Table 1 (Default) Patient Disposition Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
dst01_1_main(
  adam_db,
  arm_var = "ARM",
  status_var = "EOSSTT",
  disc_reason_var = "DCSREAS",
  lbl_overall = "All Patients",
  deco = std_deco("DST01"),
  ...
)

dst01_1_lyt(
  arm_var,
  status_var,
  disc_reason_var,
  completed_lbl = "COMPLETED",
  ongoing_lbl = "ONGOING",
  discontinued_lbl = "DISCONTINUED",
  lbl_overall,
  deco
)

dst01_1_pre(adam_db, status_var = "EOSSTT", disc_reason_var = "DCSREAS", ...)

dst01_1_post(tlg, prune_0 = TRUE, deco = std_deco("DST01"), ...)

dst01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{character}) variable. Usually one of \code{ARM}, \code{ACTARM}, \code{TRT01A}, or \code{TRT01A}.}

\item{status_var}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a
variable name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{disc_reason_var}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS},
however can also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{completed_lbl}{(\code{string}) associated with completed study and found in the columns given by \code{status}. By
Default \code{COMPLETED}.}

\item{ongoing_lbl}{(\code{string}) associated with ongoing treatment and found in the columns given by \code{status_treatment}.
By Default `ONGOING.}

\item{discontinued_lbl}{(\code{string}) associated with discontinued study and found in the columns given by \code{status}. By
Default \code{DISCONTINUED}.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
The DST01 Disposition Table provides an overview of patients
study completion. For patients who discontinued the study a reason is provided.
}
\details{
\itemize{
\item Default patient disposition table summarizing the reasons for patients withdrawal.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows.
\item Split columns by arm.
\item Include a total column by default.
\item Sort withdrawal reasons by alphabetic order.
}

Since the two parts of the tables are pruned differently, the layout function returns a list of layouts,
which allows the tables to be constructed and pruned separately before binding.
}
\section{Functions}{
\itemize{
\item \code{dst01_1_main()}: Main TLG function

\item \code{dst01_1_lyt()}: Layout

\item \code{dst01_1_pre()}: Preprocessing

\item \code{dst01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adsl} table with the columns specified by \code{status_var} and \code{disc_reason_var}.
}
}
\examples{
library(dm)
library(magrittr)

db <- syn_data \%>\%
  dst01_1_pre()

dst01_1_main(db)
dst01_1_pre(syn_data)
run(dst01_1, syn_data)
}
\keyword{datasets}
