% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_2}
\alias{dst01_2}
\title{Patient Disposition Table 2}
\usage{
dst01_2(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  status = .study$status,
  reason = .study$reason,
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{status}{variable used to define patient status}

\item{reason}{variable used to define reason for patient withdrawal}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
Patient Disposition Table 2
}
\details{
\itemize{
\item Non-standard disposition table summarizing the reasons for patient withdrawal
\item Withdrawal reasons are grouped into Safety and Non-Safety issues
\item Safety issues include Death and Adverse event
\item Numbers represent absolute numbers of patients and fraction of N
\item Remove zero-count rows
\item Split columns by ACTARM
\item Include a total column by default
\item Sort withdrawal reasons by alphabetic order
}
}
\examples{

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae \%>\%
 mutate(ANL01FL = 'Y')

dst01_2(adsl, adae)
dst01_2(adsl, adae, lbl_overall = "")

}
