% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_3}
\alias{dst01_3}
\alias{dst01_3_lyt}
\title{DST01 Table 3 (Supplementary) Patient Disposition Table 3}
\usage{
dst01_3(
  adsl,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  status_treatment = .study$status_treatment,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS", status_treatment = "EOTSTT")
)

dst01_3_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status_treatment = .study$status,
  deco = std_deco("DST01"),
  completed_lbl = "COMPLETED",
  discontinued_lbl = "DISCONTINUED",
  ongoing_lbl = "ONGOING",
  .study = list(armvar = "ARM", lbl_overall = "All patients", status_treatment =
    "EOTSTT")
)
}
\arguments{
\item{adsl}{(\code{data.frame}) Subject Level Analysis Dataset}

\item{armvar}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{status}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a variable
name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis period.}

\item{reason}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS}, however can
also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{status_treatment}{(\code{string}) variable used to define the treatment status of the patients. Default is \code{EOTSTT},
however can also be a variable with the pattern \code{EOTxxSTT} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{deco}{(\code{list}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{completed_lbl}{(\code{string}) associated with completed treatment and found in the columns given by
\code{status_treatment}. By Default \code{COMPLETED}.}

\item{discontinued_lbl}{(\code{string}) associated with discontinued treatment and found in the columns given by
\code{status_treatment}. By Default \code{DISCONTINUED}.}

\item{ongoing_lbl}{(\code{string}) associated with ongoing treatment and found in the columns given by \code{status_treatment}.
By Default `ONGOING.}
}
\value{

}
\description{
The DST01_3 Disposition Table provides an overview of patients study treatment status.
For patients who discontinued the study, the reason provided is categorized as "Safety" or "Non-Safety" issue.
}
\details{
\itemize{
\item Non-standard disposition table summarizing the reasons for patient withdrawal and treatment status.
\item Withdrawal reasons are grouped into Safety and Non-Safety issues.
\item Safety issues include Death and Adverse Events.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows.
\item Split columns by arm.
\item Include a total column by default.
\item Sort withdrawal reasons by alphabetic order.
}
}
\section{Functions}{
\itemize{
\item \code{dst01_3_lyt}: 
}}

\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")

adsl <- sd$adsl \%>\%
 mutate(ANL01FL = 'Y',
        EOTSTT = sample(c("ONGOING","COMPLETED","DISCONTINUED"),
                        nrow(sd$adsl),
                        replace = TRUE))

dst01_3(adsl)
dst01_3(adsl, lbl_overall = "")

dst01_3_lyt(armvar = "ACTARM",
status_treatment = "EOTxx01")
}
