% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\docType{data}
\name{dst01_3}
\alias{dst01_3}
\title{\code{DST01_3}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
dst01_3
}
\description{
\code{DST01_3}
}
\seealso{
\code{\link[=dst01_3_main]{dst01_3_main()}}
}
\keyword{datasets}
