% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\docType{data}
\name{dst01_3_main}
\alias{dst01_3_main}
\alias{dst01_3_lyt}
\alias{dst01_3_pre}
\alias{dst01_3}
\title{DST01 Table 3 (Supplementary) Patient Disposition Table 3.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
dst01_3_main(
  adam_db,
  armvar = .study$planarm,
  status = .study$status_var,
  disc_reason_var = .study$disc_reason_var,
  status_treatment = "EOTSTT",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", disc_reason_var = "DCSREAS", lbl_overall =
    "All Patients", status_var = "EOSSTT")
)

dst01_3_lyt(
  armvar = .study$planarm,
  status_treatment = .study$status,
  completed_lbl = "COMPLETED",
  discontinued_lbl = "DISCONTINUED",
  ongoing_lbl = "ONGOING",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", status_treatment =
    "EOTSTT")
)

dst01_3_pre(
  adam_db,
  reason = .study$disc_reason_var,
  .study = list(disc_reason_var = "DCSREAS"),
  ...
)

dst01_3
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{status}{(\code{character}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a
variable name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{disc_reason_var}{(\code{character}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS},
however can also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{status_treatment}{(\code{string}) variable used to define the treatment status of the patients. Default is \code{EOTSTT},
however can also be a variable with the pattern \code{EOTxxSTT} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{completed_lbl}{(\code{string}) associated with completed treatment and found in the columns given by
\code{status_treatment}. By Default \code{COMPLETED}.}

\item{discontinued_lbl}{(\code{string}) associated with discontinued treatment and found in the columns given by
\code{status_treatment}. By Default \code{DISCONTINUED}.}

\item{ongoing_lbl}{(\code{string}) associated with ongoing treatment and found in the columns given by \code{status_treatment}.
By Default `ONGOING.}

\item{reason}{(\code{character}) the variable name for variable with the reason for discontinuation.}

\item{...}{not used.}
}
\description{
The DST01_3 Disposition Table provides an overview of patients study treatment status. For patients who discontinued
the study, the reason provided is categorized as "Safety" or "Non-Safety" issue.
}
\details{
\itemize{
\item Non-standard disposition table summarizing the reasons for patient withdrawal and treatment status.
\item Withdrawal reasons are grouped into Safety and Non-Safety issues.
\item Safety issues include Death and Adverse Events.
\item Numbers represent absolute numbers of patients and fraction of \code{N}.
\item Remove zero-count rows.
\item Split columns by arm.
\item Include a total column by default.
\item Sort withdrawal reasons by alphabetic order.
}
}
\section{Functions}{
\itemize{
\item \code{dst01_3_main}: Main TLG function

\item \code{dst01_3_lyt}: Layout

\item \code{dst01_3_pre}: Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  dst01_3_pre()

dst01_3_main(db)
dst01_3_main(db, lbl_overall = NULL)
dst01_3_lyt(
  armvar = "ACTARM",
  status_treatment = "EOTxx01"
)
syn_test_data() \%>\%
  dst01_3_pre()
}
\keyword{datasets}
