% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst01.R
\name{dst01_3}
\alias{dst01_3}
\title{Patient Disposition Table 3}
\usage{
dst01_3(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  status_treatment = .study$status_treatment,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS", status_treatment = "EOTSTT")
)
}
\arguments{
\item{adsl}{Subject Level Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{status}{(\code{string}) variable used to define patient status. Default is \code{EOSSTT}, however can also be a variable
name with the pattern \code{EOPxxSTT} where \code{xx} must be substituted by 2 digits referring to the analysis period.}

\item{reason}{(\code{string}) variable used to define reason for patient withdrawal. Default is \code{DCSREAS}, however can
also be a variable with the pattern \code{DCPxxRS} where \code{xx} must be substituted by 2 digits referring to the analysis
period.}

\item{status_treatment}{(\code{string}) variable used to define the treatment status of the patients. Default is \code{EOTSTT},
however can also be a variable with the pattern \code{EOTxxSTT} where \code{xx} must be substituted by 2 digits referring to
the analysis period.}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
The DST01_3 Disposition Table provides an overview of patients study completion and of the treatment completion.
For patients who discontinued the study, the reason provided is categorized as "Safety" or "Non-Safety" issue.
}
\details{
\itemize{
\item Non-standard disposition table summarizing the reasons for patient withdrawal and treatment status
\item Withdrawal reasons are grouped into Safety and Non-Safety issues
\item Safety issues include Death and Adverse Events
\item Numbers represent absolute numbers of patients and fraction of \code{N}
\item Remove zero-count rows
\item Split columns by arm
\item Include a total column by default
\item Sort withdrawal reasons by alphabetic order
}
}
\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl

adae <- sd$adae |>
 mutate(ANL01FL = 'Y',
        EOTSTT = sample(c("ONGOING","COMPLETED","DISCONTINUED"),
                        nrow(sd$adae),
                        replace = TRUE))

dst01_3(adsl, adae)
dst01_3(adsl, adae, lbl_overall = "")

}
