% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtht01.R
\docType{data}
\name{dtht01_1_main}
\alias{dtht01_1_main}
\alias{dtht01_1_lyt}
\alias{dtht01_1_opt_lyt}
\alias{dtht01_1_pre}
\alias{dtht01_1}
\title{\code{DTHT01} Table 1 (Default) Death Table.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
dtht01_1_main(
  adam_db,
  lyt_ls = list(causes = dtht01_1_lyt, time_since_last_dose = dtht01_1_opt_lyt),
  armvar = .study$actualarm,
  time_since_last_dose = FALSE,
  other_category = FALSE,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DTHT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

dtht01_1_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  other_category = TRUE,
  deco = std_deco("DTHT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

dtht01_1_opt_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DTHT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

dtht01_1_pre(adam_db, ...)

dtht01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_ls}{(\verb{list of functions} returning \code{PreDataTableLayout} object. Typically one of the \verb{_lyt} function of
\code{chevron} stored in a \code{list}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{time_since_last_dose}{(\code{logical}) should the time to event information be displayed.}

\item{other_category}{(\code{logical}) should the breakdown of the \code{OTHER} category be displayed.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}
}
\description{
A description of the causes of death optionally with the breakdown of the
\code{OTHER} category and/or post-study reporting of death.
}
\details{
\itemize{
\item Numbers represent absolute numbers of subjects and fraction of \code{N}, or absolute numbers when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Does not include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{dtht01_1_main()}: Main TLG function

\item \code{dtht01_1_lyt()}: Layout

\item \code{dtht01_1_opt_lyt()}: Optional Layout

\item \code{dtht01_1_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adsl} table with the columns "DTHFL", "DTHCAT" as well as \code{LDDTHGR1} if
\code{time_since_last_dose} is \code{TRUE}.
\item \code{lyt_ls} must contain a "causes" and, if \code{time_since_last_dose} is \code{TRUE}, a "time_since_last_dose" element.
}
}
\examples{

db <- syn_test_data()

run(dtht01_1, db)
run(dtht01_1, db, other_category = TRUE, time_since_last_dose = TRUE)
}
\keyword{datasets}
