% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_template.R
\docType{data}
\name{dummy_template}
\alias{dummy_template}
\title{Dummy template.}
\format{
An object of class \code{chevron_simple} of length 1.
}
\usage{
dummy_template
}
\description{
This template creates a dummy output.
}
\examples{
run(dummy_template, syn_data)
}
\keyword{datasets}
