% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt01.R
\name{egt01_1}
\alias{egt01_1}
\alias{egt01_1_lyt}
\title{\code{EGT01} Table 1 (Default) ECG Assessments and Change from Baseline by Visit Table 1}
\usage{
egt01_1(
  adam_db,
  armvar = .study$armvar,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = var_labels_for(adam_db$adeg, summaryvars),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("EGT01"),
  .study = list(armvar = "ACTARM", evo_vars = c("AVAL", "CHG"))
)

egt01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = .study$evo_vars_lbls,
  visitvar = .study$visitvar,
  lbl_avisit = "",
  lbl_param = "",
  deco = std_deco("EGT01"),
  .study = list(armvar = "ACTARM", evo_vars = c("AVAL", "CHG"), evo_vars_lbls =
    c("Analysis \\nValue", "Change from \\nBaseline"), visitvar = "AVISIT")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_avisit}{(\code{character}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}
}
\value{

}
\description{
The \code{EGT01} table 1 summarizes several electrocardiogram parameters and their evolution throughout the study.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{egt01_1_lyt}: 
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  preprocess_data("egt01_1")

egt01_1(db)
egt01_1(db, summaryvars_lbls = c("Value at Visit", "Change from Baseline"))

}
