% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt01.R
\docType{data}
\name{egt01_1_main}
\alias{egt01_1_main}
\alias{egt01_1_lyt}
\alias{egt01_1_pre}
\alias{egt01_1}
\title{\code{EGT01} Table 1 (Default) ECG Assessments and Change from Baseline by Visit Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt01_1_main(
  adam_db,
  lyt_ls = list(egt01_1_lyt),
  armvar = "ACTARM",
  summaryvars = c("AVAL", "CHG"),
  summaryvars_lbls = c("Value at Visit", "Change from \\nBaseline"),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("EGT01"),
  ...
)

egt01_1_lyt(
  armvar,
  summaryvars,
  summaryvars_lbls,
  visitvar,
  lbl_avisit,
  lbl_param,
  deco,
  ...
)

egt01_1_pre(adam_db, ...)

egt01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_ls}{(\verb{list of functions} returning \code{PreDataTableLayout} object. Typically one of the \verb{_lyt} function of
\code{chevron} stored in a \code{list}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{lbl_avisit}{(\code{character}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}
}
\description{
The \code{EGT01} table 1 summarizes
several electrocardiogram parameters and their evolution throughout the study.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order.
}
}
\section{Functions}{
\itemize{
\item \code{egt01_1_main()}: Main TLG function

\item \code{egt01_1_lyt()}: Layout

\item \code{egt01_1_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adeg} table with a \code{"PARAM"} column as well as columns specified in
\code{summaryvars} and \code{visitvar}.
}
}
\examples{
db <- syn_test_data()
run(egt01_1, db)
run(egt01_1, db, summaryvars_lbls = c("Value at Visit", "Change from Baseline"))
}
\keyword{datasets}
