% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt02.R
\docType{data}
\name{egt02_1}
\alias{egt02_1}
\title{\code{EGT02} ECG Abnormalities Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt02_1
}
\description{
ECG Parameters outside Normal Limits Regardless of Abnormality at Baseline Table.
}
\examples{
run(egt02_1, syn_data)
}
\keyword{datasets}
