% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt02.R
\docType{data}
\name{egt02_1}
\alias{egt02_1}
\title{\code{EGT02_1}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
egt02_1
}
\description{
\code{EGT02_1}
}
\seealso{
\code{\link[=egt02_1_main]{egt02_1_main()}}
}
\keyword{datasets}
