% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt02.R
\docType{data}
\name{egt02_2}
\alias{egt02_2}
\title{\code{EGT02_2}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
egt02_2
}
\description{
\code{EGT02_2}
}
\seealso{
\code{\link[=egt02_2_main]{egt02_2_main()}}
}
\keyword{datasets}
