% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt02.R
\docType{data}
\name{egt02_2_main}
\alias{egt02_2_main}
\alias{egt02_2_lyt}
\alias{egt02_2_pre}
\alias{egt02_2_post}
\alias{egt02_2}
\title{\code{EGT02} Table 2 (Supplementary) ECG Abnormalities Table 2.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt02_2_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = NULL,
  deco = std_deco("EGT02_2"),
  ...
)

egt02_2_lyt(
  arm_var = "ACTARM",
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = NULL,
  deco = std_deco("EGT02_2")
)

egt02_2_pre(adam_db, ...)

egt02_2_post(tlg, ...)

egt02_2
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{string}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{string}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
Assessments Outside Normal Limits Among Subject Without
Abnormality at Baseline.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{egt02_2_main()}: Main TLG function

\item \code{egt02_2_lyt()}: Layout

\item \code{egt02_2_pre()}: Preprocessing

\item \code{egt02_2_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adeg} table with the \code{"PARAM"}, \code{"ANRIND"} and \code{"BNRIND"} columns.
}
}
\examples{
run(egt02_2, syn_data)
}
\keyword{datasets}
