% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt02.R
\docType{data}
\name{egt02_2_main}
\alias{egt02_2_main}
\alias{egt02_2}
\title{\code{EGT02_2} ECG Abnormalities Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt02_2_main(adam_db, arm_var = "ACTARM", lbl_overall = NULL, ...)

egt02_2
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}
}
\description{
ECG Parameters outside Normal Limits Among Patients without Abnormality at Baseline Table.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{egt02_2_main()}: Main TLG function

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adeg} table with the \code{"PARAM"}, \code{"ANRIND"} and \code{"BNRIND"} columns.
}
}
\examples{
run(egt02_2, syn_data)
}
\keyword{datasets}
