% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt03.R
\docType{data}
\name{egt03_2_main}
\alias{egt03_2_main}
\alias{egt03_2_lyt}
\alias{egt03_2_check}
\alias{egt03_2_pre}
\alias{egt03_2_post}
\alias{egt03_2}
\title{\code{EGT03} Table 2 Shift Table of ECG Interval Data - Baseline versus Maximum Post-Baseline}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt03_2_main(
  adam_db,
  armvar = "ARMCD",
  summaryvar = c(`Baseline Reference Range Indicator` = "BNRIND"),
  splitvar = c(`Analysis Reference Range Indicator` = "ANRIND"),
  deco = std_deco("EGT03"),
  ...
)

egt03_2_lyt(
  armvar,
  splitvar,
  summaryvar,
  lbl_armvar,
  lbl_splitvar,
  lbl_summaryvars,
  deco,
  ...
)

egt03_2_check(adam_db, req_tables = c("adeg"), visitvar = "AVISIT", ...)

egt03_2_pre(adam_db, ...)

egt03_2_post(tlg, prune_0 = FALSE, ...)

egt03_2
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\verb{vector of character}) the arm variables used for row split, typically \code{"ARMCD"}.}

\item{summaryvar}{(named vector of \code{character}) variables to be analyzed, typically \code{"BNRIND"}. Names are used as
subtitles. For values
where no name is provided, the label attribute of the corresponding column in \code{adeg} table of \code{adam_db} is used.}

\item{splitvar}{(named vector of \code{character}) variables to be analyzed, typically \code{"BNRIND"}. Names are used as
subtitles. For values
where no name is provided, the label attribute of the corresponding column in \code{adeg} table of \code{adam_db} is used.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{lbl_armvar}{(\code{character}) label of the \code{armvar} variable.}

\item{lbl_splitvar}{(\code{character}) label of the \code{splitvar} variable.}

\item{lbl_summaryvars}{(\code{character}) label of the \code{summaryvar} variable.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
The \code{EGT03} Table entries provide the number of patients by baseline assessment and maximum post-baseline
assessment. Percentages are based on the total number of patients in a treatment group. Baseline is the
patient's last observation prior to initiation of study drug.
}
\details{
\itemize{
\item ADEG data are subsetted to contain only "POST-BASELINE MAXIMUM" visit
\item the number of patients by baseline assessment and maximum post-baseline assessment.
\item Percentages are based on the total number of patients in a treatment group.
\item Split columns by Analysis Reference Range Indicator, typically \code{ANRIND}.
\item Does not include a total column by default.
\item Sorted  based on factor level.
}
}
\section{Functions}{
\itemize{
\item \code{egt03_2_main()}: Main TLG function

\item \code{egt03_2_lyt()}: Layout

\item \code{egt03_2_check()}: Checks

\item \code{egt03_2_pre()}: Preprocessing

\item \code{egt03_2_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adeg} table with a \code{"ARMCD"} column as well as columns specified in
\code{summaryvar} and \code{splitvar}.
}
}
\examples{
db <- syn_data
run(egt03_2, db)
}
\keyword{datasets}
