% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt05_qtcat.R
\docType{data}
\name{egt05_qtcat_1_main}
\alias{egt05_qtcat_1_main}
\alias{egt05_qtcat_1_lyt}
\alias{egt05_qtcat_1_pre}
\alias{egt05_qtcat_1_post}
\alias{egt05_qtcat_1}
\title{\code{EGT05_QTCAT} Table 1 (Default) ECG Actual Values and Changes from Baseline by Visit Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
egt05_qtcat_1_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = list(`Value at Visit` = "AVALCAT1", `Change from Baseline` = "CHGCAT1"),
  summaryvars_lbls = get_labels(adam_db$adeg, summaryvars),
  lbl_overall = NULL,
  visitvar = "AVISIT",
  lbl_avisit = var_labels_for(adam_db$adeg, visitvar),
  paramvar = "QT",
  lbl_param = var_labels_for(adam_db$adeg, "PARAM"),
  deco = std_deco("EGT05_QTCAT"),
  lbl_cat = "Category",
  lbl_headvisit = "Analysis Visit",
  ...
)

egt05_qtcat_1_lyt(
  arm_var,
  summaryvars,
  summaryvars_lbls,
  lbl_overall,
  visitvar,
  lbl_avisit,
  lbl_param,
  deco,
  lbl_cat,
  lbl_headvisit
)

egt05_qtcat_1_pre(adam_db, ...)

egt05_qtcat_1_post(tlg, prune_0 = TRUE, ...)

egt05_qtcat_1
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\code{character}) the variables to be analyzed. \code{AVALCAT1} and \code{CHGCAT1} by default.}

\item{summaryvars_lbls}{(\code{character}) the label of the variables to be analyzed.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{lbl_avisit}{(\code{string}) label of the \code{visitvar} variable.}

\item{paramvar}{(\code{string}) typically \code{"QT"} (Default). It should come from \code{"PARAMCD"}.}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{lbl_cat}{(\code{string}) label of the Category of \code{summaryvars} variable. Default as \code{Category}.}

\item{lbl_headvisit}{(\code{string}) label of Visits in the header. Default as \verb{Analysis Visit}.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{EGT05_QTCAT} table 1 summarizes several electrocardiogram parameters and their evolution
throughout the study.
}
\details{
\itemize{
\item The \verb{Value at Visit} column, displays the categories of the specific \code{"PARAMCD"} value for patients.
\item The \verb{Change from Baseline} column, displays the categories of the specific \code{"PARAMCD"} value
change from baseline for patients.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{"ACTARM"}.
\item Does not include a total column by default.
\item Sorted based on factor level; by chronological time point given by one of \code{"AVISIT"} (Default) or \code{"ATPTN"}.
Re-level to customize order.
}
}
\section{Functions}{
\itemize{
\item \code{egt05_qtcat_1_main()}: Main TLG function

\item \code{egt05_qtcat_1_lyt()}: Layout

\item \code{egt05_qtcat_1_pre()}: Preprocessing

\item \code{egt05_qtcat_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adeg} table with column specified in \code{visitvar}.
For \code{summaryvars}, please make sure \code{AVALCAT1} and \code{CHGCAT1} columns existed in input data sets.
}
}
\examples{
db <- syn_data
run(egt05_qtcat_1, db)
run(egt05_qtcat_1, db, summaryvars = list("AVALCAT1", "Change" = "CHGCAT1"))
}
\keyword{datasets}
