% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt05_qtcat.R
\name{egt05_qtcat_lyt}
\alias{egt05_qtcat_lyt}
\title{\code{EGT05_QTCAT} Layout}
\usage{
egt05_qtcat_lyt(
  arm_var,
  summaryvars,
  summaryvars_lbls,
  lbl_overall,
  visitvar,
  lbl_avisit,
  lbl_param,
  lbl_cat,
  lbl_headvisit
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\code{character}) the variables to be analyzed. \code{AVALCAT1} and \code{CHGCAT1} by default.}

\item{summaryvars_lbls}{(\code{character}) the label of the variables to be analyzed.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{lbl_avisit}{(\code{string}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{lbl_cat}{(\code{string}) label of the Category of \code{summaryvars} variable. Default as \code{Category}.}

\item{lbl_headvisit}{(\code{string}) label of Visits in the header. Default as \verb{Analysis Visit}.}
}
\description{
\code{EGT05_QTCAT} Layout
}
\keyword{internal}
