% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\docType{data}
\name{ext01_main}
\alias{ext01_main}
\alias{ext01_lyt}
\alias{ext01_pre}
\alias{ext01_post}
\alias{ext01}
\title{EXT01 Exposure Summary Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
ext01_main(
  adam_db,
  keep_lvl = "ALL",
  arm_var = "ACTARM",
  summaryvars = "AVAL",
  lbl_overall = NULL,
  ...
)

ext01_lyt(arm_var, summaryvars, summaryvars_lbls, lbl_overall)

ext01_pre(
  adam_db,
  paramcd_order = c("TNDOSE", "DOSE", "NDOSE", "TDOSE"),
  show_stats = "ALL",
  show_bins = "ALL",
  ...
)

ext01_post(tlg, prune_0 = TRUE, ...)

ext01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{\code{(string)} the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}

\item{summaryvars_lbls}{\code{(string)} the label associated with the analyzed variable.}

\item{paramcd_order}{(\code{character}) providing the \code{PARAMCD} values in the desired order.}

\item{show_stats}{(\code{character}) providing the name of the parameters whose statistical summary should be
presented. To analyze all, provide \code{show_stats = "ALL"} (Default), to analyze none, provide \code{show_stats = ""}.}

\item{show_bins}{(\code{character}) providing the name of the parameters whose categorical summary should be
presented. To analyze all, provide \code{show_bins = "ALL"} (Default), to analyze none, provide \code{show_bins = ""}.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The EXT01 table provides an overview of the of the exposure of the
patients in terms of Total dose administered or missed, and treatment duration.
}
\details{
\itemize{
\item Default Exposure table
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
\item \code{ANL01FL} is not relevant subset.
}
}
\section{Functions}{
\itemize{
\item \code{ext01_main()}: Main TLG function

\item \code{ext01_lyt()}: Layout

\item \code{ext01_pre()}: Preprocessing

\item \code{ext01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adex} table with columns specified in \code{summaryvars}.
}
}
\examples{
run(ext01, syn_data)
run(ext01, syn_data, summaryvars = c("AVAL", "AVALCAT1"))
}
\keyword{datasets}
