% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\name{ext01_1}
\alias{ext01_1}
\alias{ext01_1_lyt}
\title{EXT01 Table 1 (Default) Exposure Summary Table}
\usage{
ext01_1(
  adex,
  adsl,
  armvar = .study$armvar,
  summaryvars = "AVAL",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)

ext01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$analysis_var,
  summaryvars_lbls = .study$lbl_analysis_var,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", overall_col = "", analysis_var = "AVAL",
    lbl_analysis_var = "Analysis Value")
)
}
\arguments{
\item{adex}{(\code{data.frame}) Exposure Analysis Dataset}

\item{adsl}{(\code{data.frame}) Subject Level Analysis Dataset}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{\code{(string)} the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{summaryvars_lbls}{\code{(string)} the label associated with the analyzed variable.}
}
\value{

}
\description{
The EXT01 table provides an overview of the of the exposure of the patients in terms of Total dose administered or
missed, and treatment duration.
}
\details{
\itemize{
\item Default Exposure table
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
}
}
\section{Functions}{
\itemize{
\item \code{ext01_1_lyt}: 
}}

\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adex <- sd$adex \%>\%
 mutate(ANL01FL = 'Y')

ext01_1(adex, adsl)

adex2 <- adex \%>\%
   relevel_params(paramcd_levels = c("TNDOSE", "DOSE", "NDOSE", "TDOSE"))

ext01_1(adex2, adsl)
ext01_1(adex2, adsl, lbl_overall = "All Patients")
}
