% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\docType{data}
\name{ext01_1_main}
\alias{ext01_1_main}
\alias{ext01_1_lyt}
\alias{ext01_1_pre}
\alias{ext01_1}
\title{\code{EXT01_1} Pipeline}
\format{
a \code{tlg_pipeline_S4} object with the following slots:
\itemize{
\item \code{main} the \code{chevron::ext01_1_main} function.
\item \code{preprocess} the  \code{chevron::ext01_1_pre} function.
\item \code{postprocess} the identity function.
\item \code{check} no checks.
\item \code{adam_datasets} \code{"adsl"} and \code{"adex"}.
}
}
\usage{
ext01_1_main(
  adam_db,
  armvar = .study$actualarm,
  summaryvars = "AVAL",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

ext01_1_lyt(
  armvar = .study$actualarm,
  summaryvars = .study$analysis_var,
  summaryvars_lbls = .study$lbl_analysis_var,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, analysis_var = "AVAL",
    lbl_analysis_var = "Analysis Value")
)

ext01_1_pre(
  adam_db,
  paramcd_order = .study$paramcd_order,
  .study = list(paramcd_order = c("TNDOSE", "DOSE", "NDOSE", "TDOSE")),
  ...
)

ext01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{\code{(string)} the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{summaryvars_lbls}{\code{(string)} the label associated with the analyzed variable.}

\item{paramcd_order}{(\code{character}) providing the \code{PARAMCD} values in the desired order.}

\item{...}{not used.}
}
\description{
\code{EXT01_1} Pipeline of the class \code{tlg_pipeline_S4}
}
\details{
\itemize{
\item Default Exposure table
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
\item \code{ANL01FL} is not relevant subset
}
}
\section{Functions}{
\itemize{
\item \code{ext01_1_main}: \code{ext01_1} main function

EXT01 Table 1 (Default) Exposure Summary Table. The EXT01 table provides an overview of the of the exposure of the
patients in terms of Total dose administered or missed, and treatment duration.

\item \code{ext01_1_lyt}: \code{ext01_1} Layout

\item \code{ext01_1_pre}: \code{ext01_1} Preprocessing
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  ext01_1_pre()

ext01_1_main(db)
syn_test_data() \%>\%
  ext01_1_pre()
}
\keyword{datasets}
