% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\name{ext01_1}
\alias{ext01_1}
\title{Exposure Summary Table}
\usage{
ext01_1(
  adex,
  armvar = .study$armvar,
  summaryvars = "AVAL",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)
}
\arguments{
\item{adex}{Exposure Analysis Dataset}

\item{armvar}{variable used for column splitting}

\item{summaryvars}{}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{remove 0 count rows}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}
}
\description{
Exposure Summary Table
}
\details{
\itemize{
\item Default Exposure table
\item Split columns by arm, typically \code{ACTARM}.
}
}
\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adex <- sd$adex |>
 mutate(ANL01FL = 'Y')

ext01_1(adsl, adex)

}
