% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\docType{data}
\name{ext01_2_main}
\alias{ext01_2_main}
\alias{ext01_2_lyt}
\alias{ext01_2_pre}
\alias{ext01_2}
\title{EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
ext01_2_main(
  adam_db,
  lyt_ls = list(ext01_2_lyt),
  armvar = "ACTARM",
  lbl_overall = NULL,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  ...
)

ext01_2_lyt(
  armvar,
  summaryvars,
  summaryvars_lbls,
  lbl_overall,
  deco = std_deco("EXT01"),
  ...
)

ext01_2_pre(adam_db, show_stats = c("ALL"), show_bins = c("ALL"), ...)

ext01_2
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_ls}{(\verb{list of functions} returning \code{PreDataTableLayout} object. Typically one of the \verb{_lyt} function of
\code{chevron} stored in a \code{list}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{summaryvars}{(\code{string}) the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{summaryvars_lbls}{(\code{string}) the label associated with the analyzed variable.}

\item{show_stats}{(\verb{vector of character}) providing the name of the parameters whose statistical summary should be
presented. To analyze all, provide \code{show_stats = "ALL"} (Default), to analyze none, provide \code{show_stats = ""}.}

\item{show_bins}{(\verb{vector of character}) providing the name of the parameters whose categorical summary should be
presented. To analyze all, provide \code{show_bins = "ALL"} (Default), to analyze none, provide \code{show_bins = ""}.}
}
\description{
EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options
}
\details{
\itemize{
\item Supplementary Exposure table with binning of desired analysis values.
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
\item \code{ANL01FL} is not relevant subset
}
}
\section{Functions}{
\itemize{
\item \code{ext01_2_main()}: Main TLG function

\item \code{ext01_2_lyt()}: Layout

\item \code{ext01_2_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adex} table with columns specified in \code{summaryvars}.
}
}
\examples{
run(ext01_2, syn_test_data())
}
\keyword{datasets}
