% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\name{ext01_2}
\alias{ext01_2}
\alias{ext01_2_lyt}
\title{EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options}
\usage{
ext01_2(
  adex,
  adsl,
  armvar = .study$armvar,
  show_stats = .study$show_stats,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", show_stats = c("ALL"), lbl_overall = "")
)

ext01_2_lyt(
  armvar = .study$armvar,
  summaryvars = c("AVAL", "AVALCAT1"),
  summaryvars_lbls = c("Summary", "Categories"),
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)
}
\arguments{
\item{adex}{(\code{data.frame}) Exposure Analysis Dataset}

\item{adsl}{(\code{data.frame}) Subject Level Analysis Dataset}

\item{armvar}{(\code{string}) variable used for column splitting}

\item{show_stats}{(\verb{vector of strings}) providing the name of the parameters whose statistical summary should be
presented. To analyze all, provide \code{paramvar = "ALL"} (Default), to analyze none, provide \code{paramvar = ""}.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{list}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{summaryvars}{(\code{string}) the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{summaryvars_lbls}{(\code{string}) the label associated with the analyzed variable.}
}
\value{

}
\description{
EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options

EXT01 Layout 2 (Supplementary)
}
\details{
\itemize{
\item Supplementary Exposure table with binning of desired analysis values.
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
}
}
\section{Functions}{
\itemize{
\item \code{ext01_2_lyt}: 
}}

\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adex <- sd$adex \%>\%
  mutate(ANL01FL = 'Y')

group <- list(list("Dose administered during constant dosing interval",
                                         c(-Inf, 700, 900, 1200, Inf),
                                         c("<700", "700-900", "900-1200", ">1200")
                                          ),
                                      list("Total dose administered",
                                         c(-Inf, 5000, 7000, 9000, Inf),
                                         c("<5000", "5000-7000", "7000-9000", ">9000")
                                           )
                                       )

adex <- cut_by_group(adex, "AVAL", "PARAM", group, "AVALCAT1", as_factor = TRUE)

adex[,"PARAM"] <- factor(adex[["PARAM"]],
                   c("Total number of doses administered",
                     "Total dose administered",
                     "Dose administered during constant dosing interval",
                     "Number of doses administered during constant dosing interval"))

ext01_2(adex, adsl, lbl_overall = "", show_stats = "ALL")
ext01_2(adex, adsl, lbl_overall = "All Patients", show_stats = "ALL")

}
