% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\name{ext01_2}
\alias{ext01_2}
\alias{ext01_2_lyt}
\title{EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options}
\usage{
ext01_2(
  adam_db,
  armvar = .study$actarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(actarm = "ACTARM", lbl_overall = NULL)
)

ext01_2_lyt(
  armvar = .study$actarm,
  summaryvars = c("AVAL", "AVALCAT1"),
  summaryvars_lbls = c("Summary", "Categories"),
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(actarm = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{summaryvars}{(\code{string}) the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{summaryvars_lbls}{(\code{string}) the label associated with the analyzed variable.}
}
\description{
EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options

EXT01 Layout 2 (Supplementary)
}
\details{
\itemize{
\item Supplementary Exposure table with binning of desired analysis values.
\item The \code{n} row provides the number of non-missing values. The percentages for categorical variables is based on \code{n}.
The percentages for \verb{Total number of patients with at least one dose modification} are based on the number of
patients in the corresponding analysis population given by \code{N}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted by alphabetic order of the \code{PARAM} value. Transform to factor and re-level for custom order.
\item \code{ANL01FL} is not relevant subset
}
}
\section{Functions}{
\itemize{
\item \code{ext01_2_lyt}: 
}}

\examples{
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
  preprocess_data("ext01_2")

ext01_2(db)
}
