% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext01.R
\name{ext01_lyt}
\alias{ext01_lyt}
\title{ext01 Layout}
\usage{
ext01_lyt(arm_var, summaryvars, summaryvars_lbls, lbl_overall)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\code{character}) the name of the variable to be analyzed. By default \code{"AVAL"}.}

\item{summaryvars_lbls}{(\code{character}) the label associated with the analyzed variable.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
ext01 Layout
}
\keyword{internal}
