% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{filter_adex_drug}
\alias{filter_adex_drug}
\title{Filter \code{adex} for \code{PARCAT1}}
\usage{
filter_adex_drug(adam_db)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the adam Datasets.}
}
\description{
Filter \code{adex} for \code{PARCAT1}
}
