% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstg01.R
\docType{data}
\name{fstg01_main}
\alias{fstg01_main}
\alias{fstg01_pre}
\alias{fstg01_post}
\alias{fstg01}
\title{\code{FSTG01} Subgroup Analysis of Best Overall Response.}
\format{
An object of class \code{chevron_g} of length 1.
}
\usage{
fstg01_main(
  adam_db,
  dataset = "adrs",
  arm_var = "ARM",
  col_x = attr(tbl, "col_x"),
  col_ci = attr(tbl, "col_ci"),
  vline = 1,
  forest_header = attr(tbl, "forest_header"),
  xlim = c(0.1, 10),
  logx = TRUE,
  x_at = c(0.1, 1, 10),
  width_row_names = NULL,
  width_columns = NULL,
  width_forest = grid::unit(1, "null"),
  col_symbol_size = attr(tbl, "col_symbol_size"),
  col = getOption("ggplot2.discrete.colour")[1],
  draw = TRUE,
  newpage = TRUE,
  line_col = nestcolor::color_palette(),
  ...
)

fstg01_pre(adam_db, dataset = "adrs", ...)

fstg01_post(tlg, ...)

fstg01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{col_x}{(\code{string}) column index with estimator.}

\item{col_ci}{(\code{string}) column index with confidence intervals.}

\item{vline}{(\code{numeric}) x coordinate for vertical line.}

\item{forest_header}{(\code{character}) text displayed to the left and right of vline, respectively.}

\item{xlim}{(\code{character}) limits for x axis.}

\item{logx}{(\code{flag}) shows the x-values on logarithm scale.}

\item{x_at}{(\code{character}) x-tick locations.}

\item{col_symbol_size}{(\code{numeric}) is used to determine relative size for estimator plot symbol.}

\item{line_col}{(\code{character}) describes the colors to use for the lines or a named \code{character}
associating values of \code{arm_var} with color names.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
\code{FSTG01} Subgroup Analysis of Best Overall Response.
}
\details{
\itemize{
\item No overall value.
\item Keep zero count rows by default.
}
}
\section{Functions}{
\itemize{
\item \code{fstg01_main()}: Main TLG Function

\item \code{fstg01_pre()}: Preprocessing

\item \code{fstg01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the table specified by \code{dataset} with the columns specified by  \code{arm_var}.
}
}
\examples{
library(dplyr)
library(dunlin)

col <- c(
  "Overall" = "blue"
)

proc_data <- log_filter(syn_data, PARAMCD == "OVRINV", "adrs")
run(fstg01, proc_data, dataset = "adrs", line_col = col)

proc_data <- log_filter(syn_data, PARAMCD == "BESRSPI", "adrs")
run(fstg01, proc_data,
  rsp = c("CR", "PR"), subgroups = c("SEX", "AGEGR1", "RACE"),
  strata = c("STRATA2"), conf_level = 0.90, dataset = "adrs"
)
}
\keyword{datasets}
