% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstg02.R
\docType{data}
\name{fstg02_main}
\alias{fstg02_main}
\alias{fstg02_pre}
\alias{fstg02_post}
\alias{fstg02}
\title{\code{fstg02} Subgroup Analysis of Survival Duration.}
\format{
An object of class \code{chevron_g} of length 1.
}
\usage{
fstg02_main(
  adam_db,
  dataset = "adtte",
  arm_var = "ARM",
  subgroups = c("SEX", "AGEGR1", "RACE"),
  strata_var = NULL,
  ...
)

fstg02_pre(adam_db, ...)

fstg02_post(tlg, ...)

fstg02
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) the arm variable name used for group splitting.}

\item{subgroups}{(\code{character}) the subgroups variable name to list baseline risk factors.}

\item{strata_var}{(\code{character}) required if stratified analysis is performed.}

\item{...}{Further arguments passed to \code{g_forest} and \code{extract_rsp_subgroups} (a wrapper for
\code{h_odds_ratio_subgroups_df} and \code{h_proportion_subgroups_df}). For details, see the documentation in \code{tern}.
Commonly used arguments include \code{col_symbol_size}, \code{col}, \code{vline}, \code{groups_lists}, \code{conf_level},
\code{method}, \code{label_all}, etc.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
a \code{gTree} object.
}
\description{
The template produces the subgroup analysis of survival duration graphic.
}
\details{
\itemize{
\item No overall value.
\item Keep zero count rows by default.
}
}
\section{Functions}{
\itemize{
\item \code{fstg02_main()}: Main TLG Function

\item \code{fstg02_pre()}: Preprocessing

\item \code{fstg02_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the table specified by \code{dataset} with \code{"PARAMCD"}, \code{"ARM"},
\code{"AVAL"}, \code{"AVALU"}, \code{"CNSR"}, and the columns specified by \code{subgroups} which is denoted as
\code{c("SEX", "AGEGR1", "RACE")} by default.
}
}
\examples{
library(dplyr)
library(dunlin)

proc_data <- log_filter(
  syn_data,
  PARAMCD == "OS" & ARM \%in\% c("A: Drug X", "B: Placebo"), "adtte"
)
run(fstg02, proc_data,
  subgroups = c("SEX", "AGEGR1", "RACE"),
  conf_level = 0.90, dataset = "adtte"
)
}
\keyword{datasets}
