% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_args.R
\name{gen_args}
\alias{gen_args}
\title{General Argument Name Convention}
\usage{
gen_args(
  adam_db,
  armvar,
  dataset,
  lbl_overall,
  prune_0,
  req_tables,
  deco,
  .study,
  group,
  tlgfname,
  pmap
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting.}

\item{dataset}{(\code{string}) the name of the data set to be analyzed.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{group}{(\verb{list of lists}) for group-dependent data binning}

\item{tlgfname}{(\code{character}) function name of tlg-function}

\item{pmap}{(\verb{data frame}) preprocessing map, see \code{std_pmap} for the required format}
}
\description{
General Argument Name Convention
}
\details{
the following arguments are better provided through the study object: \code{lbl_overall}, \code{armvar}.
}
