% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_args.R
\name{gen_args}
\alias{gen_args}
\title{General Argument Name Convention}
\usage{
gen_args(
  armvar,
  lbl_overall,
  prune_0,
  deco,
  .study,
  ad_bl,
  adsl,
  adeg,
  adae,
  adcm,
  admh,
  aag,
  advs,
  adlb,
  adex,
  adsub,
  adrs,
  adtte,
  adhy,
  adsaftte,
  addv,
  adcssrs,
  adqlqc,
  acg,
  adic
)
}
\arguments{
\item{armvar}{variable used for column splitting}

\item{lbl_overall}{label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{remove 0 count rows}

\item{deco}{decoration list with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{a list with default values for the arguments of the function}

\item{ad_bl}{subject baseline dataset with one row per patient}

\item{adsl}{Subject Level Analysis Dataset}

\item{adeg}{ECG Analysis Dataset}

\item{adae}{Adverse Events Analysis Dataset}

\item{adcm}{Concomitant Medications Analysis Dataset}

\item{admh}{Medical History Analysis Dataset}

\item{aag}{Adverse Event Grouping Defs Dataset}

\item{advs}{Vital Signs Analysis Dataset}

\item{adlb}{Laboratory Analysis Dataset}

\item{adex}{Exposure Analysis Dataset}

\item{adsub}{Subcategory Analysis Dataset}

\item{adrs}{Response Analysis Dataset}

\item{adtte}{Time to Event Analysis Dataset}

\item{adhy}{Hys Law Analysis Dataset}

\item{adsaftte}{Safety Time to Event Analysis Dataset}

\item{addv}{Protocol Deviations Analysis Dataset}

\item{adcssrs}{C-SSRS Analysis Dataset}

\item{adqlqc}{EORTC QLQ-C30 V3 Analysis Dataset}

\item{acg}{Concomitant Med. Grouping Defs Dataset}

\item{adic}{Interval Censoring Analysis Dataset}
}
\description{
General Argument Name Convention
}
