% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_args.R
\name{gen_args}
\alias{gen_args}
\title{General Argument Name Convention}
\usage{
gen_args(
  adam_db,
  lyt_ls,
  main,
  preprocess,
  postprocess,
  adam_dataset,
  type,
  armvar,
  lbl_overall,
  prune_0,
  req_tables,
  deco,
  .study,
  group,
  tlg,
  tlgfname
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_ls}{(\verb{list of functions} returning \code{PreDataTableLayout} object. Typically one of the \verb{_lyt} function of
\code{chevron} stored in a \code{list}.}

\item{main}{(\code{function}) returning a \code{tlg}, with \code{adam_db} as first argument and \code{...} as last argument. Typically
one of the \verb{_main} function of \code{chevron}.}

\item{preprocess}{(\code{function}) returning a pre-processed \code{dm} object, with \code{adam_db} as first argument and \code{...} as
last argument. Typically one of the \verb{_pre} function of \code{chevron}.}

\item{postprocess}{(\code{function}) returning a post-processed \code{tlg}, with \code{tlg} as first argument.}

\item{type}{(\code{string}) indicating the subclass.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{group}{(\verb{list of lists}) for group-dependent data binning}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{tlgfname}{(\code{character}) function name of TLG function}

\item{adam_datasets}{(\code{character}) representing the names of the tables from an \code{ADaM} data set required for \code{tlg}
creation.}
}
\description{
General Argument Name Convention
}
\details{
the following arguments are better provided through the study object: \code{lbl_overall}, \code{armvar}.
}
