% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_args.R
\name{gen_args}
\alias{gen_args}
\title{General Argument Name Convention}
\usage{
gen_args(
  armvar,
  lbl_overall,
  prune_0,
  deco,
  .study,
  ad_bl,
  adsl,
  adeg,
  adae,
  adcm,
  admh,
  aag,
  advs,
  adlb,
  adex,
  adsub,
  adrs,
  adtte,
  adhy,
  adsaftte,
  addv,
  adcssrs,
  adqlqc,
  acg,
  adic,
  group
)
}
\arguments{
\item{armvar}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{list}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{ad_bl}{(\code{data.frame}) subject baseline dataset with one row per patient}

\item{adsl}{(\code{data.frame}) Subject Level Analysis Dataset}

\item{adeg}{(\code{data.frame}) ECG Analysis Dataset}

\item{adae}{(\code{data.frame}) Adverse Events Analysis Dataset}

\item{adcm}{(\code{data.frame}) Concomitant Medications Analysis Dataset}

\item{admh}{(\code{data.frame}) Medical History Analysis Dataset}

\item{aag}{(\code{data.frame}) Adverse Event Grouping Defs Dataset}

\item{advs}{(\code{data.frame}) Vital Signs Analysis Dataset}

\item{adlb}{(\code{data.frame}) Laboratory Analysis Dataset}

\item{adex}{(\code{data.frame}) Exposure Analysis Dataset}

\item{adsub}{(\code{data.frame}) Subcategory Analysis Dataset}

\item{adrs}{(\code{data.frame}) Response Analysis Dataset}

\item{adtte}{(\code{data.frame}) Time to Event Analysis Dataset}

\item{adhy}{(\code{data.frame}) Hys Law Analysis Dataset}

\item{adsaftte}{(\code{data.frame}) Safety Time to Event Analysis Dataset}

\item{addv}{(\code{data.frame}) Protocol Deviations Analysis Dataset}

\item{adcssrs}{(\code{data.frame}) C-SSRS Analysis Dataset}

\item{adqlqc}{(\code{data.frame}) EORTC QLQ-C30 V3 Analysis Dataset}

\item{acg}{(\code{data.frame}) Concomitant Med. Grouping Defs Dataset}

\item{adic}{(\code{data.frame}) Interval Censoring Analysis Dataset}

\item{group}{nest list for group-dependent data binning}
}
\description{
General Argument Name Convention
}
