% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{get_adam_datasets}
\alias{get_adam_datasets}
\alias{get_adam_datasets,chevron_tlg-method}
\title{Retrieve names of datasets associated with the object}
\usage{
get_adam_datasets(object)

\S4method{get_adam_datasets}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}
}
\description{
Retrieve names of datasets associated with the object
}
