% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_db_data}
\alias{get_db_data}
\title{Get Data from a DB}
\usage{
get_db_data(db, ...)
}
\arguments{
\item{db}{a named list or a \code{dm} object.}

\item{...}{not used.}
}
\value{
named list of datasets
}
\description{
Get Data from a DB
}
\examples{
\donttest{
  get_db_data(list(iris = iris, mtcars = mtcars, CO2 = CO2))
  get_db_data(list(iris = iris, mtcars = mtcars, CO2 = CO2), "iris")
  get_db_data(list(iris = iris, mtcars = mtcars, CO2 = CO2), "iris", "CO2")

  db <- dm::dm_nycflights13() \%>\%
    dm_filter(airports, name == "John F Kennedy Intl")

  get_db_data(db, "airports")
}

}
