% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{main,chevron_tlg-method}
\alias{main,chevron_tlg-method}
\alias{get_main}
\alias{get_main,chevron_tlg-method}
\title{Retrieve Main Function}
\usage{
\S4method{main}{chevron_tlg}(x)

get_main(object, ...)

\S4method{get_main}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Main Function
}
