% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_main}
\alias{get_main}
\alias{get_main,chevron_tlg-method}
\title{Retrieve Main Function}
\usage{
get_main(object, ...)

\S4method{get_main}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Main Function

get_main
}
