% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{get_postprocess}
\alias{get_postprocess}
\alias{get_postprocess,chevron_tlg-method}
\title{Retrieve post-processing Function}
\usage{
get_postprocess(object, ...)

\S4method{get_postprocess}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve post-processing Function
}
