% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_preprocess}
\alias{get_preprocess}
\alias{get_preprocess,tlg_pipeline_S4-method}
\title{Retrieve Pre Processing Function}
\usage{
get_preprocess(object, ...)

\S4method{get_preprocess}{tlg_pipeline_S4}(object)
}
\arguments{
\item{object}{(\code{tlg_pipeline_S4}) input.}

\item{...}{not used.}
}
\description{
Retrieve Pre Processing Function

get_preprocess
}
