% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4methods.R
\name{get_preprocess}
\alias{get_preprocess}
\alias{get_preprocess,chevron_tlg-method}
\title{Retrieve Pre Processing Function}
\usage{
get_preprocess(object, ...)

\S4method{get_preprocess}{chevron_tlg}(object)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{...}{not used.}
}
\description{
Retrieve Pre Processing Function

get_preprocess
}
