% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_list}
\alias{gg_list}
\alias{as.gg_list}
\alias{as.gg_list.list}
\title{List of \code{gg} object}
\usage{
gg_list(...)

as.gg_list(obj)

\method{as.gg_list}{list}(obj)
}
\arguments{
\item{...}{(\code{ggplot}) objects.}

\item{obj}{(\code{ggplot} or \code{list} of \code{ggplot})}
}
\value{
a \code{gg_list} object.

a \code{gg_list} object.
}
\description{
List of \code{gg} object

Convert Object to List of \code{gg_list}.
}
