% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_format_dec}
\alias{h_format_dec}
\title{Decimal Formatting}
\usage{
h_format_dec(x, digits = NA, format = NA)
}
\arguments{
\item{x}{(\code{numeric}) to format.}

\item{digits}{(\code{integer}) number of digits.}

\item{format}{(\code{string}) describing how the numbers should be formatted following the \code{sprintf} syntax.}
}
\value{
\code{string} with formatted numbers.
}
\description{
Decimal Formatting
}
\examples{
h_format_dec(c(123, 5676.56), 1, "\%f - \%f")

}
