% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_format_dec}
\alias{h_format_dec}
\title{Decimal Formatting}
\usage{
h_format_dec(digits, format, ne = FALSE)
}
\arguments{
\item{digits}{(\code{integer}) number of digits.}

\item{format}{(\code{string}) describing how the numbers should be formatted following the \code{sprintf} syntax.}

\item{ne}{(\code{flag}) indicator whether to use "NE" to replace the actual value.}
}
\value{
\code{function} formatting numbers with the defined format.
}
\description{
Decimal Formatting
}
\examples{
fun <- h_format_dec(c(1, 1), "\%s - \%s")
fun(c(123, 567.89))

}
