% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_format_dec}
\alias{h_format_dec}
\title{Decimal Formatting}
\usage{
h_format_dec(digits = NA, format = NA)
}
\arguments{
\item{digits}{(\code{integer}) number of digits.}

\item{format}{(\code{string}) describing how the numbers should be formatted following the \code{sprintf} syntax.}
}
\value{
\code{function} formatting numbers with the defined format or \code{NULL} if the format is not defined.
}
\description{
Decimal Formatting
}
\examples{
fun <- h_format_dec(1, "\%f - \%f")
fun(c(123, 567.89))

}
