% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_pad_or_round}
\alias{h_pad_or_round}
\title{Padding or Rounding Numbers}
\usage{
h_pad_or_round(x, digits = NA, ...)
}
\arguments{
\item{x}{(\code{numeric}) to format.}

\item{digits}{(\code{integer}) number of digits.}

\item{...}{not used.}
}
\value{
\code{string} corresponding to the number with the desired numbers of digits.
}
\description{
Padding or Rounding Numbers
}
\examples{
h_pad_or_round(123.1234, 10)
h_pad_or_round(123.15, 1)
h_pad_or_round(123, 1)
h_pad_or_round(123, 3)
}
