% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{ifneeded_split_col}
\alias{ifneeded_split_col}
\title{Helper function to add a column split if specified}
\usage{
ifneeded_split_col(lyt, var, lbl_var, split_fun = NULL)
}
\arguments{
\item{lyt}{(\code{rtables}) object.}

\item{var}{(\code{string}) the name of the variable initiating a new column split.}

\item{lbl_var}{(\code{string}) the label of the variable \code{var}.}

\item{split_fun}{(\code{function} or \code{NULL}) custom splitting functions.}
}
\value{
\code{rtables} object.
}
\description{
Helper function to add a column split if specified
}
\keyword{internal}
