% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{ifneeded_split_col}
\alias{ifneeded_split_col}
\title{Helper function to add a col split if specified}
\usage{
ifneeded_split_col(lyt, var, ...)
}
\arguments{
\item{lyt}{(\code{PreDataTableLayouts}) object.}

\item{var}{(\code{string}) the name of the variable initiating a new column split.}

\item{...}{additional arguments for \code{split_cols_by}.}
}
\value{
\code{PreDataTableLayouts} object.
}
\description{
Helper function to add a col split if specified
}
\keyword{internal}
