% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmg01.R
\docType{data}
\name{kmg01_main}
\alias{kmg01_main}
\alias{kmg01_pre}
\alias{kmg01_post}
\alias{kmg01}
\title{\code{KMG01} \code{Kaplan-Meier} Plot 1.}
\format{
An object of class \code{chevron_g} of length 1.
}
\usage{
kmg01_main(adam_db, dataset = "adtte", arm_var = "ARM", ...)

kmg01_pre(adam_db, dataset = "adtte", ...)

kmg01_post(tlg, ...)

kmg01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{...}{Further arguments passed to \code{g_km} and \code{control_coxph}. For details, see
the documentation in \code{tern}.
Commonly used arguments include \code{col}, \code{pval_method}, \code{ties}, \code{conf_level}, \code{conf_type},
\code{annot_coxph}, \code{annot_stats}, etc.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
a \code{gTree} object.
}
\description{
\code{KMG01} \code{Kaplan-Meier} Plot 1.
}
\details{
\itemize{
\item No overall value.
}
}
\section{Functions}{
\itemize{
\item \code{kmg01_main()}: Main TLG Function

\item \code{kmg01_pre()}: Preprocessing

\item \code{kmg01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the table specified by \code{dataset} with the columns specified by  \code{arm_var}.
}
}
\examples{
library(dplyr)
library(dunlin)

col <- c(
  "A: Drug X" = "black",
  "B: Placebo" = "blue",
  "C: Combination" = "gray"
)

syn_data2 <- log_filter(syn_data, PARAMCD == "OS", "adtte")
run(kmg01, syn_data2, dataset = "adtte", line_col = col)

syn_data3 <- log_filter(syn_data, PARAMCD == "AEREPTTE", "adaette")
run(kmg01, syn_data3, dataset = "adaette")
}
\keyword{datasets}
