% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmg01.R
\docType{data}
\name{kmg01_main}
\alias{kmg01_main}
\alias{kmg01_pre}
\alias{kmg01_post}
\alias{kmg01}
\title{\code{KMG01} Kaplan-Meier Plot 1.}
\format{
An object of class \code{chevron_g} of length 1.
}
\usage{
kmg01_main(
  adam_db,
  dataset = "adtte",
  arm_var = "ARM",
  x_name = "Time (Days)",
  y_name = "Survival Probability",
  show_statis = TRUE,
  show_censor = TRUE,
  pval_method = "wald",
  ties = "exact",
  conf_level = 0.95,
  position_coxph = c(0, 0.05),
  position_surv_med = c(0.9, 0.9),
  line_col = as.list(nestcolor::color_palette()),
  ...
)

kmg01_pre(adam_db, dataset = "adtte", ...)

kmg01_post(tlg, ...)

kmg01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{x_name}{(\code{string}) the name of the x-axis.}

\item{y_name}{(\code{string}) the name of the x-axis.}

\item{show_statis}{(\code{flag}) should the summary statistic table be displayed.}

\item{show_censor}{(\code{flag}) should the censor flag be displayed.}

\item{pval_method}{(\code{string}) p-value method for testing hazard ratio = 1. One of \code{"log-rank"}, \code{"wald"} or
\code{"likelihood"}.}

\item{ties}{(\code{string}) specifying the method for tie handling. One of \code{"efron"}, \code{"breslow"} or \code{"exact"}.}

<<<<<<< HEAD:man/kmg01.Rd
\item{conf_level}{(\code{numeric}) the confidence level.}
=======
\item{conf_level}{(\code{numeric}) the level of confidence interval, default is 0.95.}
>>>>>>> 476_harmonize_templates@main:man/kmg01_1.Rd

\item{position_coxph}{(\code{numeric}) x and y positions for plotting survival::coxph() model.}

\item{position_surv_med}{(\code{numeric}) x and y positions for plotting annotation table estimating
median survival time per group.}

\item{line_col}{(\code{list}) describing the colors to use for the lines or a named \code{list}
associating values of \code{arm_var} with color names.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
\code{KMG01} Kaplan-Meier Plot 1.
}
\details{
\itemize{
\item No overall value.
}
}
\section{Functions}{
\itemize{
\item \code{kmg01_main()}: Main TLG Function

\item \code{kmg01_pre()}: Preprocessing

\item \code{kmg01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain the table specified by \code{dataset} with the columns specified by  \code{arm_var}.
}
}
\examples{
library(dplyr)
library(dunlin)

col <- list(
  "A: Drug X" = "black",
  "B: Placebo" = "blue",
  "C: Combination" = "gray"
)

syn_data2 <- log_filter(syn_data, PARAMCD == "OS", "adtte")
run(kmg01, syn_data2, dataset = "adtte", line_col = col)

syn_data3 <- log_filter(syn_data, PARAMCD == "AEREPTTE", "adaette")
run(kmg01, syn_data3, dataset = "adaette")
}
\keyword{datasets}
