% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt01.R
\docType{data}
\name{lbt01_main}
\alias{lbt01_main}
\alias{lbt01_pre}
\alias{lbt01}
\title{\code{LBT01} Lab Results and Change from Baseline by Visit Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\source{
\code{cfbt01.R}
}
\usage{
lbt01_main(
  adam_db,
  dataset = "adlb",
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  page_var = NULL,
  ...
)

lbt01_pre(adam_db, dataset = "adlb", ...)

lbt01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{row_split_var}{(\code{character}) additional row split variables.}

\item{summaryvars}{(\code{character}) variables to be analyzed. The label attribute of the corresponding column in
table of \code{adam_db} is used as label.}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed}

\item{precision}{(named \code{list} of \code{integer}) where names are values found in the \code{PARAMCD} column and the the values
indicate the number of digits that should be represented for \code{min}, \code{max} and \code{median}. \code{Mean} and \code{sd} are
represented with one more decimal of precision.}

\item{default_precision}{(\code{integer}) the default number of digits.}

\item{page_var}{(\code{string}) variable name prior to which the row split is by page.}

\item{...}{not used.}
}
\description{
The \code{LBT01} table provides an
overview of the Lab values and its change from baseline of each respective arm
over the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{lbt01_main()}: Main TLG function

\item \code{lbt01_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain table named as \code{dataset} with the columns specified in \code{summaryvars}.
}
}
\examples{
run(lbt01, syn_data)
}
\keyword{datasets}
