% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt01.R
\name{lbt01_1}
\alias{lbt01_1}
\alias{lbt01_1_lyt}
\title{LBT01 Table 1 (Default) Laboratory Test Results and Change from Baseline by Visit}
\usage{
lbt01_1(
  adsl,
  adlb,
  armvar = .study$armvar,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = var_labels_for(adlb, summaryvars),
  visitvar = "AVISIT",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("LBT01"),
  .study = list(armvar = "ACTARM", lbl_overall = "", evo_vars = c("AVAL", "CHG"))
)

lbt01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = .study$evo_vars_lbls,
  visitvar = .study$visitvar,
  lbl_overall = .study$lbl_overall,
  lbl_AVISIT = "",
  lbl_PARAM = "",
  deco = std_deco("LBT01"),
  .study = list(armvar = "ACTARM", lbl_overall = "", evo_vars = c("AVAL", "CHG"),
    evo_vars_lbls = c("Analysis \\nValue", "Change from \\nBaseline"), visitvar =
    "AVISIT")
)
}
\arguments{
\item{adsl}{(\code{data.frame}) Subject Level Analysis Dataset}

\item{adlb}{(\code{data.frame}) Laboratory Analysis Dataset}

\item{armvar}{(\code{string}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of string}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\verb{vector of string}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point to
be displayed}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{list}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AVISIT}{(\code{string}) label of the \code{visitvar} variable.}

\item{lbl_PARAM}{(\code{string}) label of the \code{PARAM} variable.}
}
\value{

}
\description{
The LBT01 table provides an overview of the analysis values and its change from baseline of each respective arm over
the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{lbt01_1_lyt}: 
}}

\examples{
library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adlb <- sd$adlb \%>\%
 mutate(ANL01FL = 'Y')

lbt01_1(adsl, adlb)

lbt01_1(adsl, adlb, summaryvars_lbls = c("Analysis", "Change"))

}
