% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt01.R
\name{lbt01_1}
\alias{lbt01_1}
\alias{lbt01_1_lyt}
\title{LBT01 Table 1 (Default) Laboratory Test Results and Change from Baseline by Visit}
\usage{
lbt01_1(
  adam_db,
  armvar = .study$armvar,
  summaryvars = c("AVAL", "CHG"),
  summaryvars_lbls = c("Value at Visit", "Change from \\nBaseline"),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("LBT01"),
  .study = list(armvar = "ACTARM")
)

lbt01_1_lyt(
  armvar = .study$armvar,
  summaryvars = c("AVAL", "CHG"),
  summaryvars_lbls = c("Value at Visit", "Change from \\nBaseline"),
  visitvar = .study$visitvar,
  lbl_AVISIT = "Visit",
  lbl_PARAM = "Parameter",
  deco = std_deco("LBT01"),
  .study = list(armvar = "ACTARM", visitvar = "AVISIT")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the adam Datasets.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_AVISIT}{(\code{character}) label of the \code{visitvar} variable.}

\item{lbl_PARAM}{(\code{character}) label of the \code{PARAM} variable.}
}
\value{

}
\description{
The LBT01 table provides an overview of the analysis values and its change from baseline of each respective arm over
the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{lbt01_1_lyt}: 
}}

\examples{
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
   dm_select_tbl(adsl, adlb)

db <- db \%>\%
  (std_filter_fun("lbt01_1"))() \%>\%
  (std_mutate_fun("lbt01_1"))()

lbt01_1(db)

}
