% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt04.R
\docType{data}
\name{lbt04_main}
\alias{lbt04_main}
\alias{lbt04_lyt}
\alias{lbt04_pre}
\alias{lbt04_post}
\alias{lbt04}
\title{\code{LBT04} Laboratory Abnormalities Not Present at Baseline Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt04_main(adam_db, arm_var = "ACTARM", ...)

lbt04_lyt(arm_var, lbl_param, lbl_anrind)

lbt04_pre(adam_db, ...)

lbt04_post(tlg, ...)

lbt04
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{...}{not used.}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{lbl_anrind}{(\code{string}) label of the \code{ANRIND} variable.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
The \code{LBT04} table provides an
overview of laboratory abnormalities not present at baseline.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Lab test results with missing \code{ANRIND} values are excluded.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{lbt04_main()}: Main TLG function

\item \code{lbt04_lyt()}: Layout

\item \code{lbt04_pre()}: Preprocessing

\item \code{lbt04_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns \code{"PARCAT1"}, \code{"PARCAT2"}, \code{"PARAM"}, \code{"ANRIND"},
and column specified by \code{arm_var}.
}
}
\examples{
run(lbt04, syn_data)
}
\keyword{datasets}
\keyword{internal}
