% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt06.R
\docType{data}
\name{lbt06_main}
\alias{lbt06_main}
\alias{lbt06_lyt}
\alias{lbt06_pre}
\alias{lbt06_check}
\alias{lbt06_post}
\alias{lbt06}
\title{\code{LBT06} Table 1 (Default) Laboratory Abnormalities by Visit and Baseline Status.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
lbt06_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = c("AVISIT", "ANRIND"),
  paramcd = NULL,
  lbl_param = "Laboratory Test",
  lbl_anrind = "Direction of Abnormality",
  deco = std_deco("LBT06"),
  ...
)

lbt06_lyt(arm_var, summaryvars, paramcd, lbl_paramcd, lbl_anrind, deco)

lbt06_pre(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = c("AVISIT", "ANRIND"),
  lab_vars = c("ONTRTFL", "PARCAT2", "AVAL"),
  paramcd = NULL,
  ...
)

lbt06_check(
  adam_db,
  arm_var,
  summaryvars,
  lab_vars,
  paramcd,
  req_tables = c("adsl", "adlb")
)

lbt06_post(tlg, prune_0 = FALSE, ...)

lbt06
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_param}{(\code{string}) label of the \code{PARAM} variable.}

\item{lbl_anrind}{(\code{string}) label of the \code{ANRIND} variable.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{LBT06} table produces the standard laboratory abnormalities by visit and
baseline status summary.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Lab test results with missing \code{ANRIND} values are excluded.
\item Split columns by arm, typically \code{ACTARM}.
\item Do include a total column by default.
}
}
\section{Functions}{
\itemize{
\item \code{lbt06_main()}: Main TLG function

\item \code{lbt06_lyt()}: Layout

\item \code{lbt06_pre()}: Preprocessing

\item \code{lbt06_check()}: Checks

\item \code{lbt06_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adlb} table with columns ``,
and column specified by \code{arm_var}.
}
}
\examples{
run(lbt06, syn_data)

run(lbt06, syn_data, paramcd = "PARAMCD")
}
\keyword{datasets}
