% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt15.R
\name{lbt15_pre}
\alias{lbt15_pre}
\alias{lbt15}
\title{\code{LBT15} Laboratory Test Shifts to \code{NCI-CTCAE} Grade 3-4 Post-Baseline Table.}
\source{
\code{lbt04.R}
}
\usage{
lbt15_pre(adam_db, ...)

lbt15
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{...}{not used.}
}
\value{
the preprocessing function returns a \code{list} of \code{data.frame}.
}
\description{
\code{LBT15} Laboratory Test Shifts to \code{NCI-CTCAE} Grade 3-4 Post-Baseline Table.
}
\section{Functions}{
\itemize{
\item \code{lbt15_pre()}: Preprocessing

}}
\examples{
run(lbt15, syn_data)
}
