% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt15.R
\docType{data}
\name{lbt15_main}
\alias{lbt15_main}
\alias{lbt15_pre}
\alias{lbt15_post}
\alias{lbt15}
\title{\code{LBT15} Laboratory Test Shifts to \code{NCI-CTCAE} Grade 3-4 Post-Baseline Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\source{
\code{lbt04.R}

\code{lbt04.R}
}
\usage{
lbt15_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  analysis_abn_var = "ANRIND",
  baseline_abn_var = "BNRIND",
  row_split_var = "PARCAT1",
  page_var = tail(row_split_var, 1L),
  ...
)

lbt15_pre(adam_db, ...)

lbt15_post(tlg, ...)

lbt15
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{analysis_abn_var}{(\code{string}) column describing anomaly magnitude}

\item{baseline_abn_var}{(\code{string}) column describing anomaly at baseline.}

\item{row_split_var}{(\code{character}) additional row split variables.}

\item{page_var}{(\code{string}) variable name prior to which the row split is by page.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
the preprocessing function returns a \code{list} of \code{data.frame}.
}
\description{
\code{LBT15} Laboratory Test Shifts to \code{NCI-CTCAE} Grade 3-4 Post-Baseline Table.
}
\section{Functions}{
\itemize{
\item \code{lbt15_main()}: Main TLG function

\item \code{lbt15_pre()}: Preprocessing

\item \code{lbt15_post()}: Postprocessing

}}
\examples{
run(lbt15, syn_data)
}
\keyword{datasets}
