% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbt15.R
\docType{data}
\name{lbt15_main}
\alias{lbt15_main}
\alias{lbt15_pre}
\alias{lbt15}
\title{\code{LBT15} Laboratory Test Shifts to \code{NCI}-\code{CTCAE} Grade 3-4 Post-Baseline Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\source{
\code{lbt04.R}
}
\usage{
lbt15_main(
  adam_db,
  arm_var = "ACTARM",
  abn_range = "ATOXGR",
  baseline = "BTOXGR",
  ...
)

lbt15_pre(adam_db, ...)

lbt15
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{abn_range}{(\code{string}) column describing anomaly magnitude}

\item{baseline}{(\code{string}) column describing anomaly at baseline.}

\item{...}{not used.}
}
\description{
\code{LBT15} Laboratory Test Shifts to \code{NCI}-\code{CTCAE} Grade 3-4 Post-Baseline Table.
}
\section{Functions}{
\itemize{
\item \code{lbt15_main()}: Main TLG function

\item \code{lbt15_pre()}: Preprocessing

}}
\examples{
run(lbt15, syn_data)
}
\keyword{datasets}
