% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lvls}
\alias{lvls}
\title{\code{lvls} objective to get unique level of variables}
\usage{
lvls(x)
}
\arguments{
\item{x}{(\code{string}) the variable needs to get unique level}
}
\description{
\code{lvls} objective to get unique level of variables
}
