% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{args,chevron_tlg-method}
\alias{args,chevron_tlg-method}
\alias{main}
\alias{main,chevron_tlg-method}
\alias{main<-}
\alias{main<-,chevron_tlg-method}
\title{Main}
\usage{
\S4method{args}{chevron_tlg}(x, simplify = FALSE)

main(x)

\S4method{main}{chevron_tlg}(x)

main(x) <- value

\S4method{main}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{value}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{_main} function of \code{chevron}.}
}
\description{
retrieve or set \code{main} function.
}
