% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mht01.R
\docType{data}
\name{mht01_main}
\alias{mht01_main}
\alias{mht01_pre}
\alias{mht01_post}
\alias{mht01}
\title{\code{MHT01} Medical History Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
mht01_main(adam_db, arm_var = "ARM", lbl_overall = NULL, ...)

mht01_pre(adam_db, ...)

mht01_post(tlg, prune_0 = TRUE, ...)

mht01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{MHT01} table provides an overview of the subjects medical
history by SOC and Preferred Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Order by body system alphabetically and within body system and medical condition by decreasing total number of
patients with the specific condition.
}
}
\section{Functions}{
\itemize{
\item \code{mht01_main()}: Main TLG function

\item \code{mht01_pre()}: Preprocessing

\item \code{mht01_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{admh} table with columns \code{"MHBODSYS"} and \code{"MHDECOD"}.
}
}
\examples{
run(mht01, syn_data)
}
\keyword{datasets}
