% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mht01.R
\name{mht01_1}
\alias{mht01_1}
\alias{mht01_1_lyt}
\title{MHT01 Table 1 (Default) Medical History Table 1}
\usage{
mht01_1(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("MHT01"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

mht01_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_MHBODSYS = "Body System or Organ Class",
  lbl_MHDECOD = "MedDRA preferred Term",
  deco = std_deco("MHT01"),
  .study = list(armvar = "ARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_MHBODSYS}{(\code{character}) text label for MHBODSYS.}

\item{lbl_MHDECOD}{(\code{character}) text label for MHDECOD.}
}
\description{
The MHT01 Table (Medical History table by SOC and Preferred Term)
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Order by body system alphabetically and within body system and medical condition by decreasing total number of
patients with the specific condition.
}
}
\section{Functions}{
\itemize{
\item \code{mht01_1_lyt}: 
}}

\examples{
library(dm)
library(rtables)

db <- syn_test_data() \%>\%
  preprocess_data("mht01_1")

mht01_1(adam_db = db) \%>\% head(15)

mht01_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("MHT01")
)
}
