% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mht01.R
\docType{data}
\name{mht01_1_main}
\alias{mht01_1_main}
\alias{mht01_1_lyt}
\alias{mht01_1_pre}
\alias{mht01_1}
\title{\code{MHT01} Table 1 (Default) Medical History Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
mht01_1_main(
  adam_db,
  armvar = "ARM",
  lbl_overall = NULL,
  lbl_mhbodsys = "MedDRA System Organ Class",
  lbl_mhdecod = "MedDRA Preferred Term",
  prune_0 = TRUE,
  deco = std_deco("MHT01"),
  ...
)

mht01_1_lyt(armvar, lbl_overall, lbl_mhbodsys, lbl_mhdecod, deco, ...)

mht01_1_pre(adam_db, ...)

mht01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_mhbodsys}{(\code{string}) text label for \code{MHBODSYS}. If \code{NULL}, the value of the argument defaults to the label
of the corresponding column in the \code{admh} table.}

\item{lbl_mhdecod}{(\code{string}) text label for \code{MHDECOD}. If \code{NULL}, the value of the argument defaults to the label of
the corresponding column in the \code{admh} table.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}
}
\description{
The \code{MHT01} table provides an overview of the subjects medical
history by SOC and Preferred Term.
}
\details{
\itemize{
\item Numbers represent absolute numbers of patients and fraction of \code{N}, or absolute number of event when specified.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm.
\item Does not include a total column by default.
\item Order by body system alphabetically and within body system and medical condition by decreasing total number of
patients with the specific condition.
}
}
\section{Functions}{
\itemize{
\item \code{mht01_1_main()}: Main TLG function

\item \code{mht01_1_lyt()}: Layout

\item \code{mht01_1_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{admh} table with columns \code{"MHBODSYS"} and \code{"MHDECOD"}.
}
}
\examples{
run(mht01_1, syn_test_data())
}
\keyword{datasets}
