% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng01.R
\name{mng01_1}
\alias{mng01_1}
\title{\code{MNG01} Graph 1 (Default) Mean Plot 1.}
\usage{
mng01_1(
  adam_db,
  dataset = "adlb",
  xval = "AVISIT",
  yval = "AVAL",
  armvar = .study$actualarm,
  center_fun = "mean",
  error_fun = "sd",
  jitter = TRUE,
  show_n = FALSE,
  show_h_grid = .study$show_h_grid,
  show_v_grid = .study$show_v_grid,
  legend_pos = .study$legend_pos,
  .study = list(actualarm = "ARM", show_h_grid = TRUE, show_v_grid = FALSE, legend_pos
    = "top")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{dataset}{(\code{character}) the name of the data set to be analyzed.}

\item{xval}{(\code{character}) the name of the variable to be represented on the x-axis.}

\item{yval}{(\code{character}) the name of the variable to be represented on the y-axis.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{center_fun}{(\code{character}) the function defining the value to be displayed.}

\item{error_fun}{(\code{character}) the function defining the error range.}

\item{jitter}{(\code{logical}) should data point be slightly spread on the x-axis.}

\item{show_n}{(\code{logical}) should the number of observation be displayed.}

\item{show_h_grid}{(\code{logical}) should horizontal grid be displayed.}

\item{show_v_grid}{(\code{logical}) should vertical grid be displayed.}

\item{legend_pos}{(\code{character}) the position of the legend. One of \code{top}, \code{bottom}, \code{right}, \code{left} or \code{none}.}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_safety_var}{(\code{character}) the labels of the safety variables to be summarized.}
}
\value{

}
\description{
Overview of a summary statistics across time and arm for a selected parameter and data set.
}
\details{
\itemize{
\item No overall value.
}
}
\examples{
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
  dm_zoom_to("adlb") \%>\%
  filter(PARAM == "Immunoglobulin A Measurement") \%>\%
  dm_update_zoomed()

db \%>\% mng01_1(error_fun = "sd", show_n = FALSE)
db \%>\% mng01_1(error_fun = "se", show_h_grid = FALSE, show_v_grid = FALSE, jitter = FALSE, show_n = TRUE)
db \%>\% mng01_1(error_fun = "ci95", legend_pos = "right")
db \%>\% mng01_1(center_fun = "median", error_fun = "IQR")
}
