% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng01.R
\docType{data}
\name{mng01_1_main}
\alias{mng01_1_main}
\alias{mng01_1_plot}
\alias{mng01_1_pre}
\alias{mng01_1_post}
\alias{mng01_1}
\title{\code{MNG01} Graph 1 (Default) Mean Plot 1.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
mng01_1_main(
  adam_db,
  plot_fun = mng01_1_plot,
  dataset = "adlb",
  x = "AVISIT",
  y = "AVAL",
  y_name = "PARAM",
  y_unit = NA,
  armvar = .study$actualarm,
  center_fun = c("mean", "median"),
  interval_fun = c("mean_ci", "mean_sei", "mean_sdi", "median_ci", "quantiles", "range"),
  show_table = TRUE,
  jitter = TRUE,
  show_n = TRUE,
  show_h_grid = .study$show_h_grid,
  show_v_grid = .study$show_v_grid,
  legend_pos = .study$legend_pos,
  line_col = .study$color_dict,
  .study = list(actualarm = "ACTARM", show_h_grid = TRUE, show_v_grid = FALSE, legend_pos
    = "top", color_dict = nestcolor::color_palette()),
  ...
)

mng01_1_plot(
  df,
  alt_count,
  x = "AVISIT",
  y = "AVAL",
  y_name = "PARAM",
  y_unit = "AVALU",
  armvar = .study$actualarm,
  center_fun = c("mean", "median"),
  interval_fun = c("mean_ci", "mean_sei", "mean_sdi", "median_ci", "quantiles", "range"),
  show_table = TRUE,
  jitter = TRUE,
  show_n = TRUE,
  show_h_grid = .study$show_h_grid,
  show_v_grid = .study$show_v_grid,
  legend_pos = .study$legend_pos,
  line_col = .study$color_dict,
  .study = list(actualarm = "ACTARM", show_h_grid = TRUE, show_v_grid = FALSE, legend_pos
    = "top", color_dict = getOption("tern.color"))
)

mng01_1_pre(adam_db, dataset, x = "AVISIT", ...)

mng01_1_post(tlg, ...)

mng01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{plot_fun}{(\code{function} or \verb{list of functions}) creating a \code{ggplot} object.}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{x}{(\code{character}) the name of a column in the \code{dataset} to represent on the x-axis.}

\item{y}{(\code{string}) the name of the variable to be represented on the y-axis.}

\item{y_name}{(\code{string}) the variable name for \code{y}. Used for plot's subtitle.}

\item{y_unit}{(\code{string}) the name of the variable with the units of \code{y}. Used for plot's subtitle. if \code{NA}, only
\code{y_name} is displayed as subtitle.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{center_fun}{(\code{string}) the function to compute the estimate value.}

\item{interval_fun}{(\code{string}) the function defining the crossbar range.}

\item{show_table}{(\code{flag}) should the summary statistic table be displayed.}

\item{jitter}{(\code{flag}) should data point be slightly spread on the x-axis.}

\item{show_n}{(\code{flag}) should the number of observation be displayed int the table.}

\item{show_h_grid}{(\code{flag}) should horizontal grid be displayed.}

\item{show_v_grid}{(\code{flag}) should vertical grid be displayed.}

\item{legend_pos}{(\code{string}) the position of the legend.}

\item{line_col}{(\code{character}) describing the colors to use for the lines or a named \code{character} vector associating
values of \code{armvar} with color names.}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}

\item{df}{(\code{dataframe}) data set containing all analysis variables.}

\item{alt_count}{(\code{dataframe}) data set that will be used (only) to counts objects in strata.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
Overview of a summary statistics across time and arm for a selected data set.
}
\details{
\itemize{
\item No overall value.
\item Preprocessing filters for \code{ANL01FL} in the selected data set.
}
}
\section{Functions}{
\itemize{
\item \code{mng01_1_main()}: Main TLG Function

\item \code{mng01_1_plot()}: Plot Function

\item \code{mng01_1_pre()}: Preprocessing

\item \code{mng01_1_post()}: Postprocessing

}}
\examples{
library(dm)
library(dplyr)

col <- c(
  "A: Drug X" = "black",
  "B: Placebo" = "blue",
  "C: Combination" = "grey"
)

run(mng01_1, syn_test_data(), dataset = "adlb", center_fun = "median", x = c("AVISIT", "AVISITN"), line_col = col)
}
\keyword{datasets}
