% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_manipulation.R
\name{mutate_adsl_gp}
\alias{mutate_adsl_gp}
\title{Categorize Reason for Discontinuation from Study.}
\usage{
mutate_adsl_gp(
  x,
  reason = .study$disc_reason_var,
  .study = list(disc_reason_var = "DCSREAS")
)
}
\arguments{
\item{x}{(\code{dm}) object.}

\item{reason}{(\code{character}) the variable name containg the reason for discontinuation.}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{

}
\description{
Categorize Reason for Discontinuation from Study.
}
