% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{mutate_for_aet01}
\alias{mutate_for_aet01}
\title{Creating Necessary Columns for \code{aet01}}
\usage{
mutate_for_aet01(
  adam_db,
  actarm = .study$armvar,
  .study = list(armvar = "ARM")
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\description{
Creating Necessary Columns for \code{aet01}
}
