% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdl01.R
\docType{data}
\name{pdl01_main}
\alias{pdl01_main}
\alias{pdl01_pre}
\alias{pdl01}
\title{\code{PDL01} Listing 1 (Default) Major Protocol Deviations.}
\format{
An object of class \code{chevron_l} of length 1.
}
\usage{
pdl01_main(
  adam_db,
  dataset = "addv",
  arm_var = "ARM",
  disp_cols = c("ID", arm_var, "DVDECOD", "DVTERM", "DVSTDTC"),
  ...
)

pdl01_pre(adam_db, dataset = "addv", arm_var = "ARM", ...)

pdl01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{character}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{disp_cols}{(\code{character}) names of non-key columns which should be displayed when the listing is rendered.}

\item{...}{not used.}

\item{default_formatting}{(\code{list}) the default format of the listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}

\item{col_formatting}{(\code{list}) the format of specific listing columns. See \code{\link[rlistings:listings]{rlistings::as_listing}}.}
}
\description{
\code{PDL01} Listing 1 (Default) Major Protocol Deviations.
}
\section{Functions}{
\itemize{
\item \code{pdl01_main()}: Main TLG function

\item \code{pdl01_pre()}: Preprocessing

}}
\examples{
proc_data <- syn_data
proc_data$addv$DVSTDTC <- as.character(proc_data$addv$ASTDT)
res <- run(pdl01, proc_data)
}
\keyword{datasets}
