% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{postprocess}
\alias{postprocess}
\alias{postprocess,chevron_tlg-method}
\alias{postprocess<-}
\alias{postprocess<-,chevron_tlg-method}
\title{Retrieve Postprocess}
\usage{
postprocess(x)

\S4method{postprocess}{chevron_tlg}(x)

postprocess(x) <- value

\S4method{postprocess}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{value}{(\code{function}) returning a post-processed \code{tlg}.}
}
\description{
Retrieve Postprocess

Set Postprocess Function
}
