% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess,chevron_tlg-method}
\alias{preprocess<-}
\alias{preprocess<-,chevron_tlg-method}
\title{Preprocess}
\usage{
preprocess(x)

\S4method{preprocess}{chevron_tlg}(x)

preprocess(x) <- value

\S4method{preprocess}{chevron_tlg}(x) <- value
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{value}{(\code{function}) returning a pre-processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{_pre} function of \code{chevron}.}
}
\description{
retrieve or set \code{preprocess} function.
}
