% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtables_utils.R
\name{proportion_lyt}
\alias{proportion_lyt}
\title{Proportion layout}
\usage{
proportion_lyt(
  lyt,
  arm_var,
  methods,
  strata,
  conf_level,
  odds_ratio = TRUE,
  rsp_var = "IS_RSP"
)
}
\arguments{
\item{lyt}{layout created by \code{rtables}}
}
\description{
Proportion layout
}
\keyword{internal}
