% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_null.R
\docType{data}
\name{report_null}
\alias{report_null}
\alias{report_null,NULL-method}
\alias{report_null,VTableTree-method}
\alias{report_null,listing_df-method}
\alias{report_null,list-method}
\alias{report_null,ANY-method}
\alias{null_report}
\title{Creates \code{NULL} Report}
\format{
An object of class \code{ElementaryTable} with 1 rows and 1 columns.
}
\usage{
report_null(tlg, ...)

\S4method{report_null}{NULL}(tlg, ind = 2L, ...)

\S4method{report_null}{VTableTree}(tlg, ind = 2L, ...)

\S4method{report_null}{listing_df}(tlg, ind = 2L, ...)

\S4method{report_null}{list}(tlg, ind = 2L, ...)

\S4method{report_null}{ANY}(tlg, ...)

null_report
}
\arguments{
\item{tlg}{to convert to null report.}

\item{...}{not used.}

\item{ind}{(\code{integer}) indentation for the outputs of class \code{VTableTree}.}
}
\value{
the \code{tlg} object or a \code{NULL} report if the \code{tlg} is \code{NULL}, is a \code{TableTree} with 0 rows, is a \code{listing_df}
with 0 rows or is a \code{list} with 0 elements.
}
\description{
Creates \code{NULL} Report
}
\examples{
report_null(NULL)
}
\keyword{datasets}
