% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_null}
\alias{report_null}
\title{Create a Null Report}
\usage{
report_null(tlg, ...)
}
\arguments{
\item{tlg}{(\code{TableTree}) object.}

\item{...}{not used.}
}
\value{
original \code{TableTree} or a null report if no observation are found in the table.
}
\description{
Create a Null Report
}
