% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{report_null}
\alias{report_null}
\alias{null_report}
\title{Create a Null Report}
\format{
An object of class \code{TableTree} with 1 rows and 1 columns.
}
\usage{
report_null(tlg)

null_report
}
\arguments{
\item{tlg}{(\code{TableTree}) object.}
}
\value{
original \code{TableTree} or a null report if no observation are found in the table.
}
\description{
Create a Null Report
}
\keyword{datasets}
