% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rl_list}
\alias{rl_list}
\title{List of \code{rlistings} object}
\usage{
rl_list(...)
}
\arguments{
\item{...}{(\code{rlistings}) objects.}
}
\value{
a \code{rl_list} object.
}
\description{
List of \code{rlistings} object
}
