% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt01.R
\docType{data}
\name{rmpt01_1_main}
\alias{rmpt01_1_main}
\alias{rmpt01_1_lyt}
\alias{rmpt01_1_pre}
\alias{rmpt01_1_check}
\alias{rmpt01_1_post}
\alias{rmpt01_1}
\title{\code{RMPT01} Table 1 (Default) Duration of Exposure for Risk Management Plan Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt01_1_main(
  adam_db,
  anl_vars = c("AVALCAT1", "AVAL"),
  lbl_vars = c("Number of Patients", "Duration of Exposure"),
  deco = std_deco("RMPT01"),
  ...
)

rmpt01_1_lyt(anl_vars, lbl_vars, deco)

rmpt01_1_pre(adam_db, ...)

rmpt01_1_check(adam_db, arm_var, req_tables = c("adsl", "adex"))

rmpt01_1_post(tlg, prune_0 = FALSE, ...)

rmpt01_1
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{req_tables}{(\code{character}) names of the required tables.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}

\item{lbl_aval}{(\code{string}) the label of the time variable.}
}
\description{
The \code{RMPT01} table provides an overview of duration of exposure for SE patients
}
\details{
\itemize{
\item Patient time is the sum of exposure across all patients in unit: days, months, or years (days only at this time).
\item Summary statistics are by default based on the number of patients in the corresponding \code{N} row
(number of non-missing values).
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt01_1_main()}: Main TLG function

\item \code{rmpt01_1_lyt()}: Layout

\item \code{rmpt01_1_pre()}: Preprocessing

\item \code{rmpt01_1_check()}: Checks

\item \code{rmpt01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adex} table with the \code{"PARAMCD"}, \code{"AVAL"} and \code{"AVALU"} columns.
}
}
\examples{
library(dplyr)

set.seed(1)
proc_data <- syn_data
proc_data$adex <- proc_data$adex \%>\%
  group_by(USUBJID) \%>\%
  mutate(
    id = seq_along(AVAL),
    AVAL = sample(x = seq(1, 200), size = n(), replace = TRUE)
  ) \%>\%
  ungroup() \%>\%
  mutate(
    PARAMCD = factor(case_when(id == 1 ~ "TDURD", TRUE ~ .data$PARAMCD)),
    AVALU = factor(case_when(PARAMCD == "TDURD" ~ "DAYS", TRUE ~ .data$AVALU)),
    AVALCAT1 = ""
  ) \%>\%
  filter(.data$PARAMCD == "TDURD")

run(rmpt01_1, proc_data)
}
\keyword{datasets}
