% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt01.R
\docType{data}
\name{rmpt01_1_main}
\alias{rmpt01_1_main}
\alias{rmpt01_1_lyt}
\alias{rmpt01_1_pre}
\alias{rmpt01_1_post}
\alias{rmpt01_1}
\title{\code{RMPT01} Table 1 (Default) Duration of Exposure for Risk Management Plan Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt01_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  lbl_aval = "Duration of exposure",
  prune_0 = FALSE,
  deco = std_deco("RMPT01"),
  ...
)

rmpt01_1_lyt(arm_var, lbl_overall, lbl_aval, deco)

rmpt01_1_pre(adam_db, ...)

rmpt01_1_post(tlg, prune_0 = FALSE, ...)

rmpt01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{lbl_aval}{(\code{string}) the label of the time variable.}

\item{prune_0}{(\code{flag}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
Duration of Exposure: Safety-Evaluable Patients
}
\details{
\itemize{
\item Person time is the sum of exposure across all patients in unit: <days, months, or years>.
\item Summary statistics are by default based on the number of patients in the
corresponding \code{N} row (number of non-missing values).
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt01_1_main()}: Main TLG function

\item \code{rmpt01_1_lyt()}: Layout

\item \code{rmpt01_1_pre()}: Preprocessing

\item \code{rmpt01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{adex} table with the \code{"AVAL"}, \code{"AVALCAT1"},
"PARAMCD"\verb{, }PARCAT2\verb{, and }SAFFL` columns.
}
}
\examples{
run(rmpt01_1, syn_data)
}
\keyword{datasets}
