% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt01.R
\name{rmpt01_lyt}
\alias{rmpt01_lyt}
\title{\code{rmpt01} Layout}
\usage{
rmpt01_lyt(
  summaryvars,
  lbl_summaryvars,
  show_tot,
  split_var,
  col_split_var,
  overall_col_lbl
)
}
\arguments{
\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adex}
table of \code{adam_db} is used as label.}

\item{lbl_summaryvars}{(\code{character}) label associated with the analyzed variables.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{character}) recursive split row variable.}

\item{col_split_var}{(\code{string}) additional column splitting variable.}

\item{overall_col_lbl}{(\code{string}) name of the overall column. If \code{NULL}, no overall level is added.}
}
\description{
\code{rmpt01} Layout
}
\keyword{internal}
