% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt01.R
\name{rmpt01_lyt}
\alias{rmpt01_lyt}
\title{\code{rmpt01} Layout}
\usage{
rmpt01_lyt(summaryvars, split_var)
}
\arguments{
\item{summaryvars}{(\code{character}) variables to be analyzed. The label attribute of the corresponding columns in \code{adex}
table of \code{adam_db} is used as label.}

\item{split_var}{(\code{character}) recursive split row variable.}

\item{lbl_summaryvars}{(\code{character}) label associated with the analyzed variables.}

\item{lbl_parcat}{(\code{string}) the label associated with the variable initiating a new row split.}
}
\description{
\code{rmpt01} Layout
}
\keyword{internal}
