% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt03.R
\name{rmpt03_main}
\alias{rmpt03_main}
\alias{rmpt03_pre}
\alias{rmpt03}
\title{\code{rmpt03}Duration of Exposure for Risk Management Plan Table.}
\usage{
rmpt03_main(
  adam_db,
  summaryvars = "AGEGR1",
  show_tot = TRUE,
  row_split_var = NULL,
  col_split_var = "SEX",
  overall_col_lbl = "All Genders",
  ...
)

rmpt03_pre(adam_db, summaryvars = "AGEGR1", ...)

rmpt03
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adex}
table of \code{adam_db} is used as label.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{row_split_var}{(\code{string}) the name of the column that containing variable to split exposure by.}

\item{col_split_var}{(\code{string}) additional column splitting variable.}

\item{overall_col_lbl}{(\code{string}) name of the overall column. If \code{NULL}, no overall level is added.}

\item{...}{not used.}
}
\value{
the main function returns an \code{rtables} object.

the preprocessing function returns a \code{list} of \code{data.frame}.
}
\description{
The \code{rmpt03} table provides an overview of duration of exposure.
}
\details{
\itemize{
\item Person time is the sum of exposure across all patients.
\item Summary statistics are by default based on the number of patients in the corresponding \code{N} row
(number of non-missing values).
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt03_main()}: Main TLG function

\item \code{rmpt03_pre()}: Preprocessing

}}
\examples{
pre_data <- dunlin::propagate(syn_data, "adsl", "AGEGR1", "USUBJID")
run(rmpt03, pre_data)
}
