% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt03.R
\docType{data}
\name{rmpt03_main}
\alias{rmpt03_main}
\alias{rmpt03_pre}
\alias{rmpt03_post}
\alias{rmpt03}
\title{\code{rmpt03}Duration of Exposure for Risk Management Plan Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt03_main(
  adam_db,
  summaryvars = "AGEGR1",
  show_tot = TRUE,
  split_var = NULL,
  col_split_var = "SEX",
  col_split_fun = function (df, spl, vals = NULL, labels = NULL, ...) 
 {
     if
    (is(spl, "MultiVarSplit")) 
        
    stop("Combining levels of a MultiVarSplit does not make sense.", call. = FALSE)
    
    ret <- .apply_split_inner(spl, df, vals = vals, labels = labels, trim = trim)
    
    for (i in seq_len(nrow(combosdf))) {
         lcombo <- combosdf[i, "levelcombo",
    drop = TRUE][[1]]
         spld <- spl_payload(spl)
         if (is(lcombo,
    "AllLevelsSentinel")) 
             subdf <- df
         else if (is(spl,
    "VarLevelSplit")) {
             subdf <- df[df[[spld]] \%in\% lcombo, ]
         }
 
           else {
             stopifnot(all(lcombo \%in\% c(ret$labels, ret$vals)))
    
            subdf <- do.call(rbind, ret$datasplit[names(ret$datasplit) \%in\% lcombo |
    ret$vals \%in\% lcombo])
         }
         ret <- .add_combo_part_info(ret, subdf,
    combosdf[i, "valname", drop = TRUE], lcombo, combosdf[i, "label", drop = TRUE],
    combosdf[i, "exargs", drop = TRUE][[1]], first)
     }
     if
    (!is.null(keep_levels)) {
         keep_inds <- value_names(ret$values) \%in\%
    keep_levels
         ret <- lapply(ret, function(x) x[keep_inds])
     }
     ret
 },
  ...
)

rmpt03_pre(adam_db, ...)

rmpt03_post(tlg, prune_0 = FALSE, ...)

rmpt03
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adex}
table of \code{adam_db} is used as label.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{string}) the name of the column that containing variable to split exposure by.}

\item{col_split_var}{(\code{character}) additional column splitting variables.}

\item{col_split_fun}{(\code{function} or \code{NULL}) function for column split.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{rmpt03} table provides an overview of duration of exposure.
}
\details{
\itemize{
\item Person time is the sum of exposure across all patients.
\item Summary statistics are by default based on the number of patients in the corresponding \code{N} row
(number of non-missing values).
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt03_main()}: Main TLG function

\item \code{rmpt03_pre()}: Preprocessing

\item \code{rmpt03_post()}: Postprocessing

}}
\examples{
pre_data <- dunlin::propagate(syn_data, "adsl", "AGEGR1", "USUBJID")
run(rmpt03, pre_data)
}
\keyword{datasets}
