% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt03.R
\docType{data}
\name{rmpt03_main}
\alias{rmpt03_main}
\alias{rmpt03_pre}
\alias{rmpt03_post}
\alias{rmpt03}
\title{\code{rmpt03}Duration of Exposure for Risk Management Plan Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt03_main(
  adam_db,
  summaryvars = "AGEGR1",
  show_tot = TRUE,
  split_var = NULL,
  ...
)

rmpt03_pre(adam_db, parcat = NULL, ...)

rmpt03_post(tlg, prune_0 = FALSE, ...)

rmpt03
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adsl}
table of \code{adam_db} is used as label.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{string}) the name of the column that containing variable to split exposure by.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{rmpt03} table provides an overview of duration of exposure.
}
\details{
\itemize{
\item Person time is the sum of exposure across all patients.
\item Summary statistics are by default based on the number of patients in the corresponding \code{N} row
(number of non-missing values).
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt03_main()}: Main TLG function

\item \code{rmpt03_pre()}: Preprocessing

\item \code{rmpt03_post()}: Postprocessing

}}
\examples{
pre_data <- dunlin::propagate(syn_data, "adsl", "AGEGR1", "USUBJID")
run(rmpt03, pre_data)
}
\keyword{datasets}
