% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt04.R
\docType{data}
\name{rmpt04_main}
\alias{rmpt04_main}
\alias{rmpt04_pre}
\alias{rmpt04_post}
\alias{rmpt04}
\title{\code{RMPT04}Extent of Exposure by Ethnic Origin for Risk Management Plan Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt04_main(adam_db, summaryvars = "ETHNIC", show_tot = TRUE, ...)

rmpt04_pre(adam_db, ...)

rmpt04_post(tlg, prune_0 = TRUE, ...)

rmpt04
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{summaryvars}{(\code{character}) variables to be analyzed. The label attribute of the corresponding columns in \code{adsl}
table of \code{adam_db} is used as label.}

\item{show_tot}{(\code{flag}) whether to show the total number of patients and patient time.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{RMPT04} table provides an overview of duration of exposure extent.
}
\section{Functions}{
\itemize{
\item \code{rmpt04_main()}: Main TLG function

\item \code{rmpt04_pre()}: Preprocessing

\item \code{rmpt04_post()}: Postprocessing

}}
\examples{
run(rmpt04, syn_data, show_tot = FALSE)
}
\keyword{datasets}
