% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt05.R
\docType{data}
\name{rmpt05_main}
\alias{rmpt05_main}
\alias{rmpt05_pre}
\alias{rmpt05_post}
\alias{rmpt05}
\title{\code{RMPT05} Extent of Exposure by Race for Risk Management Plan Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt05_main(
  adam_db,
  summaryvars = "RACE",
  show_tot = TRUE,
  split_var = NULL,
  ...
)

rmpt05_pre(adam_db, ...)

rmpt05_post(tlg, prune_0 = FALSE, ...)

rmpt05
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adsl}
table of \code{adam_db} is used as label.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{character}) recursive split row variable.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
The \code{RMPT05} table provides an overview of duration of exposure extent.
}
\section{Functions}{
\itemize{
\item \code{rmpt05_main()}: Main TLG function

\item \code{rmpt05_pre()}: Preprocessing

\item \code{rmpt05_post()}: Postprocessing

}}
\examples{
run(rmpt05, syn_data)
}
\keyword{datasets}
