% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt05.R
\docType{data}
\name{rmpt05_main}
\alias{rmpt05_main}
\alias{rmpt05}
\title{\code{RMPT05} Extent of Exposure by Race for Risk Management Plan Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt05_main(
  adam_db,
  summaryvars = "RACE",
  show_tot = TRUE,
  split_var = NULL,
  col_split_var = NULL,
  overall_col_lbl = NULL,
  ...
)

rmpt05
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{summaryvars}{(\code{string}) variables to be analyzed. The label attribute of the corresponding columns in \code{adex}
table of \code{adam_db} is used as label.}

\item{show_tot}{(\code{flag}) whether to display the cumulative total.}

\item{split_var}{(\code{character}) recursive split row variable.}

\item{col_split_var}{(\code{character}) additional column splitting variables.}

\item{overall_col_lbl}{(\code{string}) name of the overall column.}

\item{...}{not used.}
}
\description{
The \code{RMPT05} table provides an overview of duration of exposure extent.
}
\details{
\itemize{
\item Person time is the sum of exposure across all patients.
\item Summary statistics are by default based on the number of patients in the corresponding \code{N} row
(number of non-missing values).
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{rmpt05_main()}: Main TLG function

}}
\examples{
run(rmpt05, syn_data)
}
\keyword{datasets}
