% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpt06.R
\docType{data}
\name{rmpt06_main}
\alias{rmpt06_main}
\alias{rmpt06_pre}
\alias{rmpt06_post}
\alias{rmpt06}
\title{\code{RMPT06} Table 1 (Default) Seriousness, Outcomes, Severity, Frequency with 95\% CI for Risk Management Plan.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rmpt06_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  aesi_vars = NULL,
  grade_groups = NULL,
  ...
)

rmpt06_pre(adam_db, ...)

rmpt06_post(tlg, prune_0 = FALSE, ...)

rmpt06
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
\code{RMPT06} Table 1 (Default) Seriousness, Outcomes, Severity, Frequency with 95\% CI for Risk Management Plan.
}
\section{Functions}{
\itemize{
\item \code{rmpt06_main()}: Main TLG function

\item \code{rmpt06_pre()}: Preprocessing

\item \code{rmpt06_post()}: Postprocessing

}}
\examples{
run(rmpt06, syn_data)
}
\keyword{datasets}
