% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspt01.R
\docType{data}
\name{rspt01_1_main}
\alias{rspt01_1_main}
\alias{rspt01_1_lyt}
\alias{rspt01_1_pre}
\alias{rspt01_1_post}
\alias{rspt01_1}
\alias{proportion_lyt}
\title{\code{RSPT01} Binary Outcomes Summary}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
rspt01_1_main(
  adam_db,
  dataset = "adrs",
  arm_var = "ARM",
  ref_group = NULL,
  odds_ratio = TRUE,
  strat_analysis = FALSE,
  strata = NULL,
  conf_level = 0.95,
  methods = list(),
  deco = std_deco("RSPT01"),
  ...
)

rspt01_1_lyt(
  arm_var,
  ref_group,
  odds_ratio,
  strat_analysis,
  strata,
  conf_level,
  methods,
  deco
)

rspt01_1_pre(
  adam_db,
  arm_var = "ARM",
  dataset = "adrs",
  responder = c("CR", "PR"),
  ...
)

rspt01_1_post(tlg, prune_0 = TRUE, ...)

rspt01_1

proportion_lyt(
  lyt,
  arm_var,
  methods,
  strata,
  conf_level,
  odds_ratio = TRUE,
  strat_analysis = FALSE
)
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{ref_group}{(\code{string}) The name of the reference group, the value should
be identical to the values in \code{arm_var}, if not speficied, it will by default
use the first level or value of \code{arm_var}.}

\item{odds_ratio}{(\code{flag}) should the odds ratio be calculated, default is TRUE}

\item{strat_analysis}{(\code{flag}) should the stratified analysis be performed,}

\item{strata}{(\code{string}) stratification factors, e.g. strata = c("STRATA1", "STRATA2"), by default as NULL}

\item{conf_level}{(\code{numeric}) the level of confidence interval, default is 0.95.}

\item{methods}{(\code{list}) a named list, use a named list to control, for example:
methods = list(prop_conf_method = "wald",
diff_conf_method = "wald",
strat_diff_conf_method = "ha",
diff_pval_method = "fisher",
strat_diff_pval_method = "schouten")
see more details in \code{tern}}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{responder}{(\code{string}) responder defined in the table, by default use c("CR", "PR").}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
RSPT01 template may be used to summarize any binary outcome or response variable at
a single time point. Typical application for oncology
}
\details{
\itemize{
\item No overall value.
}
}
\section{Functions}{
\itemize{
\item \code{rspt01_1_main()}: Main TLG function

\item \code{rspt01_1_lyt()}: Layout

\item \code{rspt01_1_pre()}: Preprocessing

\item \code{rspt01_1_post()}: Postprocessing

\item \code{proportion_lyt()}: get proportion layout

}}
\examples{
rspt01_1_pre(syn_data)
library(dplyr)
library(dunlin)

syn_data2 <- log_filter(syn_data, PARAMCD == "BESRSPI", "adrs")
run(rspt01_1, syn_data2)
run(rspt01_1, syn_data2, odds_ratio = FALSE, strat_analysis = TRUE, strata = c("STRATA1", "STRATA2"),
 methods = list(diff_pval_method = "fisher"))
}
\keyword{datasets}
