% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspt01.R
\name{rspt01_lyt}
\alias{rspt01_lyt}
\title{rspt01 Layout}
\usage{
rspt01_lyt(
  arm_var,
  ref_group,
  odds_ratio,
  perform_analysis,
  strata,
  conf_level,
  methods,
  rsp_var
)
}
\arguments{
\item{arm_var}{(\code{string}) variable used for column splitting}
}
\description{
rspt01 Layout
}
\keyword{internal}
