% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{run}
\alias{run}
\alias{run,chevron_tlg-method}
\title{Run the pipeline}
\usage{
run(
  object,
  adam_db,
  auto_pre = TRUE,
  verbose = FALSE,
  ...,
  user_args = list(...)
)

\S4method{run}{chevron_tlg}(
  object,
  adam_db,
  auto_pre = TRUE,
  verbose = FALSE,
  unwrap = FALSE,
  ...,
  user_args = list(...)
)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{auto_pre}{(\code{flag}) whether to perform the default pre processing step.}

\item{verbose}{(\code{flag}) whether to print argument information.}

\item{...}{extra arguments to pass to the pre-processing, main and post-processing functions.}

\item{user_args}{(\code{list}) arguments from \code{...}.}
}
\value{
an \code{rtables} (for \code{chevron_t}), \code{rlistings} (for \code{chevron_l}), \code{grob} (for \code{chevron_g}) or \code{ElementaryTable}
(null report) depending on the class of \code{chevron_tlg} object passed as \code{object} argument.
}
\description{
Run the pipeline
}
\examples{
run(mng01, syn_data, auto_pre = TRUE, dataset = "adlb")
}
