% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{run}
\alias{run}
\alias{run,chevron_tlg-method}
\title{run}
\usage{
run(object, adam_db, ...)

\S4method{run}{chevron_tlg}(object, adam_db, ...)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{...}{extra arguments to pass to the check, pre-processing or \code{tlg} functions.}
}
\description{
run
}
\examples{
run(aet01_1, syn_test_data())

}
