% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R,
%   R/chevron_t-S4methods.R
\name{run}
\alias{run}
\alias{run,chevron_tlg-method}
\alias{run,chevron_t-method}
\title{Run the pipeline}
\usage{
run(object, adam_db, ...)

\S4method{run}{chevron_tlg}(object, adam_db, ...)

\S4method{run}{chevron_t}(object, adam_db, ...)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{...}{extra arguments to pass to the check, pre-processing or \code{tlg} functions.}
}
\description{
Run the pipeline
}
\examples{
run(mng01_1, syn_test_data(), dataset = "adlb")
run(aet01_1, syn_test_data())
}
