% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{run}
\alias{run}
\alias{run,chevron_tlg-method}
\title{Run the pipeline}
\usage{
run(object, adam_db, auto_pre = TRUE, ...)

\S4method{run}{chevron_tlg}(object, adam_db, auto_pre = TRUE, ...)
}
\arguments{
\item{object}{(\code{chevron_tlg}) input.}

\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{auto_pre}{(\code{flag}) whether to perform the default pre processing step.}

\item{...}{extra arguments to pass to the pre-processing, main and post-processing functions.}
}
\description{
Run the pipeline
}
\examples{
run(mng01_1, syn_data, auto_pre = TRUE, dataset = "adlb")
}
