% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_study_object.R
\name{sample_study_object}
\alias{sample_study_object}
\title{Extraction of .study object default values}
\usage{
sample_study_object()
}
\value{
(\code{data.frame})
}
\description{
Get the names and the values of the default arguments stored in the study object of each exported function of
chevron.
}
\examples{
sample_study_object()
}
