% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{script_funs}
\alias{script_funs}
\alias{script_funs,chevron_tlg-method}
\title{Create Script for \code{TLG} Generation}
\usage{
script_funs(x, adam_db, args, details = FALSE)

\S4method{script_funs}{chevron_tlg}(x, adam_db, args, details = FALSE)
}
\description{
Create Script for \code{TLG} Generation
}
\examples{
script_funs(aet04, adam_db = "syn_data", args = "args")
}
