% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{script}
\alias{script}
\alias{script,chevron_tlg-method}
\title{Create Pre Processing Script}
\usage{
script(x, con = stdout())

\S4method{script}{chevron_tlg}(x, con = stdout())
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}

\item{con}{(\code{connection}) where the output is saved.}
}
\description{
Create Pre Processing Script
}
