% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevron_tlg-S4methods.R
\name{script}
\alias{script}
\alias{script,chevron_tlg-method}
\title{Create Pre Processing Script}
\usage{
script(x)

\S4method{script}{chevron_tlg}(x)
}
\arguments{
\item{x}{(\code{chevron_tlg}) input.}
}
\description{
Create Pre Processing Script
}
