% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_decoration}
\alias{set_decoration}
\title{Temporary function to add decorator to a table}
\usage{
set_decoration(x, deco)
}
\arguments{
\item{x}{(\code{rtables}) object.}

\item{deco}{(\code{list}) typically generated with \code{std_deco()}.}
}
\value{
\code{rtables} with set title, subtitle and footnotes. If one of this attribute is NULL, the slot is empty.
}
\description{
Temporary function to add decorator to a table
}
\keyword{internal}
