% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_deco}
\alias{std_deco}
\title{Standard Documentation Lookup}
\usage{
std_deco(id, ...)
}
\arguments{
\item{id}{standard id of output}

\item{...}{not used at the moment}
}
\description{
Standard Documentation Lookup
}
