% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{std_filter_fun}
\alias{std_filter_fun}
\title{Retrieve Standard Subsetting for Templates}
\usage{
std_filter_fun(tlgfname, pmap = std_pmap)
}
\arguments{
\item{tlgfname}{(\code{character}) function name of tlg-function}

\item{pmap}{('data frame') preprocessing map, see \code{std_pmap} for the required format}
}
\description{
Retrieve Standard Subsetting for Templates
}
\examples{
std_filter_fun("aet02_1")
}
