% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_data_preprocessing.R
\name{std_pmap}
\alias{std_pmap}
\title{Standard Preprocessing Map}
\usage{
std_pmap()
}
\description{
The preprocessing map contains information how the ADaM data needs to be preprocessed for each function.
}
\examples{
std_pmap()
}
