% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_postprocess}
\alias{std_postprocess}
\title{Standard Post processing}
\usage{
std_postprocess(tlg, ind = 2L, ...)
}
\arguments{
\item{tlg}{(\code{TableTree}) object.}

\item{ind}{(\code{integer}) the indentation of the table.}

\item{...}{not used at the moment.}
}
\value{
a post-processed \code{tlg}.
}
\description{
Standard Post processing
}
\note{
Standard post processing includes:
\itemize{
\item \code{NULL} report creation if necessary
\item indentation
}
}
\keyword{internal}
