% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syn_data}
\alias{syn_data}
\title{Example \code{adam} Synthetic Data}
\format{
A named \code{list} of 13 \code{data.frames}:
- \code{adsl}
- \code{adae}
- \code{adaette}
- \code{adcm}
- \code{addv}
- \code{adeg}
- \code{adex}
- \code{adlb}
- \code{admh}
- \code{adrs}
- \code{adsub}
- \code{adtte}
- \code{advs}
}
\source{
based on \code{scda::synthetic_cdisc_data("rcd_2022_06_27")}
}
\usage{
syn_data
}
\description{
Example \code{adam} Synthetic Data
}
\keyword{datasets}
