% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLG_pipeline-S4class.R
\docType{class}
\name{tlg_pipeline_s4-class}
\alias{tlg_pipeline_s4-class}
\alias{.tlg_pipeline_s4}
\title{\code{tlg_pipeline_s4} Class}
\description{
The \code{tlg_pipeline_s4} class associates a \code{preprocess} function, a main \code{tlg} function and \code{AdAM} tables names and
optionally a \code{postprocess} function.
}
\section{Slots}{

\describe{
\item{\code{main}}{(\code{function}) returning a \code{tlg}. Typically one of the \verb{_main} function of \code{chevron}.}

\item{\code{preprocess}}{(\code{function}) returning a pre processed \code{dm} object amenable to \code{tlg} creation. Typically one of the
\verb{_pre} function of \code{chevron}.}

\item{\code{postprocess}}{(\code{function}) returning a post processed \code{tlg}.}

\item{\code{adam_datasets}}{(\code{character}) representing the name of the table from an \code{AdAM} data set required for \code{tlg}
creation.}
}}

\note{
To ensure the correct execution of the pipeline workflow additional validation criteria are:
\itemize{
\item the first argument of the \code{main} function must be \code{adam_db}, the input \code{dm} object to preprocess.
\item the first argument of the \code{preprocess} function must be \code{adam_db}, the input \code{dm} object to create \code{tlg} output.
\item the first argument of the \code{postprocess} function must be \code{tlg}, the input \code{TableTree} object to post process.
}
}
