% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttet01.R
\docType{data}
\name{ttet01_main}
\alias{ttet01_main}
\alias{ttet01_lyt}
\alias{ttet01_pre}
\alias{ttet01_post}
\alias{ttet01}
\title{\code{TTET01} Binary Outcomes Summary}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
ttet01_main(
  adam_db,
  dataset = "adtte",
  arm_var = "ARM",
  ref_group = NULL,
  summarize_event = TRUE,
  perform_analysis = c("unstrat"),
  strata = NULL,
  pval_method = "log-rank",
  conf_level = 0.95,
  conf_type = "log-log",
  quantiles = c(0.25, 0.75),
  ties = "efron",
  timepoint = c(6, 12),
  method = "both",
  ...
)

ttet01_lyt(
  arm_var,
  ref_group,
  summarize_event,
  perform_analysis,
  strata,
  pval_method,
  conf_level,
  conf_type,
  quantiles,
  ties,
  timeunit,
  timepoint,
  method
)

ttet01_pre(adam_db, arm_var = "ARM", dataset = "adtte", ...)

ttet01_post(tlg, ...)

ttet01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{ref_group}{(\code{string}) The name of the reference group, the value should
be identical to the values in \code{arm_var}, if not speficied, it will by default
use the first level or value of \code{arm_var}.}

\item{perform_analysis}{(\code{string}) option to display statistical comparisons using stratified analyses,
or unstratified analyses, or both, e.g. c("unstrat", "strat"). Only unstratified will be displayed by default}

\item{conf_level}{(\code{numeric}) the level of confidence interval, default is 0.95.}

\item{...}{not used.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{odds_ratio}{(\code{flag}) should the odds ratio be calculated, default is TRUE}

\item{strat}{(\code{string}) stratification factors, e.g. strat = c("STRATA1", "STRATA2"), by default as NULL}

\item{methods}{(\code{list}) a named list, use a named list to control, for example:
methods = list(prop_conf_method = "wald",
diff_conf_method = "wald",
strat_diff_conf_method = "ha",
diff_pval_method = "fisher",
strat_diff_pval_method = "schouten")
\code{prop_conf_method} controls the methods of calculating proportion confidence interval,
\code{diff_conf_method} controls the methods of calculating unstratified difference confidence interval,
\code{strat_diff_conf_method} controls the methods of calculating stratified difference confidence interval,
\code{diff_pval_method} controls the methods of calculating unstratified p-value for odds ratio,
\code{strat_diff_pval_method} controls the methods of calculating stratified p-value for odds ratio,
see more details in \code{tern}}

\item{responder}{(\code{string}) responder defined in the table, by default use c("CR", "PR").}
}
\description{
TTET01 template may be used to summarize any binary outcome or response variable at
a single time point. Typical application for oncology
}
\details{
\itemize{
\item No overall value.
}
}
\section{Functions}{
\itemize{
\item \code{ttet01_main()}: Main TLG function

\item \code{ttet01_lyt()}: Layout

\item \code{ttet01_pre()}: Preprocessing

\item \code{ttet01_post()}: Postprocessing

}}
\examples{
ttet01_pre(syn_data)
library(dplyr)
library(dunlin)

syn_data2 <- log_filter(syn_data, PARAMCD == "PFS", "adtte")
run(ttet01, syn_data2)
run(ttet01, syn_data2,
  summarize_event = FALSE, perform_analysis = c("unstrat", "strat"),
  strata = c("STRATA1", "STRATA2")
)
}
\keyword{datasets}
