% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttet01.R
\docType{data}
\name{ttet01_main}
\alias{ttet01_main}
\alias{ttet01_lyt}
\alias{ttet01_pre}
\alias{ttet01_post}
\alias{ttet01}
\title{\code{TTET01} Binary Outcomes Summary}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
ttet01_main(
  adam_db,
  dataset = "adtte",
  arm_var = "ARM",
  ref_group = NULL,
  summarize_event = TRUE,
  perform_analysis = c("unstrat"),
  strata = NULL,
  pval_method = "log-rank",
  conf_level = 0.95,
  conf_type = "log-log",
  quantiles = c(0.25, 0.75),
  ties = "efron",
  timepoint = c(6, 12),
  method = "both",
  ...
)

ttet01_lyt(
  arm_var,
  ref_group,
  summarize_event,
  perform_analysis,
  strata,
  pval_method,
  conf_level,
  conf_type,
  quantiles,
  ties,
  timeunit,
  timepoint,
  method
)

ttet01_pre(adam_db, arm_var = "ARM", dataset = "adtte", ...)

ttet01_post(tlg, ...)

ttet01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{ref_group}{(\code{string}) The name of the reference group, the value should
be identical to the values in \code{arm_var}, if not speficied, it will by default
use the first level or value of \code{arm_var}.}

\item{summarize_event}{(\code{flag}) should the event discription be displayed, default is TRUE}

\item{perform_analysis}{(\code{string}) option to display statistical comparisons using stratified analyses,
or unstratified analyses, or both, e.g. c("unstrat", "strat"). Only unstratified will be displayed by default}

\item{strata}{(\code{string}) stratification factors, e.g. strata = c("STRATA1", "STRATA2"), by default as NULL}

\item{pval_method}{(string) p-value method for testing hazard ratio = 1.
Default method is "log-rank", can also be set to "wald" or "likelihood".}

\item{conf_level}{(\code{numeric}) the level of confidence interval, default is 0.95.}

\item{conf_type}{(\code{string}) confidence interval type. Options are "plain" (default), "log", "log-log",
see more in survival::survfit(). Note option "none" is no longer supported.}

\item{quantiles}{(\code{numeric}) of length two to specify the quantiles of survival time.}

\item{ties}{string) specifying the method for tie handling. Default is "efron",
can also be set to "breslow" or "exact". see more in survival::coxph()}

\item{timepoint}{(number) survival time point of interest.}

\item{method}{(string) either surv (survival estimations),
surv_diff (difference in survival with the control) or both.}

\item{...}{not used.}

\item{timeunit}{(string) time uint get from AVALU, by default is "Months"}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}
}
\description{
TTET01 template may be used to summarize any binary outcome or response variable at
a single time point. Typical application for oncology
}
\details{
\itemize{
\item No overall value.
}
}
\section{Functions}{
\itemize{
\item \code{ttet01_main()}: Main TLG function

\item \code{ttet01_lyt()}: Layout

\item \code{ttet01_pre()}: Preprocessing

\item \code{ttet01_post()}: Postprocessing

}}
\examples{
ttet01_pre(syn_data)
library(dplyr)
library(dunlin)

syn_data2 <- log_filter(syn_data, PARAMCD == "PFS", "adtte")
run(ttet01, syn_data2)
run(ttet01, syn_data2,
  summarize_event = FALSE, perform_analysis = c("unstrat", "strat"),
  strata = c("STRATA1", "STRATA2")
)
}
\keyword{datasets}
