% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_labels_for}
\alias{var_labels_for}
\title{Retrieve variables for certain variables}
\usage{
var_labels_for(df, vars)
}
\arguments{
\item{df}{data frame}

\item{vars}{variable names in \code{df}}
}
\description{
Retrieve variables for certain variables
}
