% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_labels_for}
\alias{var_labels_for}
\title{Retrieve Variables for Certain variables}
\usage{
var_labels_for(df, vars)
}
\arguments{
\item{df}{data froma}

\item{vars}{variable names in data frame \code{df}}
}
\description{
Retrieve Variables for Certain variables
}
