% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egt01.R, R/vst01.R
\docType{data}
\name{egt01_pre}
\alias{egt01_pre}
\alias{vst01_main}
\alias{vst01_pre}
\alias{vst01}
\title{\code{VST01} Vital Sign Results and change from Baseline By Visit Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\source{
\code{cfbt01.R}
}
\usage{
egt01_pre(adam_db, dataset = "adeg", ...)

vst01_main(
  adam_db,
  dataset = "advs",
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(default = 2L),
  page_by = TRUE,
  .stats = c("n", "mean_sd", "median", "range"),
  ...
)

vst01_pre(adam_db, dataset = "advs", ...)

vst01
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{dataset}{(\code{string}) the name of a table in the \code{adam_db} object.}

\item{...}{additional arguments like \code{.indent_mods}, \code{.labels}.}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{row_split_var}{(\code{character}) row split variable other than \code{PARAMCD}.}

\item{summaryvars}{(\code{character}) variables to be analyzed. The label attribute of the corresponding column in
table of \code{adam_db} is used as label.}

\item{visitvar}{(\code{string}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed}

\item{precision}{(named \code{list} of \code{integer}) where names are values found in the \code{PARAMCD} column and the the values
indicate the number of digits that should be represented for \code{min}, \code{max} and \code{median}. \code{Mean} and \code{sd} are
represented with one more decimal of precision.}

\item{page_by}{(\code{flag}) indicator whether the parameter row split is by page.}

\item{.stats}{(\code{character}) statistics names, see \code{summarize_vars()}.}
}
\description{
The \code{VST01} table provides an
overview of the Vital Sign values and its change from baseline of each respective arm
over the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{egt01_pre()}: Preprocessing

\item \code{vst01_main()}: Main TLG function

\item \code{vst01_pre()}: Preprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain table named as \code{dataset} with the columns specified in \code{summaryvars}.
}
}
\examples{
run(vst01, syn_data)
}
\keyword{datasets}
