% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst01.R, R/vst02.R
\docType{data}
\name{vst01_1_main}
\alias{vst01_1_main}
\alias{vst01_1_lyt}
\alias{vst01_1_pre}
\alias{vst01_1}
\alias{vst02_1_main}
\title{\code{VST01_1} Pipeline}
\format{
a \code{tlg_pipeline_s4} object with the following slots:
\itemize{
\item \code{main} the \code{chevron::vst01_1_main} function.
\item \code{preprocess} the  \code{chevron::vst01_1_pre} function.
\item \code{postprocess} the identity function.
\item \code{check} no checks.
\item \code{adam_datasets} \code{"adsl"} and \code{"advs"}.
}
}
\usage{
vst01_1_main(
  adam_db,
  armvar = .study$actualarm,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = var_labels_for(adam_db$advs, summaryvars),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("VST01"),
  .study = list(actualarm = "ACTARM", evo_vars = c("AVAL", "CHG"))
)

vst01_1_lyt(
  armvar = .study$actualarm,
  summaryvars = .study$evo_vars,
  summaryvars_lbls = .study$evo_vars_lbls,
  visitvar = .study$visitvar,
  lbl_avisit = "",
  lbl_param = "",
  deco = std_deco("VST01"),
  .study = list(actualarm = "ACTARM", evo_vars = c("AVAL", "CHG"), evo_vars_lbls =
    c("Analysis \\nValue", "Change from \\nBaseline"), visitvar = "AVISIT")
)

vst01_1_pre(adam_db, ...)

vst01_1

vst02_1_main(
  adam_db,
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("VST02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{lbl_avisit}{(\code{character}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}

\item{...}{not used.}

\item{lbl_vs_assessment}{(\code{character}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{character}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}
}
\description{
\code{VST01_1} Pipeline of the class \code{tlg_pipeline_s4}
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}

\itemize{
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{vst01_1_main}: \code{vst01_1} main function

\code{VST01} Table 1 (Default) Vital Sign Results and change from Baseline By Visit Table 1. The \code{VST01} table provides an
overview of the Vital Sign values and its change from baseline of each respective arm
over the course of the trial.

\item \code{vst01_1_lyt}: \code{vst01_1} Layout

\item \code{vst01_1_pre}: \code{vst01_1} Preprocessing

\item \code{vst02_1_main}: \code{vst01_1} main function

\code{VST02} Table 1 (Default) Vital Sign Abnormalities Table 1. Assessments Outside Normal Limits Regardless of
Abnormality at Baseline Table.
}}

\examples{
library(dm)

db <- syn_test_data() \%>\%
  vst01_1_pre()

vst01_1_main(db)
vst01_1_main(db, summaryvars_lbls = c("Value at Visit", "Change from Baseline"))
syn_test_data() \%>\%
  vst01_1_pre()
library(dm)

db <- syn_test_data() \%>\%
  dm_zoom_to("advs") \%>\%
  filter(PARAM \%in\% c("Temperature")) \%>\%
  dm_update_zoomed() \%>\%
  vst02_1_pre()

vst02_1_main(adam_db = db)
}
\keyword{datasets}
