% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst01.R
\docType{data}
\name{vst01_1_main}
\alias{vst01_1_main}
\alias{vst01_1_lyt}
\alias{vst01_1_pre}
\alias{vst01_1_post}
\alias{vst01_1}
\title{\code{VST01} Table 1 (Default) Vital Sign Results and change from Baseline By Visit Table 1.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
vst01_1_main(
  adam_db,
  armvar = "ACTARM",
  summaryvars = c(`Value at Visit` = "AVAL", `Change from \\nBaseline` = "CHG"),
  visitvar = "AVISIT",
  deco = std_deco("VST01"),
  ...
)

vst01_1_lyt(
  armvar,
  summaryvars,
  summaryvars_lbls,
  visitvar,
  lbl_avisit,
  lbl_param,
  deco,
  ...
)

vst01_1_pre(adam_db, ...)

vst01_1_post(tlg, prune_0 = TRUE, ...)

vst01_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{summaryvars}{(\verb{vector of character}) the variables to be analyzed. For this table, \code{AVAL} and \code{CHG} by default.}

\item{visitvar}{(\code{character}) typically one of \code{"AVISIT"} (Default) or \code{"ATPTN"} depending on the type of time point
to be displayed.}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{...}{not used.}

\item{summaryvars_lbls}{(\verb{vector of character}) the label of the variables to be analyzed.}

\item{lbl_avisit}{(\code{character}) label of the \code{visitvar} variable.}

\item{lbl_param}{(\code{character}) label of the \code{PARAM} variable.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{logical}) remove 0 count rows}
}
\description{
The \code{VST01} table provides an
overview of the Vital Sign values and its change from baseline of each respective arm
over the course of the trial.
}
\details{
\itemize{
\item The \verb{Analysis Value} column, displays the number of patients, the mean, standard deviation, median and range of
the analysis value for each visit.
\item The \verb{Change from Baseline} column, displays the number of patient and the mean, standard deviation,
median and range of changes relative to the baseline.
\item Remove zero-count rows unless overridden with \code{prune_0 = FALSE}.
\item Split columns by arm, typically \code{ACTARM}.
\item Does not include a total column by default.
\item Sorted  based on factor level; first by \code{PARAM} labels in alphabetic order then by chronological time point given
by \code{AVISIT}. Re-level to customize order
}
}
\section{Functions}{
\itemize{
\item \code{vst01_1_main()}: Main TLG function

\item \code{vst01_1_lyt()}: Layout

\item \code{vst01_1_pre()}: Preprocessing

\item \code{vst01_1_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{advs} table with the columns specified in \code{summaryvars}.
}
}
\examples{
run(vst01_1, syn_data)
}
\keyword{datasets}
