% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst02.R
\docType{data}
\name{vst02_1_main}
\alias{vst02_1_main}
\alias{vst02_lyt}
\alias{vst02_pre}
\alias{vst02_post}
\alias{vst02_1}
\title{\code{VST02} Vital Sign Abnormalities Table.}
\format{
An object of class \code{chevron_t} of length 1.
}
\usage{
vst02_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = NULL,
  ...
)

vst02_lyt(
  arm_var,
  exclude_base_abn,
  lbl_vs_assessment,
  lbl_vs_abnormality,
  lbl_overall
)

vst02_pre(adam_db, ...)

vst02_post(tlg, prune_0 = FALSE, ...)

vst02_1
}
\arguments{
\item{adam_db}{(\code{list} of \code{data.frames}) object containing the \code{ADaM} datasets}

\item{arm_var}{(\code{string}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{string}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{string}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{string}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{...}{not used.}

\item{exclude_base_abn}{(\code{flag}) whether to exclude subjects with baseline abnormality from numerator and
denominator.}

\item{tlg}{(\code{TableTree}, \code{Listing} or \code{ggplot}) object typically produced by a \code{main} function.}

\item{prune_0}{(\code{flag}) remove 0 count rows}
}
\description{
Vital Sign Parameters outside Normal Limits Regardless of Abnormality at Baseline.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{vst02_1_main()}: Main TLG function

\item \code{vst02_lyt()}: Layout

\item \code{vst02_pre()}: Preprocessing

\item \code{vst02_post()}: Postprocessing

}}
\note{
\itemize{
\item \code{adam_db} object must contain an \code{advs} table with the \code{"PARAM"}, \code{"ANRIND"} and \code{"BNRIND"} columns.
}
}
\examples{
run(vst02_1, syn_data)
}
\keyword{datasets}
