% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst02.R
\docType{data}
\name{vst02_1_lyt}
\alias{vst02_1_lyt}
\alias{vst02_1_pre}
\alias{vst02_1}
\title{\code{VST02_1} Pipeline}
\format{
a \code{tlg_pipeline_s4} object with the following slots:
\itemize{
\item \code{main} the \code{chevron::vst02_1_main} function.
\item \code{preprocess} the  \code{chevron::vst02_1_pre} function.
\item \code{postprocess} the identity function.
\item \code{check} no checks.
\item \code{adam_datasets} \code{"adsl"} and \code{"advs"}.
}
}
\usage{
vst02_1_lyt(
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("VST02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

vst02_1_pre(adam_db, ...)

vst02_1
}
\arguments{
\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{character}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{character}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{...}{not used.}
}
\description{
\code{VST02_1} Pipeline of the class \code{tlg_pipeline_s4}
}
\section{Functions}{
\itemize{
\item \code{vst02_1_lyt}: \code{vst02_1} Layout

\item \code{vst02_1_pre}: \code{vst02_1} Preprocessing
}}

\examples{
vst02_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
vst02_1_pre(syn_test_data())
}
\keyword{datasets}
