% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst02.R
\name{vst02_1}
\alias{vst02_1}
\alias{vst02_1_lyt}
\title{\code{VST02} Table 1 (Default) Vital Sign Abnormalities Table 1}
\usage{
vst02_1(
  adam_db,
  armvar = .study$armvar,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("VST02"),
  .study = list(armvar = "ARM", lbl_overall = NULL)
)

vst02_1_lyt(
  armvar = .study$armvar,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("VST02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the ADaM datasets}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{character}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{character}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}
}
\value{



}
\description{
Assessments Outside Normal Limits Regardless of Abnormality at Baseline Table.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{vst02_1_lyt}: 
}}

\examples{
library(dm)
library(dplyr)

db <- syn_test_data() \%>\%
  preprocess_data("vst02_1")

vst02_1(adam_db = db)

vst02_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)

}
