% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst02.R
\docType{data}
\name{vst02_1_main}
\alias{vst02_1_main}
\alias{vst02_1_lyt}
\alias{vst02_1_pre}
\alias{vst02_1}
\title{\code{VST02} Table 1 (Default) Vital Sign Abnormalities Table 1.}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
vst02_1_main(
  adam_db,
  lyt_fun = list(vst02_1_lyt),
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("VST02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

vst02_1_lyt(
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("VST02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

vst02_1_pre(adam_db, ...)

vst02_1
}
\arguments{
\item{adam_db}{(\code{dm}) object containing the \code{ADaM} datasets}

\item{lyt_fun}{(\verb{list of functions}) creating \verb{Pre-data Table Layout}.}

\item{armvar}{(\code{character}) variable used for column splitting}

\item{lbl_vs_assessment}{(\code{character}) the label of the assessment variable.}

\item{lbl_vs_abnormality}{(\code{character}) the label of the abnormality variable.}

\item{lbl_overall}{(\code{character}) label used for overall column, if set to \code{NULL} the overall column is omitted}

\item{prune_0}{(\code{logical}) remove 0 count rows}

\item{deco}{(\code{character}) decoration with \code{title}, \code{subtitles} and \code{main_footer} content}

\item{.study}{(\code{list})  with default values for the arguments of the function}

\item{...}{not used.}
}
\description{
Assessments Outside Normal Limits Regardless of
Abnormality at Baseline Table.
}
\details{
\itemize{
\item Only count LOW or HIGH values.
\item Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".
\item Does not include a total column by default.
\item Does not remove zero-count rows unless overridden with \code{prune_0 = TRUE}.
}
}
\section{Functions}{
\itemize{
\item \code{vst02_1_main()}: Main TLG function

\item \code{vst02_1_lyt()}: Layout

\item \code{vst02_1_pre()}: Preprocessing

}}
\examples{
run(vst02_1, syn_test_data())
}
\keyword{datasets}
