% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst02.R
\docType{data}
\name{vst02_2}
\alias{vst02_2}
\title{\code{VST02_2}}
\format{
An object of class \code{chevron_tlg} of length 1.
}
\usage{
vst02_2
}
\description{
\code{VST02_2}
}
\seealso{
\code{\link[=vst02_2_main]{vst02_2_main()}}
}
\keyword{datasets}
